@@ -1,13 +1,22 @@
 package cmd
 
 import (
+	"cmp"
 	"fmt"
+	"net/url"
 	"os"
+	"path/filepath"
 	"regexp"
+	"slices"
 
-	"github.com/smartcontractkit/chainlink/core/store"
-
+	"github.com/pkg/errors"
 	"github.com/urfave/cli"
+
+	"github.com/smartcontractkit/chainlink/v2/core/build"
+	"github.com/smartcontractkit/chainlink/v2/core/logger"
+	"github.com/smartcontractkit/chainlink/v2/core/services/chainlink"
+	"github.com/smartcontractkit/chainlink/v2/core/static"
+	"github.com/smartcontractkit/chainlink/v2/core/utils"
 )
 
 func removeHidden(cmds ...cli.Command) []cli.Command {
@@ -22,450 +31,291 @@ func removeHidden(cmds ...cli.Command) []cli.Command {
 }
 
 // NewApp returns the command-line parser/function-router for the given client
-func NewApp(client *Client) *cli.App {
+func NewApp(s *Shell) *cli.App {
 	app := cli.NewApp()
 	app.Usage = "CLI for Chainlink"
-	app.Version = fmt.Sprintf("%v@%v", store.Version, store.Sha)
+	app.Version = fmt.Sprintf("%v@%v", static.Version, static.Sha)
+	// TOML
+	var opts chainlink.GeneralConfigOpts
+
 	app.Flags = []cli.Flag{
 		cli.BoolFlag{
 			Name:  "json, j",
 			Usage: "json output as opposed to table",
 		},
+		cli.StringFlag{
+			Name:  "admin-credentials-file",
+			Usage: fmt.Sprintf("optional, applies only in client mode when making remote API calls. If provided, `FILE` containing admin credentials will be used for logging in, allowing to avoid an additional login step. If `FILE` is missing, it will be ignored. Defaults to %s", filepath.Join("<RootDir>", "apicredentials")),
+		},
+		cli.StringFlag{
+			Name:  "remote-node-url",
+			Usage: "optional, applies only in client mode when making remote API calls. If provided, `URL` will be used as the remote Chainlink API endpoint",
+			Value: "http://localhost:6688",
+		},
+		cli.BoolFlag{
+			Name:  "insecure-skip-verify",
+			Usage: "optional, applies only in client mode when making remote API calls. If turned on, SSL certificate verification will be disabled. This is mostly useful for people who want to use Chainlink with a self-signed TLS certificate",
+		},
+		cli.StringSliceFlag{
+			Name:  "config, c",
+			Usage: "TOML configuration file(s) via flag, or raw TOML via env var. If used, legacy env vars must not be set. Multiple files can be used (-c configA.toml -c configB.toml), and they are applied in order with duplicated fields overriding any earlier values. If the 'CL_CONFIG' env var is specified, it is always processed last with the effect of being the final override. [$CL_CONFIG]",
+			// Note: we cannot use the EnvVar field since it will combine with the flags.
+			Hidden: true,
+		},
+		cli.StringSliceFlag{
+			Name:   "secrets, s",
+			Usage:  "TOML configuration file for secrets. Must be set if and only if config is set. Multiple files can be used (-s secretsA.toml -s secretsB.toml), and they are applied in order. No overrides are allowed.",
+			Hidden: true,
+		},
 	}
 	app.Before = func(c *cli.Context) error {
+		s.configFiles = c.StringSlice("config")
+		s.configFilesIsSet = c.IsSet("config")
+		s.secretsFiles = c.StringSlice("secrets")
+		s.secretsFileIsSet = c.IsSet("secrets")
+
+		// Default to using a stdout logger only.
+		// This is overidden for server commands which may start a rotating
+		// logger instead.
+		lggr, closeFn := logger.NewLogger()
+
+		cfg, err := opts.New()
+		if err != nil {
+			return err
+		}
+
+		s.Logger = lggr
+		s.CloseLogger = closeFn
+		s.Config = cfg
+
 		if c.Bool("json") {
-			client.Renderer = RendererJSON{Writer: os.Stdout}
+			s.Renderer = RendererJSON{Writer: os.Stdout}
+		}
+
+		cookieJar, err := NewUserCache("cookies", func() logger.Logger { return s.Logger })
+		if err != nil {
+			return fmt.Errorf("error initialize chainlink cookie cache: %w", err)
+		}
+
+		urlStr := c.String("remote-node-url")
+		remoteNodeURL, err := url.Parse(urlStr)
+		if err != nil {
+			return errors.Wrapf(err, "%s is not a valid URL", urlStr)
+		}
+
+		insecureSkipVerify := c.Bool("insecure-skip-verify")
+		clientOpts := ClientOpts{RemoteNodeURL: *remoteNodeURL, InsecureSkipVerify: insecureSkipVerify}
+		cookieAuth := NewSessionCookieAuthenticator(clientOpts, DiskCookieStore{Config: cookieJar}, s.Logger)
+		sessionRequestBuilder := NewFileSessionRequestBuilder(s.Logger)
+
+		credentialsFile := c.String("admin-credentials-file")
+		sr, err := sessionRequestBuilder.Build(credentialsFile)
+		if err != nil && !errors.Is(errors.Cause(err), ErrNoCredentialFile) && !os.IsNotExist(err) {
+			return errors.Wrapf(err, "failed to load API credentials from file %s", credentialsFile)
+		}
+
+		s.HTTP = NewAuthenticatedHTTPClient(s.Logger, clientOpts, cookieAuth, sr)
+		s.CookieAuthenticator = cookieAuth
+		s.FileSessionRequestBuilder = sessionRequestBuilder
+
+		// Allow for initServerConfig to be called if the flag is provided.
+		if c.Bool("applyInitServerConfig") {
+			cfg, err = initServerConfig(&opts, s.configFiles, s.secretsFiles)
+			if err != nil {
+				return err
+			}
+			s.Config = cfg
+		}
+
+		return nil
+	}
+	app.After = func(c *cli.Context) error {
+		if s.CloseLogger != nil {
+			return s.CloseLogger()
 		}
 		return nil
 	}
 	app.Commands = removeHidden([]cli.Command{
 		{
-			Name:  "admin",
-			Usage: "Commands for remotely taking admin related actions",
-			Subcommands: []cli.Command{
-				{
-					Name:   "chpass",
-					Usage:  "Change your account password remotely",
-					Action: client.ChangePassword,
-				},
-				{
-					Name:   "info",
-					Usage:  "Display the Account's address with its ETH & LINK balances",
-					Action: client.DisplayAccountBalance,
-				},
-				{
-					Name:   "login",
-					Usage:  "Login to remote client by creating a session cookie",
-					Action: client.RemoteLogin,
-					Flags: []cli.Flag{
-						cli.StringFlag{
-							Name:  "file, f",
-							Usage: "text file holding the API email and password needed to create a session cookie",
-						},
-					},
-				},
-			},
+			Name:        "admin",
+			Usage:       "Commands for remotely taking admin related actions",
+			Subcommands: initAdminSubCmds(s),
 		},
-
-		cli.Command{
-			Name:    "agreements",
-			Aliases: []string{"agree"},
-			Usage:   "Commands for handling service agreements",
-			Hidden:  !client.Config.Dev(),
-			Subcommands: []cli.Command{
-				{
-					Name:   "create",
-					Usage:  "Creates a Service Agreement",
-					Action: client.CreateServiceAgreement,
-				},
-			},
-		},
-
-		cli.Command{
-			Name:    "attempts",
-			Aliases: []string{"txas"},
-			Usage:   "Commands for managing Ethereum Transaction Attempts",
-			Hidden:  !client.Config.Dev(),
-			Subcommands: []cli.Command{
-				{
-					Name:   "list",
-					Usage:  "List the Transaction Attempts in descending order",
-					Action: client.IndexTxAttempts,
-					Flags: []cli.Flag{
-						cli.IntFlag{
-							Name:  "page",
-							Usage: "page of results to display",
-						},
-					},
-				},
-			},
-		},
-
 		{
-			Name:  "bridges",
-			Usage: "Commands for Bridges communicating with External Adapters",
-			Subcommands: []cli.Command{
-				{
-					Name:   "create",
-					Usage:  "Create a new Bridge to an External Adapter",
-					Action: client.CreateBridge,
-				},
-				{
-					Name:   "destroy",
-					Usage:  "Destroys the Bridge for an External Adapter",
-					Action: client.RemoveBridge,
-				},
-				{
-					Name:   "list",
-					Usage:  "List all Bridges to External Adapters",
-					Action: client.IndexBridges,
-					Flags: []cli.Flag{
-						cli.IntFlag{
-							Name:  "page",
-							Usage: "page of results to display",
-						},
-					},
-				},
-				{
-					Name:   "show",
-					Usage:  "Show an Bridge's details",
-					Action: client.ShowBridge,
-				},
-			},
+			Name:        "attempts",
+			Aliases:     []string{"txas"},
+			Usage:       "Commands for managing Ethereum Transaction Attempts",
+			Subcommands: initAttemptsSubCmds(s),
 		},
-
 		{
-			Name:  "config",
-			Usage: "Commands for the node's configuration",
-			Subcommands: []cli.Command{
-				{
-					Name:   "list",
-					Usage:  "Show the node's environment variables",
-					Action: client.GetConfiguration,
-				},
-				{
-					Name:   "setgasprice",
-					Usage:  "Set the minimum gas price to use for outgoing transactions",
-					Action: client.SetMinimumGasPrice,
-					Flags: []cli.Flag{
-						cli.BoolFlag{
-							Name:  "gwei",
-							Usage: "Specify amount in gwei",
-						},
-					},
-				},
-			},
+			Name:        "blocks",
+			Aliases:     []string{},
+			Usage:       "Commands for managing blocks",
+			Subcommands: initBlocksSubCmds(s),
 		},
-
-		cli.Command{
-			Name:   "createextrakey",
-			Usage:  "Create a key in the node's keystore alongside the existing key; to create an original key, just run the node",
-			Hidden: !client.Config.Dev(),
-			Action: client.CreateExtraKey,
-		},
-
 		{
-			Name:  "jobs",
-			Usage: "Commands for managing Jobs",
-			Subcommands: []cli.Command{
-				{
-					Name:   "archive",
-					Usage:  "Archive a Job and all its associated Runs",
-					Action: client.ArchiveJobSpec,
-				},
-				{
-					Name:   "create",
-					Usage:  "Create Job from a Job Specification JSON",
-					Action: client.CreateJobSpec,
-				},
-				{
-					Name:   "list",
-					Usage:  "List all jobs",
-					Action: client.IndexJobSpecs,
-					Flags: []cli.Flag{
-						cli.IntFlag{
-							Name:  "page",
-							Usage: "page of results to display",
-						},
-					},
-				},
-				{
-					Name:   "show",
-					Usage:  "Show a specific Job's details",
-					Action: client.ShowJobSpec,
+			Name:        "bridges",
+			Usage:       "Commands for Bridges communicating with External Adapters",
+			Subcommands: initBrideSubCmds(s),
+		},
+		{
+			Name:        "config",
+			Usage:       "Commands for the node's configuration",
+			Subcommands: initRemoteConfigSubCmds(s),
+		},
+		{
+			Name:   "health",
+			Usage:  "Prints a health report",
+			Action: s.Health,
+			Flags: []cli.Flag{
+				cli.BoolFlag{
+					Name:  "json, j",
+					Usage: "json output",
 				},
 			},
 		},
+		{
+			Name:        "jobs",
+			Usage:       "Commands for managing Jobs",
+			Subcommands: initJobsSubCmds(s),
+		},
+		{
+			Name:  "keys",
+			Usage: "Commands for managing various types of keys used by the Chainlink node",
+			Subcommands: []cli.Command{
+				// TODO unify init vs keysCommand
+				// out of scope for initial refactor because it breaks usage messages.
+				initEthKeysSubCmd(s),
+				initP2PKeysSubCmd(s),
+				initCSAKeysSubCmd(s),
+				initOCRKeysSubCmd(s),
+				initOCR2KeysSubCmd(s),
 
+				keysCommand("Cosmos", NewCosmosKeysClient(s)),
+				keysCommand("Solana", NewSolanaKeysClient(s)),
+				keysCommand("StarkNet", NewStarkNetKeysClient(s)),
+				keysCommand("DKGSign", NewDKGSignKeysClient(s)),
+				keysCommand("DKGEncrypt", NewDKGEncryptKeysClient(s)),
+
+				initVRFKeysSubCmd(s),
+			},
+		},
 		{
 			Name:        "node",
 			Aliases:     []string{"local"},
 			Usage:       "Commands for admin actions that must be run locally",
 			Description: "Commands can only be run from on the same machine as the Chainlink node.",
-			Subcommands: []cli.Command{
-				{
-					Name:        "deleteuser",
-					Usage:       "Erase the *local node's* user and corresponding session to force recreation on next node launch.",
-					Description: "Does not work remotely over API.",
-					Action:      client.DeleteUser,
+			Subcommands: initLocalSubCmds(s, build.IsProd()),
+			Flags: []cli.Flag{
+				cli.StringSliceFlag{
+					Name:  "config, c",
+					Usage: "TOML configuration file(s) via flag, or raw TOML via env var. If used, legacy env vars must not be set. Multiple files can be used (-c configA.toml -c configB.toml), and they are applied in order with duplicated fields overriding any earlier values. If the 'CL_CONFIG' env var is specified, it is always processed last with the effect of being the final override. [$CL_CONFIG]",
 				},
-				{
-					Name:    "import",
-					Aliases: []string{"i"},
-					Usage:   "Import a key file to use with the node",
-					Action:  client.ImportKey,
-				},
-				{
-					Name:   "setnextnonce",
-					Usage:  "Manually set the next nonce for a key. This should NEVER be necessary during normal operation. USE WITH CAUTION: Setting this incorrectly can break your node.",
-					Action: client.SetNextNonce,
-					Flags: []cli.Flag{
-						cli.StringFlag{
-							Name:  "address",
-							Usage: "address of the key for which to set the nonce",
-						},
-						cli.Uint64Flag{
-							Name:  "nextNonce",
-							Usage: "the next nonce in the sequence",
-						},
-					},
-				},
-				{
-					Name:    "start",
-					Aliases: []string{"node", "n"},
-					Flags: []cli.Flag{
-						cli.StringFlag{
-							Name:  "api, a",
-							Usage: "text file holding the API email and password, each on a line",
-						},
-						cli.BoolFlag{
-							Name:  "debug, d",
-							Usage: "set logger level to debug",
-						},
-						cli.StringFlag{
-							Name:  "password, p",
-							Usage: "text file holding the password for the node's account",
-						},
-						cli.StringFlag{
-							Name:  "vrfpassword, vp",
-							Usage: "textfile holding the password for the vrf keys; enables chainlink VRF oracle",
-						},
-						cli.Int64Flag{
-							Name:  "replay-from-block, r",
-							Usage: "historical block height from which to replay log-initiated jobs",
-							Value: -1,
-						},
-					},
-					Usage:  "Run the chainlink node",
-					Action: client.RunNode,
-				},
-				cli.Command{
-					Name:   "p2p",
-					Usage:  "Local commands for administering the node's p2p layer",
-					Hidden: !client.Config.Dev(),
-					Subcommands: cli.Commands{
-						{
-							Name: "create",
-							Usage: format(`Create a p2p key, encrypted with password from the
-               password file, and store it in the database.`),
-							Flags:  flags("password, p"),
-							Action: client.CreateP2PKey,
-						},
-					},
-				},
-				cli.Command{
-					Name: "vrf",
-					Usage: format(`Local commands for administering the database of VRF proof
-           keys. These commands will not affect the extant in-memory keys of
-           any live node.`),
-					Hidden: !client.Config.Dev(),
-					Subcommands: cli.Commands{
-						{
-							Name: "create",
-							Usage: format(`Create a VRF key, encrypted with password from the
-               password file, and store it in the database.`),
-							Flags:  flags("password, p"),
-							Action: client.CreateVRFKey,
-						},
-						{
-							Name:   "import",
-							Usage:  "Import key from keyfile.",
-							Flags:  append(flags("password, p"), flags("file, f")...),
-							Action: client.ImportVRFKey,
-						},
-						{
-							Name:   "export",
-							Usage:  "Export key to keyfile.",
-							Flags:  append(flags("file, f"), flags("publicKey, pk")...),
-							Action: client.ExportVRFKey,
-						},
-						{
-							Name:   "delete",
-							Usage:  "Remove key from database, if present",
-							Flags:  flags("publicKey, pk"),
-							Action: client.DeleteVRFKey,
-						},
-						{
-							Name: "list", Usage: "List the public keys in the db",
-							Action: client.ListKeys,
-						},
-						{
-							Name: "",
-						},
-						{
-							Name: "xxxCreateWeakKeyPeriodYesIReallyKnowWhatIAmDoingAndDoNotCareAboutThisKeyMaterialFallingIntoTheWrongHandsExclamationPointExclamationPointExclamationPointExclamationPointIAmAMasochistExclamationPointExclamationPointExclamationPointExclamationPointExclamationPoint",
-							Usage: format(`
-               For testing purposes ONLY! DO NOT USE FOR ANY OTHER PURPOSE!
-
-               Creates a key with weak key-devrivation-function parameters, so that it can be
-               decrypted quickly during tests. As a result, it would be cheap to brute-force
-               the encryption password for the key, if the ciphertext fell into the wrong
-               hands!`),
-							Flags:  append(flags("password, p"), flags("file, f")...),
-							Action: client.CreateAndExportWeakVRFKey,
-							Hidden: !client.Config.Dev(), // For when this suite gets promoted out of dev mode
-						},
-					},
-				},
-				{
-					Name:   "rebroadcast-transactions",
-					Usage:  "Manually rebroadcast txs matching nonce range with the specified gas price. This is useful in emergencies e.g. high gas prices and/or network congestion to forcibly clear out the pending TX queue",
-					Action: client.RebroadcastTransactions,
-					Flags: []cli.Flag{
-						cli.Uint64Flag{
-							Name:  "beginningNonce, b",
-							Usage: "beginning of nonce range to rebroadcast",
-						},
-						cli.Uint64Flag{
-							Name:  "endingNonce, e",
-							Usage: "end of nonce range to rebroadcast (inclusive)",
-						},
-						cli.Uint64Flag{
-							Name:  "gasPriceWei, g",
-							Usage: "gas price (in Wei) to rebroadcast transactions at",
-						},
-						cli.StringFlag{
-							Name:  "password, p",
-							Usage: "text file holding the password for the node's account",
-						},
-						cli.StringFlag{
-							Name:  "address, a",
-							Usage: "The address (in hex format) for the key which we want to rebroadcast transactions",
-						},
-						cli.Uint64Flag{
-							Name:  "gasLimit",
-							Usage: "OPTIONAL: gas limit to use for each transaction ",
-						},
-					},
-				},
-				{
-					Name:        "db",
-					Usage:       "Commands for managing the database.",
-					Description: "Potentially destructive commands for managing the database, only intended for dev/testing purposes.",
-					Hidden:      !client.Config.Dev(),
-					Subcommands: []cli.Command{
-						{
-							Name:   "reset",
-							Usage:  "Drop, create and migrate database. Useful for setting up the database in order to run tests or resetting the dev database. WARNING: This will ERASE ALL DATA for the specified DATABASE_URL.",
-							Hidden: !client.Config.Dev(),
-							Action: client.ResetDatabase,
-							Flags:  []cli.Flag{},
-						},
-						{
-							Name:   "preparetest",
-							Usage:  "Reset database and load fixtures.",
-							Hidden: !client.Config.Dev(),
-							Action: client.PrepareTestDatabase,
-							Flags:  []cli.Flag{},
-						},
-					},
+				cli.StringSliceFlag{
+					Name:  "secrets, s",
+					Usage: "TOML configuration file for secrets. Must be set if and only if config is set. Multiple files can be used (-s secretsA.toml -s secretsB.toml), and fields from the files will be merged. No overrides are allowed.",
 				},
 			},
-		},
+			Before: func(c *cli.Context) error {
+				errNoDuplicateFlags := fmt.Errorf("multiple commands with --config or --secrets flags. only one command may specify these flags. when secrets are used, they must be specific together in the same command")
+				if c.IsSet("config") {
+					if s.configFilesIsSet || s.secretsFileIsSet {
+						return errNoDuplicateFlags
+					}
+					s.configFiles = c.StringSlice("config")
+				}
 
-		cli.Command{
-			Name:   "initiators",
-			Usage:  "Commands for managing External Initiators",
-			Hidden: !client.Config.Dev() && !client.Config.FeatureExternalInitiators(),
-			Subcommands: []cli.Command{
-				{
-					Name:   "create",
-					Usage:  "Create an authentication key for a user of External Initiators",
-					Action: client.CreateExternalInitiator,
-				},
-				{
-					Name:   "destroy",
-					Usage:  "Remove an authentication key by name",
-					Action: client.DeleteExternalInitiator,
-				},
+				if c.IsSet("secrets") {
+					if s.configFilesIsSet || s.secretsFileIsSet {
+						return errNoDuplicateFlags
+					}
+					s.secretsFiles = c.StringSlice("secrets")
+				}
+
+				// flags here, or ENV VAR only
+				cfg, err := initServerConfig(&opts, s.configFiles, s.secretsFiles)
+				if err != nil {
+					return err
+				}
+				s.Config = cfg
+
+				logFileMaxSizeMB := s.Config.Log().File().MaxSize() / utils.MB
+				if logFileMaxSizeMB > 0 {
+					err = utils.EnsureDirAndMaxPerms(s.Config.Log().File().Dir(), os.FileMode(0700))
+					if err != nil {
+						return err
+					}
+				}
+
+				// Swap out the logger, replacing the old one.
+				err = s.CloseLogger()
+				if err != nil {
+					return err
+				}
+
+				lggrCfg := logger.Config{
+					LogLevel:       s.Config.Log().Level(),
+					Dir:            s.Config.Log().File().Dir(),
+					JsonConsole:    s.Config.Log().JSONConsole(),
+					UnixTS:         s.Config.Log().UnixTimestamps(),
+					FileMaxSizeMB:  int(logFileMaxSizeMB),
+					FileMaxAgeDays: int(s.Config.Log().File().MaxAgeDays()),
+					FileMaxBackups: int(s.Config.Log().File().MaxBackups()),
+				}
+				l, closeFn := lggrCfg.New()
+
+				s.Logger = l
+				s.CloseLogger = closeFn
+
+				return nil
 			},
 		},
-
 		{
-			Name:  "runs",
-			Usage: "Commands for managing Runs",
-			Subcommands: []cli.Command{
-				{
-					Name:        "create",
-					Aliases:     []string{"r"},
-					Usage:       "Create a new Run for a Job given an Job ID and optional JSON body",
-					Description: "Takes a Job ID and a JSON string or path to a JSON file",
-					Action:      client.CreateJobRun,
-				},
-				{
-					Name:   "list",
-					Usage:  "List all Runs",
-					Action: client.IndexJobRuns,
-					Flags: []cli.Flag{
-						cli.IntFlag{
-							Name:  "page",
-							Usage: "page of results to display",
-						},
-						cli.StringFlag{
-							Name:  "jobid",
-							Usage: "filter all Runs to match the given jobid",
-						},
-					},
-				},
-				{
-					Name:    "show",
-					Aliases: []string{"sr"},
-					Usage:   "Show a Run for a specific ID",
-					Action:  client.ShowJobRun,
-				},
-				{
-					Name:   "cancel",
-					Usage:  "Cancel a Run with a specified ID",
-					Action: client.CancelJobRun,
-				},
-			},
+			Name:        "initiators",
+			Usage:       "Commands for managing External Initiators",
+			Subcommands: initInitiatorsSubCmds(s),
 		},
-
 		{
 			Name:  "txs",
-			Usage: "Commands for handling Ethereum transactions",
+			Usage: "Commands for handling transactions",
 			Subcommands: []cli.Command{
-				{
-					Name:   "create",
-					Usage:  "Send <amount> Eth from node ETH account <fromAddress> to destination <toAddress>.",
-					Action: client.SendEther,
-				},
-				{
-					Name:   "list",
-					Usage:  "List the Ethereum Transactions in descending order",
-					Action: client.IndexTransactions,
-					Flags: []cli.Flag{
-						cli.IntFlag{
-							Name:  "page",
-							Usage: "page of results to display",
-						},
-					},
-				},
-				{
-					Name:   "show",
-					Usage:  "get information on a specific Ethereum Transaction",
-					Action: client.ShowTransaction,
-				},
+				initEVMTxSubCmd(s),
+				initCosmosTxSubCmd(s),
+				initSolanaTxSubCmd(s),
+			},
+		},
+		{
+			Name:  "chains",
+			Usage: "Commands for handling chain configuration",
+			Subcommands: cli.Commands{
+				chainCommand("EVM", EVMChainClient(s), cli.Int64Flag{Name: "id", Usage: "chain ID"}),
+				chainCommand("Cosmos", CosmosChainClient(s), cli.StringFlag{Name: "id", Usage: "chain ID"}),
+				chainCommand("Solana", SolanaChainClient(s),
+					cli.StringFlag{Name: "id", Usage: "chain ID, options: [mainnet, testnet, devnet, localnet]"}),
+				chainCommand("StarkNet", StarkNetChainClient(s), cli.StringFlag{Name: "id", Usage: "chain ID"}),
+			},
+		},
+		{
+			Name:  "nodes",
+			Usage: "Commands for handling node configuration",
+			Subcommands: cli.Commands{
+				initEVMNodeSubCmd(s),
+				initCosmosNodeSubCmd(s),
+				initSolanaNodeSubCmd(s),
+				initStarkNetNodeSubCmd(s),
+			},
+		},
+		{
+			Name:        "forwarders",
+			Usage:       "Commands for managing forwarder addresses.",
+			Subcommands: initFowardersSubCmds(s),
+		},
+		{
+			Name:  "help-all",
+			Usage: "Shows a list of all commands and sub-commands",
+			Action: func(c *cli.Context) error {
+				printCommands("", c.App.Commands)
+				return nil
 			},
 		},
 	}...)
@@ -479,5 +329,25 @@ func format(s string) string {
 	return string(whitespace.ReplaceAll([]byte(s), []byte(" ")))
 }
 
-// flags is an abbreviated way to express a CLI flag
-func flags(s string) []cli.Flag { return []cli.Flag{cli.StringFlag{Name: s}} }
+func initServerConfig(opts *chainlink.GeneralConfigOpts, configFiles []string, secretsFiles []string) (chainlink.GeneralConfig, error) {
+	err := opts.Setup(configFiles, secretsFiles)
+	if err != nil {
+		return nil, err
+	}
+	return opts.New()
+}
+
+func printCommands(parent string, cs cli.Commands) {
+	slices.SortFunc(cs, func(a, b cli.Command) int {
+		return cmp.Compare(a.Name, b.Name)
+	})
+	for i := range cs {
+		c := cs[i]
+		name := c.Name
+		if parent != "" {
+			name = parent + " " + name
+		}
+		fmt.Printf("%s # %s\n", name, c.Usage)
+		printCommands(name, c.Subcommands)
+	}
+}
