@@ -3,22 +3,17 @@ package cmd
 import (
 	"fmt"
 	"io"
-	"reflect"
-	"strconv"
-
-	"github.com/smartcontractkit/chainlink/core/logger"
-	"github.com/smartcontractkit/chainlink/core/store/models"
-	"github.com/smartcontractkit/chainlink/core/store/orm"
-	"github.com/smartcontractkit/chainlink/core/store/presenters"
-	"github.com/smartcontractkit/chainlink/core/utils"
-	"github.com/smartcontractkit/chainlink/core/web"
+	"strings"
 
 	"github.com/olekukonko/tablewriter"
+
+	"github.com/smartcontractkit/chainlink/v2/core/utils"
+	webpresenters "github.com/smartcontractkit/chainlink/v2/core/web/presenters"
 )
 
 // Renderer implements the Render method.
 type Renderer interface {
-	Render(interface{}) error
+	Render(interface{}, ...string) error
 }
 
 // RendererJSON is used to render JSON data.
@@ -27,11 +22,14 @@ type RendererJSON struct {
 }
 
 // Render writes the given input as a JSON string.
-func (rj RendererJSON) Render(v interface{}) error {
+func (rj RendererJSON) Render(v interface{}, _ ...string) error {
 	b, err := utils.FormatJSON(v)
 	if err != nil {
 		return err
 	}
+
+	// Append a new line
+	b = append(b, []byte("\n")...)
 	if _, err = rj.Write(b); err != nil {
 		return err
 	}
@@ -43,95 +41,60 @@ type RendererTable struct {
 	io.Writer
 }
 
+type TableRenderer interface {
+	RenderTable(rt RendererTable) error
+}
+
 // Render returns a formatted table of text for a given Job or presenter
 // and relevant information.
-func (rt RendererTable) Render(v interface{}) error {
+func (rt RendererTable) Render(v interface{}, headers ...string) error {
+	for _, h := range headers {
+		fmt.Println(h)
+	}
+
 	switch typed := v.(type) {
-	case *[]models.JobSpec:
-		return rt.renderJobs(*typed)
-	case *presenters.JobSpec:
-		return rt.renderJob(*typed)
-	case *[]presenters.JobRun:
-		return rt.renderJobRuns(*typed)
-	case *presenters.JobRun:
-		return rt.renderJobRun(*typed)
-	case *models.BridgeType:
-		return rt.renderBridge(*typed)
-	case *models.BridgeTypeAuthentication:
-		return rt.renderBridgeAuthentication(*typed)
-	case *[]models.BridgeType:
-		return rt.renderBridges(*typed)
-	case *[]presenters.AccountBalance:
-		return rt.renderAccountBalances(*typed)
-	case *presenters.ServiceAgreement:
-		return rt.renderServiceAgreement(*typed)
-	case *[]models.TxAttempt:
-		return rt.renderTxAttempts(*typed)
-	case *[]presenters.Tx:
-		return rt.renderTxs(*typed)
-	case *presenters.Tx:
-		return rt.renderTx(*typed)
-	case *presenters.ExternalInitiatorAuthentication:
+	case *webpresenters.ExternalInitiatorAuthentication:
 		return rt.renderExternalInitiatorAuthentication(*typed)
-	case *web.ConfigPatchResponse:
-		return rt.renderConfigPatchResponse(typed)
-	case *presenters.ConfigWhitelist:
-		return rt.renderConfiguration(*typed)
+	case *webpresenters.PipelineRunResource:
+		return rt.renderPipelineRun(*typed)
+	case *webpresenters.ServiceLogConfigResource:
+		return rt.renderLogPkgConfig(*typed)
+	case *[]VRFKeyPresenter:
+		return rt.renderVRFKeys(*typed)
+	case TableRenderer:
+		return typed.RenderTable(rt)
 	default:
 		return fmt.Errorf("unable to render object of type %T: %v", typed, typed)
 	}
 }
 
-func (rt RendererTable) renderJobs(jobs []models.JobSpec) error {
-	table := rt.newTable([]string{"ID", "Created At", "Initiators", "Tasks"})
-	for _, v := range jobs {
-		table.Append(jobRowToStrings(v))
+func (rt RendererTable) renderLogPkgConfig(serviceLevelLog webpresenters.ServiceLogConfigResource) error {
+	table := rt.newTable([]string{"ID", "Service", "LogLevel"})
+	for i, svcName := range serviceLevelLog.ServiceName {
+		table.Append([]string{
+			serviceLevelLog.ID,
+			svcName,
+			serviceLevelLog.LogLevel[i],
+		})
 	}
 
-	render("Jobs", table)
+	render("ServiceLogConfig", table)
 	return nil
 }
 
-func (rt RendererTable) renderConfiguration(cwl presenters.ConfigWhitelist) error {
-	table := rt.newTable([]string{"Key", "Value"})
+func (rt RendererTable) renderVRFKeys(keys []VRFKeyPresenter) error {
+	var rows [][]string
 
-	table.Append([]string{
-		"ACCOUNT_ADDRESS",
-		cwl.AccountAddress,
-	})
-
-	schemaT := reflect.TypeOf(orm.ConfigSchema{})
-	cwlT := reflect.TypeOf(cwl.Whitelist)
-	cwlV := reflect.ValueOf(cwl.Whitelist)
-
-	for index := 0; index < cwlT.NumField(); index++ {
-		item := cwlT.FieldByIndex([]int{index})
-		schemaItem, ok := schemaT.FieldByName(item.Name)
-		if !ok {
-			logger.Panicf("Field %s missing from store.Schema", item.Name)
-		}
-		envName, ok := schemaItem.Tag.Lookup("env")
-		if !ok {
-			continue
-		}
-		field := cwlV.FieldByIndex(item.Index)
-
-		if stringer, ok := field.Interface().(fmt.Stringer); ok {
-			if stringer != reflect.Zero(reflect.TypeOf(stringer)).Interface() {
-				table.Append([]string{
-					envName,
-					stringer.String(),
-				})
-			}
-		} else {
-			table.Append([]string{
-				envName,
-				fmt.Sprintf("%v", field),
-			})
-		}
+	for _, key := range keys {
+		rows = append(rows, []string{
+			key.Compressed,
+			key.Uncompressed,
+			key.Hash,
+		})
 	}
 
-	render("Configuration", table)
+	renderList([]string{"Compressed", "Uncompressed", "Hash"}, rows, rt.Writer)
+
 	return nil
 }
 
@@ -145,165 +108,40 @@ func render(name string, table *tablewriter.Table) {
 	table.Render()
 }
 
-func jobRowToStrings(job models.JobSpec) []string {
-	p := presenters.JobSpec{JobSpec: job}
-	return []string{
-		p.ID.String(),
-		p.FriendlyCreatedAt(),
-		p.FriendlyInitiators(),
-		p.FriendlyTasks(),
+func renderList(fields []string, items [][]string, writer io.Writer) {
+	var maxLabelLength int
+	for _, field := range fields {
+		if len(field) > maxLabelLength {
+			maxLabelLength = len(field)
+		}
+	}
+	var itemsRendered []string
+	var maxLineLength int
+	for _, row := range items {
+		var lines []string
+		for i, field := range fields {
+			diff := maxLabelLength - len(field)
+			spaces := strings.Repeat(" ", diff)
+			line := fmt.Sprintf("%v: %v%v", field, spaces, row[i])
+			for _, l := range strings.Split(line, "\n") {
+				if len(l) > maxLineLength {
+					maxLineLength = len(l)
+				}
+			}
+			lines = append(lines, line)
+		}
+		itemsRendered = append(itemsRendered, strings.Join(lines, "\n"))
+	}
+	divider := "\n" + strings.Repeat("-", maxLineLength) + "\n"
+	listRendered := divider + strings.Join(itemsRendered, divider) + divider
+	_, err := writer.Write([]byte(listRendered))
+	if err != nil {
+		// Handles errcheck
+		return
 	}
 }
 
-func bridgeRowToStrings(bridge models.BridgeType) []string {
-	return []string{
-		bridge.Name.String(),
-		bridge.URL.String(),
-		strconv.FormatUint(uint64(bridge.Confirmations), 10),
-	}
-}
-
-func (rt RendererTable) renderBridges(bridges []models.BridgeType) error {
-	table := rt.newTable([]string{"Name", "URL", "Confirmations"})
-	for _, v := range bridges {
-		table.Append(bridgeRowToStrings(v))
-	}
-
-	render("Bridges", table)
-	return nil
-}
-
-func (rt RendererTable) renderBridge(bridge models.BridgeType) error {
-	table := rt.newTable([]string{"Name", "URL", "Default Confirmations", "Outgoing Token"})
-	table.Append([]string{
-		bridge.Name.String(),
-		bridge.URL.String(),
-		strconv.FormatUint(uint64(bridge.Confirmations), 10),
-		bridge.OutgoingToken,
-	})
-	render("Bridge", table)
-	return nil
-}
-
-func (rt RendererTable) renderBridgeAuthentication(bridge models.BridgeTypeAuthentication) error {
-	table := rt.newTable([]string{"Name", "URL", "Default Confirmations", "Incoming Token", "Outgoing Token"})
-	table.Append([]string{
-		bridge.Name.String(),
-		bridge.URL.String(),
-		strconv.FormatUint(uint64(bridge.Confirmations), 10),
-		bridge.IncomingToken,
-		bridge.OutgoingToken,
-	})
-	render("Bridge", table)
-	return nil
-}
-
-func (rt RendererTable) renderJob(job presenters.JobSpec) error {
-	if err := rt.renderJobSingles(job); err != nil {
-		return err
-	}
-
-	if err := rt.renderJobInitiators(job); err != nil {
-		return err
-	}
-
-	err := rt.renderJobTasks(job)
-	return err
-}
-
-func (rt RendererTable) renderJobRun(run presenters.JobRun) error {
-	err := rt.renderJobRuns([]presenters.JobRun{run})
-	return err
-}
-
-func (rt RendererTable) renderJobSingles(j presenters.JobSpec) error {
-	table := rt.newTable([]string{"ID", "Created At", "Start At", "End At", "Min Payment"})
-	table.Append([]string{
-		j.ID.String(),
-		j.FriendlyCreatedAt(),
-		j.FriendlyStartAt(),
-		j.FriendlyEndAt(),
-		j.FriendlyMinPayment(),
-	})
-	render("Job", table)
-	return nil
-}
-
-func (rt RendererTable) renderJobInitiators(j presenters.JobSpec) error {
-	table := rt.newTable([]string{"Type", "Schedule", "Run At", "Address"})
-	for _, i := range j.Initiators {
-		p := presenters.Initiator{Initiator: i}
-		table.Append([]string{
-			p.Type,
-			p.Schedule.String(),
-			p.FriendlyRunAt(),
-			p.FriendlyAddress(),
-		})
-	}
-
-	render("Initiators", table)
-	return nil
-}
-
-func (rt RendererTable) renderJobTasks(j presenters.JobSpec) error {
-	table := rt.newTable([]string{"Type", "Config", "Value"})
-	table.SetAutoWrapText(false)
-	for _, t := range j.Tasks {
-		p := presenters.TaskSpec{TaskSpec: t}
-		keys, values := p.FriendlyParams()
-		table.Append([]string{p.Type.String(), keys, values})
-	}
-
-	render("Tasks", table)
-	return nil
-}
-
-func (rt RendererTable) renderJobRuns(runs []presenters.JobRun) error {
-	table := rt.newTable([]string{"ID", "Status", "Created", "Completed", "Result", "Error"})
-	for _, jr := range runs {
-		table.Append([]string{
-			jr.ID.String(),
-			string(jr.GetStatus()),
-			utils.ISO8601UTC(jr.CreatedAt),
-			utils.NullISO8601UTC(jr.FinishedAt),
-			jr.Result.Data.String(),
-			jr.ErrorString(),
-		})
-	}
-
-	render("Runs", table)
-	return nil
-}
-
-func (rt RendererTable) renderAccountBalances(balances []presenters.AccountBalance) error {
-	table := rt.newTable([]string{"Address", "ETH", "LINK"})
-	for _, ab := range balances {
-		table.Append([]string{
-			ab.Address,
-			ab.EthBalance.String(),
-			ab.LinkBalance.String(),
-		})
-	}
-	render("Account Balance", table)
-	return nil
-}
-
-func (rt RendererTable) renderServiceAgreement(sa presenters.ServiceAgreement) error {
-	table := rt.newTable([]string{"ID", "Created At", "Payment", "Expiration", "Aggregator", "AggInit", "AggFulfill"})
-	table.Append([]string{
-		sa.ID,
-		sa.FriendlyCreatedAt(),
-		sa.FriendlyPayment(),
-		sa.FriendlyExpiration(),
-		sa.FriendlyAggregator(),
-		sa.FriendlyAggregatorInitMethod(),
-		sa.FriendlyAggregatorFulfillMethod(),
-	})
-	render("Service Agreement", table)
-	return nil
-}
-
-func (rt RendererTable) renderExternalInitiatorAuthentication(eia presenters.ExternalInitiatorAuthentication) error {
+func (rt RendererTable) renderExternalInitiatorAuthentication(eia webpresenters.ExternalInitiatorAuthentication) error {
 	table := rt.newTable([]string{"Name", "URL", "AccessKey", "Secret", "OutgoingToken", "OutgoingSecret"})
 	table.Append([]string{
 		eia.Name,
@@ -323,59 +161,21 @@ func (rt RendererTable) newTable(headers []string) *tablewriter.Table {
 	return table
 }
 
-func (rt RendererTable) renderTxAttempts(attempts []models.TxAttempt) error {
-	table := rt.newTable([]string{"TxID", "Hash", "GasPrice", "SentAt", "Confirmed"})
-	for _, a := range attempts {
-		table.Append([]string{
-			fmt.Sprint(a.TxID),
-			a.Hash.Hex(),
-			fmt.Sprint(a.GasPrice),
-			fmt.Sprint(a.SentAt),
-			fmt.Sprint(a.Confirmed),
-		})
+func (rt RendererTable) renderPipelineRun(run webpresenters.PipelineRunResource) error {
+	table := rt.newTable([]string{"ID", "Created At", "Finished At"})
+
+	var finishedAt string
+	if !run.FinishedAt.IsZero() {
+		finishedAt = run.FinishedAt.ValueOrZero().String()
 	}
 
-	render("Ethereum Transaction Attempts", table)
-	return nil
-}
-
-func (rt RendererTable) renderTx(tx presenters.Tx) error {
-	table := rt.newTable([]string{"From", "Nonce", "To", "Confirmed"})
-	table.Append([]string{
-		tx.From.Hex(),
-		tx.Nonce,
-		tx.To.Hex(),
-		fmt.Sprint(tx.Confirmed),
-	})
-
-	render(fmt.Sprintf("Ethereum Transaction %v", tx.Hash.Hex()), table)
-	return nil
-}
-
-func (rt RendererTable) renderTxs(txs []presenters.Tx) error {
-	table := rt.newTable([]string{"Hash", "Nonce", "From", "GasPrice", "SentAt", "Confirmed"})
-	for _, tx := range txs {
-		table.Append([]string{
-			tx.Hash.Hex(),
-			tx.Nonce,
-			tx.From.Hex(),
-			tx.GasPrice,
-			tx.SentAt,
-			fmt.Sprint(tx.Confirmed),
-		})
+	row := []string{
+		run.GetID(),
+		run.CreatedAt.String(),
+		finishedAt,
 	}
+	table.Append(row)
 
-	render("Ethereum Transactions", table)
-	return nil
-}
-
-func (rt RendererTable) renderConfigPatchResponse(config *web.ConfigPatchResponse) error {
-	table := rt.newTable([]string{"Config", "Old Value", "New Value"})
-	table.Append([]string{
-		"EthGasPriceDefault",
-		config.EthGasPriceDefault.From,
-		config.EthGasPriceDefault.To,
-	})
-	render("Configuration Changes", table)
+	render("Pipeline Run", table)
 	return nil
 }
