@@ -2,346 +2,106 @@ package cltest
 
 import (
 	"bytes"
-	"crypto/rand"
-	"encoding/binary"
 	"encoding/json"
 	"flag"
 	"fmt"
 	"math/big"
+	mathrand "math/rand"
 	"net/url"
-	"strings"
+	"strconv"
 	"testing"
 	"time"
-
-	"github.com/smartcontractkit/chainlink/core/adapters"
-	"github.com/smartcontractkit/chainlink/core/assets"
-	"github.com/smartcontractkit/chainlink/core/internal/mocks"
-	"github.com/smartcontractkit/chainlink/core/logger"
-	"github.com/smartcontractkit/chainlink/core/services/fluxmonitor"
-	strpkg "github.com/smartcontractkit/chainlink/core/store"
-	"github.com/smartcontractkit/chainlink/core/store/models"
-	"github.com/smartcontractkit/chainlink/core/utils"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/stretchr/testify/assert"
+	"github.com/google/uuid"
+	ragep2ptypes "github.com/smartcontractkit/libocr/ragep2p/types"
 	"github.com/stretchr/testify/require"
-	"github.com/tidwall/gjson"
-	"github.com/tidwall/sjson"
 	"github.com/urfave/cli"
+	"gopkg.in/guregu/null.v4"
+
+	"github.com/jmoiron/sqlx"
+
+	"github.com/smartcontractkit/chainlink-common/pkg/sqlutil"
+	"github.com/smartcontractkit/chainlink-common/pkg/utils/jsonserializable"
+	txmgrcommon "github.com/smartcontractkit/chainlink/v2/common/txmgr"
+	txmgrtypes "github.com/smartcontractkit/chainlink/v2/common/txmgr/types"
+	"github.com/smartcontractkit/chainlink/v2/core/auth"
+	"github.com/smartcontractkit/chainlink/v2/core/bridges"
+	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/assets"
+	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/gas"
+	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/headtracker"
+	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/txmgr"
+	evmtypes "github.com/smartcontractkit/chainlink/v2/core/chains/evm/types"
+	evmutils "github.com/smartcontractkit/chainlink/v2/core/chains/evm/utils"
+	ubig "github.com/smartcontractkit/chainlink/v2/core/chains/evm/utils/big"
+	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/generated/flux_aggregator_wrapper"
+	"github.com/smartcontractkit/chainlink/v2/core/internal/testutils"
+	"github.com/smartcontractkit/chainlink/v2/core/internal/testutils/configtest"
+	"github.com/smartcontractkit/chainlink/v2/core/logger"
+	"github.com/smartcontractkit/chainlink/v2/core/services/job"
+	"github.com/smartcontractkit/chainlink/v2/core/services/keeper"
+	"github.com/smartcontractkit/chainlink/v2/core/services/keystore"
+	"github.com/smartcontractkit/chainlink/v2/core/services/keystore/keys/ethkey"
+	"github.com/smartcontractkit/chainlink/v2/core/services/pipeline"
+	"github.com/smartcontractkit/chainlink/v2/core/store/models"
+	"github.com/smartcontractkit/chainlink/v2/core/utils"
 )
 
-// NewJob return new NoOp JobSpec
-func NewJob() models.JobSpec {
-	j := models.NewJob()
-	j.Tasks = []models.TaskSpec{{Type: adapters.TaskTypeNoOp}}
-	return j
-}
-
-// NewTask given the tasktype and json params return a TaskSpec
-func NewTask(t *testing.T, taskType string, json ...string) models.TaskSpec {
-	if len(json) == 0 {
-		json = append(json, ``)
+func NewEIP55Address() evmtypes.EIP55Address {
+	a := testutils.NewAddress()
+	e, err := evmtypes.NewEIP55Address(a.Hex())
+	if err != nil {
+		panic(err)
 	}
-	params := JSONFromString(t, json[0])
-	params, err := params.Add("type", taskType)
-	require.NoError(t, err)
+	return e
+}
 
-	return models.TaskSpec{
-		Type:   models.MustNewTaskType(taskType),
-		Params: params,
+func NewPeerID() (id ragep2ptypes.PeerID) {
+	err := id.UnmarshalText([]byte("12D3KooWL3XJ9EMCyZvmmGXL2LMiVBtrVa2BuESsJiXkSj7333Jw"))
+	if err != nil {
+		panic(err)
 	}
+	return id
 }
 
-// NewJobWithExternalInitiator creates new Job with external initiator
-func NewJobWithExternalInitiator(ei *models.ExternalInitiator) models.JobSpec {
-	j := NewJob()
-	j.Initiators = []models.Initiator{{
-		JobSpecID: j.ID,
-		Type:      models.InitiatorExternal,
-		InitiatorParams: models.InitiatorParams{
-			Name: ei.Name,
-		},
-	}}
-	return j
-}
-
-// NewJobWithSchedule create new job with the given schedule
-func NewJobWithSchedule(sched string) models.JobSpec {
-	j := NewJob()
-	j.Initiators = []models.Initiator{{
-		JobSpecID: j.ID,
-		Type:      models.InitiatorCron,
-		InitiatorParams: models.InitiatorParams{
-			Schedule: models.Cron(sched),
-		},
-	}}
-	return j
-}
-
-// NewJobWithWebInitiator create new Job with web initiator
-func NewJobWithWebInitiator() models.JobSpec {
-	j := NewJob()
-	j.Initiators = []models.Initiator{{
-		JobSpecID: j.ID,
-		Type:      models.InitiatorWeb,
-	}}
-	return j
-}
-
-// NewJobWithLogInitiator create new Job with ethlog initiator
-func NewJobWithLogInitiator() models.JobSpec {
-	j := NewJob()
-	j.Initiators = []models.Initiator{{
-		JobSpecID: j.ID,
-		Type:      models.InitiatorEthLog,
-		InitiatorParams: models.InitiatorParams{
-			Address: NewAddress(),
-		},
-	}}
-	return j
-}
-
-// NewJobWithRunLogInitiator creates a new JobSpec with the RunLog initiator
-func NewJobWithRunLogInitiator() models.JobSpec {
-	j := NewJob()
-	j.Initiators = []models.Initiator{{
-		JobSpecID: j.ID,
-		Type:      models.InitiatorRunLog,
-		InitiatorParams: models.InitiatorParams{
-			Address: NewAddress(),
-		},
-	}}
-	return j
-}
-
-// NewJobWithRunAtInitiator create new Job with RunAt initiator
-func NewJobWithRunAtInitiator(t time.Time) models.JobSpec {
-	j := NewJob()
-	j.Initiators = []models.Initiator{{
-		JobSpecID: j.ID,
-		Type:      models.InitiatorRunAt,
-		InitiatorParams: models.InitiatorParams{
-			Time: models.NewAnyTime(t),
-		},
-	}}
-	return j
-}
-
-// NewJobWithFluxMonitorInitiator create new Job with FluxMonitor initiator
-func NewJobWithFluxMonitorInitiator() models.JobSpec {
-	j := NewJob()
-	j.Initiators = []models.Initiator{{
-		JobSpecID: j.ID,
-		Type:      models.InitiatorFluxMonitor,
-		InitiatorParams: models.InitiatorParams{
-			Address:           NewAddress(),
-			RequestData:       models.JSON{Result: gjson.Parse(`{"data":{"coin":"ETH","market":"USD"}}`)},
-			Feeds:             models.JSON{Result: gjson.Parse(`["https://lambda.staging.devnet.tools/bnc/call"]`)},
-			Threshold:         0.5,
-			AbsoluteThreshold: 0.01,
-			IdleTimer: models.IdleTimerConfig{
-				Duration: models.MustMakeDuration(time.Minute),
-			},
-			PollTimer: models.PollTimerConfig{
-				Period: models.MustMakeDuration(time.Minute),
-			},
-			Precision: 2,
-		},
-	}}
-	return j
-}
-
-// NewJobWithFluxMonitorInitiator create new Job with FluxMonitor initiator
-func NewJobWithFluxMonitorInitiatorWithBridge() models.JobSpec {
-	j := NewJob()
-	j.Initiators = []models.Initiator{{
-		JobSpecID: j.ID,
-		Type:      models.InitiatorFluxMonitor,
-		InitiatorParams: models.InitiatorParams{
-			Address:           NewAddress(),
-			RequestData:       models.JSON{Result: gjson.Parse(`{"data":{"coin":"ETH","market":"USD"}}`)},
-			Feeds:             models.JSON{Result: gjson.Parse(`[{"bridge":"testbridge"}]`)},
-			Threshold:         0.5,
-			AbsoluteThreshold: 0.01,
-			Precision:         2,
-		},
-	}}
-	return j
-}
-
-// NewJobWithRandomnessLog create new Job with VRF initiator
-func NewJobWithRandomnessLog() models.JobSpec {
-	j := NewJob()
-	j.Initiators = []models.Initiator{{
-		JobSpecID: j.ID,
-		Type:      models.InitiatorRandomnessLog,
-		InitiatorParams: models.InitiatorParams{
-			Address: NewAddress(),
-		},
-	}}
-	return j
-}
-
-// NewTx returns a Tx using a specified from address and sentAt
-func NewTx(from common.Address, sentAt uint64) *models.Tx {
-	tx := &models.Tx{
-		From:     from,
-		Nonce:    0,
-		Data:     []byte{},
-		Value:    utils.NewBig(big.NewInt(0)),
-		GasLimit: 250000,
-		SentAt:   sentAt,
-	}
-	copy(tx.Hash[:], randomBytes(common.HashLength))
-	return tx
-}
-
-func NewTransaction(nonce uint64, sentAtV ...uint64) *models.Tx {
-	from := common.HexToAddress("0xf208000000000000000000000000000000000000")
-	to := common.HexToAddress("0x7000000000000000000000000000000000000000")
-
-	value := new(big.Int).Exp(big.NewInt(10), big.NewInt(36), nil)
-	gasLimit := uint64(50000)
-	data := hexutil.MustDecode("0xda7ada7a")
-
-	sentAt := uint64(0)
-	if len(sentAtV) > 0 {
-		sentAt = sentAtV[0]
-	}
-
-	transaction := types.NewTransaction(nonce, to, value, gasLimit, new(big.Int), data)
-	return &models.Tx{
-		From:        from,
-		SentAt:      sentAt,
-		To:          *transaction.To(),
-		Nonce:       transaction.Nonce(),
-		Data:        transaction.Data(),
-		Value:       utils.NewBig(transaction.Value()),
-		GasLimit:    transaction.Gas(),
-		GasPrice:    utils.NewBig(transaction.GasPrice()),
-		Hash:        transaction.Hash(),
-		SignedRawTx: hexutil.MustDecode("0xcafe11"),
-	}
-}
-
-// CreateTx creates a Tx from a specified address, and sentAt
-func CreateTx(
-	t testing.TB,
-	store *strpkg.Store,
-	from common.Address,
-	sentAt uint64,
-) *models.Tx {
-	return CreateTxWithNonceAndGasPrice(t, store, from, sentAt, 0, 1)
-}
-
-// CreateTxWithNonceAndGasPrice creates a Tx from a specified address, sentAt, nonce and gas price
-func CreateTxWithNonceAndGasPrice(
-	t testing.TB,
-	store *strpkg.Store,
-	from common.Address,
-	sentAt uint64,
-	nonce uint64,
-	gasPrice int64,
-) *models.Tx {
-	return CreateTxWithNonceGasPriceAndRecipient(t, store, from, common.Address{}, sentAt, nonce, gasPrice)
-}
-
-// CreateTxWithNonceGasPriceAndRecipient creates a Tx from a specified sender, recipient, sentAt, nonce and gas price
-func CreateTxWithNonceGasPriceAndRecipient(
-	t testing.TB,
-	store *strpkg.Store,
-	from common.Address,
-	to common.Address,
-	sentAt uint64,
-	nonce uint64,
-	gasPrice int64,
-) *models.Tx {
-	data := make([]byte, 36)
-	binary.LittleEndian.PutUint64(data, sentAt)
-
-	transaction := types.NewTransaction(nonce, to, big.NewInt(0), 250000, big.NewInt(gasPrice), data)
-	tx := &models.Tx{
-		From:        from,
-		SentAt:      sentAt,
-		To:          *transaction.To(),
-		Nonce:       transaction.Nonce(),
-		Data:        transaction.Data(),
-		Value:       utils.NewBig(transaction.Value()),
-		GasLimit:    transaction.Gas(),
-		GasPrice:    utils.NewBig(transaction.GasPrice()),
-		Hash:        transaction.Hash(),
-		SignedRawTx: hexutil.MustDecode("0xcafe22"),
-	}
-
-	tx, err := store.CreateTx(tx)
-	require.NoError(t, err)
-	_, err = store.AddTxAttempt(tx, tx)
-	require.NoError(t, err)
-	return tx
-}
-
-func AddTxAttempt(
-	t testing.TB,
-	store *strpkg.Store,
-	tx *models.Tx,
-	etx *types.Transaction,
-	blkNum uint64,
-) *models.TxAttempt {
-	transaction := types.NewTransaction(tx.Nonce, common.Address{}, big.NewInt(0), 250000, big.NewInt(1), tx.Data)
-
-	newTxAttempt := &models.Tx{
-		From:        tx.From,
-		SentAt:      blkNum,
-		To:          *transaction.To(),
-		Nonce:       transaction.Nonce(),
-		Data:        transaction.Data(),
-		Value:       utils.NewBig(transaction.Value()),
-		GasLimit:    transaction.Gas(),
-		GasPrice:    utils.NewBig(transaction.GasPrice()),
-		Hash:        transaction.Hash(),
-		SignedRawTx: []byte{byte(len(tx.Attempts))},
-	}
-
-	txAttempt, err := store.AddTxAttempt(tx, newTxAttempt)
-	require.NoError(t, err)
-	return txAttempt
-}
-
-// NewHash return random Keccak256
-func NewHash() common.Hash {
-	return common.BytesToHash(randomBytes(32))
-}
-
-// NewAddress return a random new address
-func NewAddress() common.Address {
-	return common.BytesToAddress(randomBytes(20))
-}
-
-func randomBytes(n int) []byte {
-	b := make([]byte, n)
-	rand.Read(b)
-	return b
+type BridgeOpts struct {
+	Name string
+	URL  string
 }
 
 // NewBridgeType create new bridge type given info slice
-func NewBridgeType(t testing.TB, info ...string) (*models.BridgeTypeAuthentication, *models.BridgeType) {
-	btr := &models.BridgeTypeRequest{}
+func NewBridgeType(t testing.TB, opts BridgeOpts) (*bridges.BridgeTypeAuthentication, *bridges.BridgeType) {
+	btr := &bridges.BridgeTypeRequest{}
 
-	if len(info) > 0 {
-		btr.Name = models.MustNewTaskType(info[0])
+	// Must randomise default to avoid unique constraint conflicts with other parallel tests
+	rnd := uuid.New().String()
+
+	if opts.Name != "" {
+		btr.Name = bridges.MustParseBridgeName(opts.Name)
 	} else {
-		btr.Name = models.MustNewTaskType("defaultFixtureBridgeType")
+		btr.Name = bridges.MustParseBridgeName(fmt.Sprintf("test_bridge_%s", rnd))
 	}
 
-	if len(info) > 1 {
-		btr.URL = WebURL(t, info[1])
+	if opts.URL != "" {
+		btr.URL = WebURL(t, opts.URL)
 	} else {
-		btr.URL = WebURL(t, "https://bridge.example.com/api")
+		btr.URL = WebURL(t, fmt.Sprintf("https://bridge.example.com/api?%s", rnd))
 	}
 
-	bta, bt, err := models.NewBridgeType(btr)
+	bta, bt, err := bridges.NewBridgeType(btr)
+	require.NoError(t, err)
+	return bta, bt
+}
+
+// MustCreateBridge creates a bridge
+// Be careful not to specify a name here unless you ABSOLUTELY need to
+// This is because name is a unique index and identical names used across transactional tests will lock/deadlock
+func MustCreateBridge(t testing.TB, ds sqlutil.DataSource, opts BridgeOpts) (bta *bridges.BridgeTypeAuthentication, bt *bridges.BridgeType) {
+	bta, bt = NewBridgeType(t, opts)
+	orm := bridges.NewORM(ds)
+	err := orm.CreateBridgeType(testutils.Context(t), bt)
 	require.NoError(t, err)
 	return bta, bt
 }
@@ -365,169 +125,11 @@ func JSONFromBytes(t testing.TB, body []byte) models.JSON {
 	return j
 }
 
-// MustJSONSet uses sjson.Set to set a path in a JSON string and returns the string
-// See https://github.com/tidwall/sjson
-func MustJSONSet(t *testing.T, json, path string, value interface{}) string {
-	json, err := sjson.Set(json, path, value)
+func MustJSONMarshal(t *testing.T, val interface{}) string {
+	t.Helper()
+	bs, err := json.Marshal(val)
 	require.NoError(t, err)
-	return json
-}
-
-// MustJSONDel uses sjson.Delete to remove a path from a JSON string and returns the string
-func MustJSONDel(t *testing.T, json, path string) string {
-	json, err := sjson.Delete(json, path)
-	require.NoError(t, err)
-	return json
-}
-
-// NewRunLog create models.Log for given jobid, address, block, and json
-func NewRunLog(
-	t *testing.T,
-	jobID *models.ID,
-	emitter common.Address,
-	requester common.Address,
-	blk int,
-	json string,
-) models.Log {
-	return models.Log{
-		Address:     emitter,
-		BlockNumber: uint64(blk),
-		Data:        StringToVersionedLogData20190207withoutIndexes(t, "internalID", requester, json),
-		TxHash:      NewHash(),
-		BlockHash:   NewHash(),
-		Topics: []common.Hash{
-			models.RunLogTopic20190207withoutIndexes,
-			models.IDToTopic(jobID),
-		},
-	}
-}
-
-// NewRandomnessRequestLog(t, r, emitter, blk) is a RandomnessRequest log for
-// the randomness request log represented by r.
-func NewRandomnessRequestLog(t *testing.T, r models.RandomnessRequestLog,
-	emitter common.Address, blk int) models.Log {
-	rawData, err := r.RawData()
-	require.NoError(t, err)
-	return models.Log{
-		Address:     emitter,
-		BlockNumber: uint64(blk),
-		Data:        rawData,
-		TxHash:      NewHash(),
-		BlockHash:   NewHash(),
-		Topics:      []common.Hash{models.RandomnessRequestLogTopic, r.JobID},
-	}
-}
-
-func NewLink(t *testing.T, amount string) *assets.Link {
-	link := assets.NewLink(0)
-	link, ok := link.SetString(amount, 10)
-	assert.True(t, ok)
-	return link
-}
-
-func NewEth(t *testing.T, amount string) *assets.Eth {
-	eth := assets.NewEth(0)
-	eth, ok := eth.SetString(amount, 10)
-	assert.True(t, ok)
-	return eth
-}
-
-func StringToVersionedLogData20190207withoutIndexes(
-	t *testing.T,
-	internalID string,
-	requester common.Address,
-	str string,
-) []byte {
-	requesterBytes := requester.Hash().Bytes()
-	buf := bytes.NewBuffer(requesterBytes)
-
-	requestID := hexutil.MustDecode(StringToHash(internalID).Hex())
-	buf.Write(requestID)
-
-	payment := hexutil.MustDecode(minimumContractPayment.ToHash().Hex())
-	buf.Write(payment)
-
-	callbackAddr := utils.EVMWordUint64(0)
-	buf.Write(callbackAddr)
-
-	callbackFunc := utils.EVMWordUint64(0)
-	buf.Write(callbackFunc)
-
-	expiration := utils.EVMWordUint64(4000000000)
-	buf.Write(expiration)
-
-	version := utils.EVMWordUint64(1)
-	buf.Write(version)
-
-	dataLocation := utils.EVMWordUint64(common.HashLength * 8)
-	buf.Write(dataLocation)
-
-	cbor, err := JSONFromString(t, str).CBOR()
-	require.NoError(t, err)
-	buf.Write(utils.EVMWordUint64(uint64(len(cbor))))
-	paddedLength := common.HashLength * ((len(cbor) / common.HashLength) + 1)
-	buf.Write(common.RightPadBytes(cbor, paddedLength))
-
-	return buf.Bytes()
-}
-
-// BigHexInt create hexutil.Big value from given value
-func BigHexInt(val interface{}) hexutil.Big {
-	switch x := val.(type) {
-	case int: // Single case allows compiler to narrow x's type.
-		return hexutil.Big(*big.NewInt(int64(x)))
-	case uint32:
-		return hexutil.Big(*big.NewInt(int64(x)))
-	case uint64:
-		return hexutil.Big(*big.NewInt(0).SetUint64(x))
-	case int64:
-		return hexutil.Big(*big.NewInt(x))
-	default:
-		logger.Panicf("Could not convert %v of type %T to hexutil.Big", val, val)
-		return hexutil.Big{}
-	}
-}
-
-func Int(val interface{}) *utils.Big {
-	switch x := val.(type) {
-	case int:
-		return (*utils.Big)(big.NewInt(int64(x)))
-	case uint32:
-		return (*utils.Big)(big.NewInt(int64(x)))
-	case uint64:
-		return (*utils.Big)(big.NewInt(int64(x)))
-	case int64:
-		return (*utils.Big)(big.NewInt(x))
-	default:
-		logger.Panicf("Could not convert %v of type %T to utils.Big", val, val)
-		return &utils.Big{}
-	}
-}
-
-func MustEVMUintHexFromBase10String(t *testing.T, strings ...string) string {
-	var allBytes []byte
-	for _, s := range strings {
-		i, ok := big.NewInt(0).SetString(s, 10)
-		require.True(t, ok)
-		bs, err := utils.EVMWordBigInt(i)
-		require.NoError(t, err)
-		allBytes = append(allBytes, bs...)
-	}
-	return fmt.Sprintf("0x%0x", allBytes)
-}
-
-type MockSigner struct{}
-
-func (s MockSigner) SignHash(common.Hash) (models.Signature, error) {
-	return models.NewSignature("0xb7a987222fc36c4c8ed1b91264867a422769998aadbeeb1c697586a04fa2b616025b5ca936ec5bdb150999e298b6ecf09251d3c4dd1306dedec0692e7037584800")
-}
-
-func ServiceAgreementFromString(str string) (models.ServiceAgreement, error) {
-	us, err := models.NewUnsignedServiceAgreementFromRequest(strings.NewReader(str))
-	if err != nil {
-		return models.ServiceAgreement{}, err
-	}
-	return models.BuildServiceAgreement(us, MockSigner{})
+	return string(bs)
 }
 
 func EmptyCLIContext() *cli.Context {
@@ -535,244 +137,434 @@ func EmptyCLIContext() *cli.Context {
 	return cli.NewContext(nil, set, nil)
 }
 
-// NewJobRun returns a newly initialized job run for test purposes only
-func NewJobRun(job models.JobSpec) models.JobRun {
-	initiator := job.Initiators[0]
-	now := time.Now()
-	run := models.MakeJobRun(&job, now, &initiator, nil, &models.RunRequest{})
-	return run
-}
-
-// NewJobRunPendingBridge returns a new job run in the pending bridge state
-func NewJobRunPendingBridge(job models.JobSpec) models.JobRun {
-	run := NewJobRun(job)
-	run.SetStatus(models.RunStatusPendingBridge)
-	run.TaskRuns[0].Status = models.RunStatusPendingBridge
-	return run
-}
-
-// CreateJobRunWithStatus returns a new job run with the specified status that has been persisted
-func CreateJobRunWithStatus(t testing.TB, store *strpkg.Store, job models.JobSpec, status models.RunStatus) models.JobRun {
-	run := NewJobRun(job)
-	run.SetStatus(status)
-	require.NoError(t, store.CreateJobRun(&run))
-	return run
-}
-
-func BuildInitiatorRequests(t *testing.T, initrs []models.Initiator) []models.InitiatorRequest {
-	bytes, err := json.Marshal(initrs)
-	require.NoError(t, err)
-
-	var dst []models.InitiatorRequest
-	err = json.Unmarshal(bytes, &dst)
-	require.NoError(t, err)
-	return dst
-}
-
-func BuildTaskRequests(t *testing.T, initrs []models.TaskSpec) []models.TaskSpecRequest {
-	bytes, err := json.Marshal(initrs)
-	require.NoError(t, err)
-
-	var dst []models.TaskSpecRequest
-	err = json.Unmarshal(bytes, &dst)
-	require.NoError(t, err)
-	return dst
-}
-
-func NewRunInput(value models.JSON) models.RunInput {
-	jobRunID := models.NewID()
-	taskRunID := models.NewID()
-	return *models.NewRunInput(jobRunID, *taskRunID, value, models.RunStatusUnstarted)
-}
-
-func NewRunInputWithString(t testing.TB, value string) models.RunInput {
-	jobRunID := models.NewID()
-	taskRunID := models.NewID()
-	data := JSONFromString(t, value)
-	return *models.NewRunInput(jobRunID, *taskRunID, data, models.RunStatusUnstarted)
-}
-
-func NewRunInputWithResult(value interface{}) models.RunInput {
-	jobRunID := models.NewID()
-	taskRunID := models.NewID()
-	return *models.NewRunInputWithResult(jobRunID, *taskRunID, value, models.RunStatusUnstarted)
-}
-
-func NewRunInputWithResultAndJobRunID(value interface{}, jobRunID *models.ID) models.RunInput {
-	taskRunID := models.NewID()
-	return *models.NewRunInputWithResult(jobRunID, *taskRunID, value, models.RunStatusUnstarted)
-}
-
-func NewPollingDeviationChecker(t *testing.T, s *strpkg.Store) *fluxmonitor.PollingDeviationChecker {
-	fluxAggregator := new(mocks.FluxAggregator)
-	runManager := new(mocks.RunManager)
-	fetcher := new(mocks.Fetcher)
-	initr := models.Initiator{
-		JobSpecID: models.NewID(),
-		InitiatorParams: models.InitiatorParams{
-			PollTimer: models.PollTimerConfig{
-				Period: models.MustMakeDuration(time.Second),
-			},
-		},
-	}
-	checker, err := fluxmonitor.NewPollingDeviationChecker(s, fluxAggregator, initr, nil, runManager, fetcher, func() {})
-	require.NoError(t, err)
-	return checker
-}
-
-func NewGethHeader(height interface{}) *types.Header {
-	var h int64
-	switch v := height.(type) {
-	case int64:
-		h = v
-	case int:
-		h = int64(v)
-	case uint64:
-		h = int64(v)
-	default:
-		panic(fmt.Sprintf("invalid type: %t", height))
-	}
-
-	return &types.Header{
-		Number:     big.NewInt(h),
-		ParentHash: NewHash(),
-		Time:       uint64(time.Now().Unix()),
-	}
-}
-
-func NewHeadFromGethHeader(gethHeader *types.Header) *models.Head {
-	return &models.Head{
-		Hash:       NewHash(),
-		Number:     gethHeader.Number.Int64(),
-		ParentHash: gethHeader.ParentHash,
-		Timestamp:  time.Unix(int64(gethHeader.Time), 0),
-	}
-}
-
-func MustInsertTaskRun(t *testing.T, store *strpkg.Store) models.ID {
-	taskRunID := models.NewID()
-
-	job := NewJobWithWebInitiator()
-	require.NoError(t, store.CreateJob(&job))
-	jobRun := NewJobRun(job)
-	jobRun.TaskRuns = []models.TaskRun{models.TaskRun{ID: taskRunID, Status: models.RunStatusUnstarted, TaskSpecID: job.Tasks[0].ID}}
-	require.NoError(t, store.CreateJobRun(&jobRun))
-
-	return *taskRunID
-}
-
-func MustInsertKey(t *testing.T, store *strpkg.Store, address common.Address) models.Key {
-	a, err := models.NewEIP55Address(address.Hex())
-	require.NoError(t, err)
-	key := models.Key{
-		Address: a,
-		JSON:    JSONFromString(t, "{}"),
-	}
-	require.NoError(t, store.DB.Save(&key).Error)
-	return key
-}
-
-func NewEthTx(t *testing.T, store *strpkg.Store) models.EthTx {
-	return models.EthTx{
-		FromAddress:    GetDefaultFromAddress(t, store),
-		ToAddress:      NewAddress(),
+func NewEthTx(fromAddress common.Address) txmgr.Tx {
+	return txmgr.Tx{
+		FromAddress:    fromAddress,
+		ToAddress:      testutils.NewAddress(),
 		EncodedPayload: []byte{1, 2, 3},
-		Value:          assets.NewEthValue(142),
-		GasLimit:       uint64(1000000000),
+		Value:          big.Int(assets.NewEthValue(142)),
+		FeeLimit:       uint64(1000000000),
+		State:          txmgrcommon.TxUnstarted,
 	}
 }
 
-func MustInsertUnconfirmedEthTxWithBroadcastAttempt(t *testing.T, store *strpkg.Store, nonce int64) models.EthTx {
-	timeNow := time.Now()
-	etx := NewEthTx(t, store)
+func NewLegacyTransaction(nonce uint64, to common.Address, value *big.Int, gasLimit uint32, gasPrice *big.Int, data []byte) *types.Transaction {
+	tx := types.LegacyTx{
+		Nonce:    nonce,
+		To:       &to,
+		Value:    value,
+		Gas:      uint64(gasLimit),
+		GasPrice: gasPrice,
+		Data:     data,
+	}
+	return types.NewTx(&tx)
+}
 
-	etx.BroadcastAt = &timeNow
-	n := nonce
-	etx.Nonce = &n
-	etx.State = models.EthTxUnconfirmed
-	require.NoError(t, store.DB.Save(&etx).Error)
-	attempt := NewEthTxAttempt(t, etx.ID)
+func MustInsertUnconfirmedEthTx(t *testing.T, txStore txmgr.TestEvmTxStore, nonce int64, fromAddress common.Address, opts ...interface{}) txmgr.Tx {
+	broadcastAt := time.Now()
+	chainID := &FixtureChainID
+	for _, opt := range opts {
+		switch v := opt.(type) {
+		case time.Time:
+			broadcastAt = v
+		case *big.Int:
+			chainID = v
+		}
+	}
+	etx := NewEthTx(fromAddress)
 
-	tx := types.NewTransaction(uint64(nonce), NewAddress(), big.NewInt(142), 242, big.NewInt(342), []byte{1, 2, 3})
+	etx.BroadcastAt = &broadcastAt
+	etx.InitialBroadcastAt = &broadcastAt
+	n := evmtypes.Nonce(nonce)
+	etx.Sequence = &n
+	etx.State = txmgrcommon.TxUnconfirmed
+	etx.ChainID = chainID
+	require.NoError(t, txStore.InsertTx(testutils.Context(t), &etx))
+	return etx
+}
+
+func MustInsertUnconfirmedEthTxWithBroadcastLegacyAttempt(t *testing.T, txStore txmgr.TestEvmTxStore, nonce int64, fromAddress common.Address, opts ...interface{}) txmgr.Tx {
+	etx := MustInsertUnconfirmedEthTx(t, txStore, nonce, fromAddress, opts...)
+	attempt := NewLegacyEthTxAttempt(t, etx.ID)
+	ctx := testutils.Context(t)
+
+	tx := NewLegacyTransaction(uint64(nonce), testutils.NewAddress(), big.NewInt(142), 242, big.NewInt(342), []byte{1, 2, 3})
 	rlp := new(bytes.Buffer)
 	require.NoError(t, tx.EncodeRLP(rlp))
 	attempt.SignedRawTx = rlp.Bytes()
 
-	attempt.State = models.EthTxAttemptBroadcast
-	require.NoError(t, store.DB.Save(&attempt).Error)
-	etx, err := store.FindEthTxWithAttempts(etx.ID)
+	attempt.State = txmgrtypes.TxAttemptBroadcast
+	require.NoError(t, txStore.InsertTxAttempt(ctx, &attempt))
+	etx, err := txStore.FindTxWithAttempts(ctx, etx.ID)
 	require.NoError(t, err)
 	return etx
 }
 
-func MustInsertConfirmedEthTxWithAttempt(t *testing.T, store *strpkg.Store, nonce int64, broadcastBeforeBlockNum int64) models.EthTx {
+func MustInsertConfirmedEthTxWithLegacyAttempt(t *testing.T, txStore txmgr.TestEvmTxStore, nonce int64, broadcastBeforeBlockNum int64, fromAddress common.Address) txmgr.Tx {
 	timeNow := time.Now()
-	etx := NewEthTx(t, store)
+	etx := NewEthTx(fromAddress)
+	ctx := testutils.Context(t)
 
 	etx.BroadcastAt = &timeNow
-	etx.Nonce = &nonce
-	etx.State = models.EthTxConfirmed
-	require.NoError(t, store.DB.Save(&etx).Error)
-	attempt := NewEthTxAttempt(t, etx.ID)
+	etx.InitialBroadcastAt = &timeNow
+	n := evmtypes.Nonce(nonce)
+	etx.Sequence = &n
+	etx.State = txmgrcommon.TxConfirmed
+	etx.MinConfirmations.SetValid(6)
+	require.NoError(t, txStore.InsertTx(ctx, &etx))
+	attempt := NewLegacyEthTxAttempt(t, etx.ID)
 	attempt.BroadcastBeforeBlockNum = &broadcastBeforeBlockNum
-	attempt.State = models.EthTxAttemptBroadcast
-	require.NoError(t, store.DB.Save(&attempt).Error)
-	etx.EthTxAttempts = append(etx.EthTxAttempts, attempt)
+	attempt.State = txmgrtypes.TxAttemptBroadcast
+	require.NoError(t, txStore.InsertTxAttempt(ctx, &attempt))
+	etx.TxAttempts = append(etx.TxAttempts, attempt)
 	return etx
 }
 
-func GetDefaultFromAddress(t *testing.T, store *strpkg.Store) common.Address {
-	keys, err := store.Keys()
-	require.NoError(t, err)
-	require.Len(t, keys, 1)
-	key := keys[0]
-	return key.Address.Address()
-}
-
-func NewEthTxAttempt(t *testing.T, etxID int64) models.EthTxAttempt {
-	gasPrice := utils.NewBig(big.NewInt(1))
-	return models.EthTxAttempt{
-		EthTxID:  etxID,
-		GasPrice: *gasPrice,
+func NewLegacyEthTxAttempt(t *testing.T, etxID int64) txmgr.TxAttempt {
+	gasPrice := assets.NewWeiI(1)
+	return txmgr.TxAttempt{
+		ChainSpecificFeeLimit: 42,
+		TxID:                  etxID,
+		TxFee:                 gas.EvmFee{Legacy: gasPrice},
 		// Just a random signed raw tx that decodes correctly
 		// Ignore all actual values
 		SignedRawTx: hexutil.MustDecode("0xf889808504a817c8008307a12094000000000000000000000000000000000000000080a400000000000000000000000000000000000000000000000000000000000000000000000025a0838fe165906e2547b9a052c099df08ec891813fea4fcdb3c555362285eb399c5a070db99322490eb8a0f2270be6eca6e3aedbc49ff57ef939cf2774f12d08aa85e"),
-		Hash:        NewHash(),
+		Hash:        evmutils.NewHash(),
+		State:       txmgrtypes.TxAttemptInProgress,
 	}
 }
 
-func MustInsertBroadcastEthTxAttempt(t *testing.T, etxID int64, store *strpkg.Store, gasPrice int64) models.EthTxAttempt {
-	attempt := NewEthTxAttempt(t, etxID)
-	attempt.State = models.EthTxAttemptBroadcast
-	attempt.GasPrice = *utils.NewBig(big.NewInt(gasPrice))
-	require.NoError(t, store.DB.Create(&attempt).Error)
-	return attempt
-}
-
-func MustInsertEthReceipt(t *testing.T, s *strpkg.Store, blockNumber int64, blockHash common.Hash, txHash common.Hash) models.EthReceipt {
-	r := models.EthReceipt{
-		BlockNumber:      blockNumber,
-		BlockHash:        blockHash,
-		TxHash:           txHash,
-		TransactionIndex: uint(NewRandomInt64()),
-		Receipt:          []byte(`{"foo":42}`),
+func NewDynamicFeeEthTxAttempt(t *testing.T, etxID int64) txmgr.TxAttempt {
+	gasTipCap := assets.NewWeiI(1)
+	gasFeeCap := assets.NewWeiI(1)
+	return txmgr.TxAttempt{
+		TxType: 0x2,
+		TxID:   etxID,
+		TxFee: gas.EvmFee{
+			DynamicTipCap: gasTipCap,
+			DynamicFeeCap: gasFeeCap,
+		},
+		// Just a random signed raw tx that decodes correctly
+		// Ignore all actual values
+		SignedRawTx:           hexutil.MustDecode("0xf889808504a817c8008307a12094000000000000000000000000000000000000000080a400000000000000000000000000000000000000000000000000000000000000000000000025a0838fe165906e2547b9a052c099df08ec891813fea4fcdb3c555362285eb399c5a070db99322490eb8a0f2270be6eca6e3aedbc49ff57ef939cf2774f12d08aa85e"),
+		Hash:                  evmutils.NewHash(),
+		State:                 txmgrtypes.TxAttemptInProgress,
+		ChainSpecificFeeLimit: 42,
 	}
-	require.NoError(t, s.DB.Save(&r).Error)
-	return r
 }
 
-func MustInsertFatalErrorEthTx(t *testing.T, store *strpkg.Store) models.EthTx {
-	etx := NewEthTx(t, store)
-	errStr := "something exploded"
-	etx.Error = &errStr
-	etx.State = models.EthTxFatalError
+type RandomKey struct {
+	Nonce    int64
+	Disabled bool
 
-	require.NoError(t, store.DB.Save(&etx).Error)
-	return etx
+	chainIDs []ubig.Big // nil: Fixture, set empty for none
 }
 
-func MustInsertRandomKey(t *testing.T, store *strpkg.Store) models.Key {
-	k := models.Key{Address: models.EIP55Address(NewAddress().Hex()), JSON: JSONFromString(t, `{"key": "factory"}`)}
-	require.NoError(t, store.CreateKeyIfNotExists(k))
-	return k
+func (r RandomKey) MustInsert(t testing.TB, keystore keystore.Eth) (ethkey.KeyV2, common.Address) {
+	ctx := testutils.Context(t)
+	if r.chainIDs == nil {
+		r.chainIDs = []ubig.Big{*ubig.New(&FixtureChainID)}
+	}
+
+	key := MustGenerateRandomKey(t)
+	keystore.XXXTestingOnlyAdd(ctx, key)
+
+	for _, cid := range r.chainIDs {
+		require.NoError(t, keystore.Add(ctx, key.Address, cid.ToInt()))
+		require.NoError(t, keystore.Enable(ctx, key.Address, cid.ToInt()))
+		if r.Disabled {
+			require.NoError(t, keystore.Disable(ctx, key.Address, cid.ToInt()))
+		}
+	}
+
+	return key, key.Address
+}
+
+func (r RandomKey) MustInsertWithState(t testing.TB, keystore keystore.Eth) (ethkey.State, common.Address) {
+	ctx := testutils.Context(t)
+	k, address := r.MustInsert(t, keystore)
+	state, err := keystore.GetStateForKey(ctx, k)
+	require.NoError(t, err)
+	return state, address
+}
+
+// MustInsertRandomKey inserts a randomly generated (not cryptographically secure) key for testing.
+// By default, it is enabled for the fixture chain. Pass chainIDs to override.
+// Use MustInsertRandomKeyNoChains for a key associate with no chains.
+func MustInsertRandomKey(t testing.TB, keystore keystore.Eth, chainIDs ...ubig.Big) (ethkey.KeyV2, common.Address) {
+	r := RandomKey{}
+	if len(chainIDs) > 0 {
+		r.chainIDs = chainIDs
+	}
+	return r.MustInsert(t, keystore)
+}
+
+func MustInsertRandomKeyNoChains(t testing.TB, keystore keystore.Eth) (ethkey.KeyV2, common.Address) {
+	return RandomKey{chainIDs: []ubig.Big{}}.MustInsert(t, keystore)
+}
+
+func MustInsertRandomKeyReturningState(t testing.TB, keystore keystore.Eth) (ethkey.State, common.Address) {
+	return RandomKey{}.MustInsertWithState(t, keystore)
+}
+
+func MustGenerateRandomKey(t testing.TB) ethkey.KeyV2 {
+	key, err := ethkey.NewV2()
+	require.NoError(t, err)
+	return key
+}
+
+func MustGenerateRandomKeyState(_ testing.TB) ethkey.State {
+	return ethkey.State{Address: NewEIP55Address()}
+}
+
+func MustInsertHead(t *testing.T, ds sqlutil.DataSource, number int64) evmtypes.Head {
+	h := evmtypes.NewHead(big.NewInt(number), evmutils.NewHash(), evmutils.NewHash(), 0, ubig.New(&FixtureChainID))
+	horm := headtracker.NewORM(FixtureChainID, ds)
+
+	err := horm.IdempotentInsertHead(testutils.Context(t), &h)
+	require.NoError(t, err)
+	return h
+}
+
+func MustInsertV2JobSpec(t *testing.T, db *sqlx.DB, transmitterAddress common.Address) job.Job {
+	t.Helper()
+
+	addr, err := evmtypes.NewEIP55Address(transmitterAddress.Hex())
+	require.NoError(t, err)
+
+	pipelineSpec := pipeline.Spec{}
+	err = db.Get(&pipelineSpec, `INSERT INTO pipeline_specs (dot_dag_source,created_at) VALUES ('',NOW()) RETURNING *`)
+	require.NoError(t, err)
+
+	oracleSpec := MustInsertOffchainreportingOracleSpec(t, db, addr)
+	jb := job.Job{
+		OCROracleSpec:   &oracleSpec,
+		OCROracleSpecID: &oracleSpec.ID,
+		ExternalJobID:   uuid.New(),
+		Type:            job.OffchainReporting,
+		SchemaVersion:   1,
+		PipelineSpec:    &pipelineSpec,
+		PipelineSpecID:  pipelineSpec.ID,
+	}
+
+	jorm := job.NewORM(db, nil, nil, nil, logger.TestLogger(t))
+	err = jorm.InsertJob(testutils.Context(t), &jb)
+	require.NoError(t, err)
+	return jb
+}
+
+func MustInsertOffchainreportingOracleSpec(t *testing.T, db *sqlx.DB, transmitterAddress evmtypes.EIP55Address) job.OCROracleSpec {
+	t.Helper()
+
+	ocrKeyID := models.MustSha256HashFromHex(DefaultOCRKeyBundleID)
+	spec := job.OCROracleSpec{}
+	require.NoError(t, db.Get(&spec, `INSERT INTO ocr_oracle_specs (created_at, updated_at, contract_address, p2pv2_bootstrappers, is_bootstrap_peer, encrypted_ocr_key_bundle_id, transmitter_address, observation_timeout, blockchain_timeout, contract_config_tracker_subscribe_interval, contract_config_tracker_poll_interval, contract_config_confirmations, database_timeout, observation_grace_period, contract_transmitter_transmit_timeout, evm_chain_id) VALUES (
+NOW(),NOW(),$1,'{}',false,$2,$3,0,0,0,0,0,0,0,0,0
+) RETURNING *`, NewEIP55Address(), &ocrKeyID, &transmitterAddress))
+	return spec
+}
+
+func MakeDirectRequestJobSpec(t *testing.T) *job.Job {
+	t.Helper()
+	drs := &job.DirectRequestSpec{EVMChainID: (*ubig.Big)(testutils.FixtureChainID)}
+	spec := &job.Job{
+		Type:              job.DirectRequest,
+		SchemaVersion:     1,
+		ExternalJobID:     uuid.New(),
+		DirectRequestSpec: drs,
+		Pipeline:          pipeline.Pipeline{},
+		PipelineSpec:      &pipeline.Spec{},
+	}
+	return spec
+}
+
+func MustInsertKeeperJob(t *testing.T, db *sqlx.DB, korm *keeper.ORM, from evmtypes.EIP55Address, contract evmtypes.EIP55Address) job.Job {
+	t.Helper()
+	ctx := testutils.Context(t)
+
+	var keeperSpec job.KeeperSpec
+	err := korm.DataSource().GetContext(ctx, &keeperSpec, `INSERT INTO keeper_specs (contract_address, from_address, created_at, updated_at,evm_chain_id) VALUES ($1, $2, NOW(), NOW(), $3) RETURNING *`, contract, from, testutils.SimulatedChainID.Int64())
+	require.NoError(t, err)
+
+	var pipelineSpec pipeline.Spec
+	err = korm.DataSource().GetContext(ctx, &pipelineSpec, `INSERT INTO pipeline_specs (dot_dag_source,created_at) VALUES ('',NOW()) RETURNING *`)
+	require.NoError(t, err)
+
+	jb := job.Job{
+		KeeperSpec:     &keeperSpec,
+		KeeperSpecID:   &keeperSpec.ID,
+		ExternalJobID:  uuid.New(),
+		Type:           job.Keeper,
+		SchemaVersion:  1,
+		PipelineSpec:   &pipelineSpec,
+		PipelineSpecID: pipelineSpec.ID,
+	}
+
+	cfg := configtest.NewTestGeneralConfig(t)
+	tlg := logger.TestLogger(t)
+	prm := pipeline.NewORM(db, tlg, cfg.JobPipeline().MaxSuccessfulRuns())
+	btORM := bridges.NewORM(db)
+	jrm := job.NewORM(db, prm, btORM, nil, tlg)
+	err = jrm.InsertJob(testutils.Context(t), &jb)
+	require.NoError(t, err)
+	jb.PipelineSpec.JobID = jb.ID
+	return jb
+}
+
+func MustInsertKeeperRegistry(t *testing.T, db *sqlx.DB, korm *keeper.ORM, ethKeyStore keystore.Eth, keeperIndex, numKeepers, blockCountPerTurn int32) (keeper.Registry, job.Job) {
+	t.Helper()
+	ctx := testutils.Context(t)
+	key, _ := MustInsertRandomKey(t, ethKeyStore, *ubig.New(testutils.SimulatedChainID))
+	from := key.EIP55Address
+	contractAddress := NewEIP55Address()
+	jb := MustInsertKeeperJob(t, db, korm, from, contractAddress)
+	registry := keeper.Registry{
+		ContractAddress:   contractAddress,
+		BlockCountPerTurn: blockCountPerTurn,
+		CheckGas:          150_000,
+		FromAddress:       from,
+		JobID:             jb.ID,
+		KeeperIndex:       keeperIndex,
+		NumKeepers:        numKeepers,
+		KeeperIndexMap: map[evmtypes.EIP55Address]int32{
+			from: keeperIndex,
+		},
+	}
+	err := korm.UpsertRegistry(ctx, &registry)
+	require.NoError(t, err)
+	return registry, jb
+}
+
+func MustInsertUpkeepForRegistry(t *testing.T, db *sqlx.DB, registry keeper.Registry) keeper.UpkeepRegistration {
+	ctx := testutils.Context(t)
+	korm := keeper.NewORM(db, logger.TestLogger(t))
+	upkeepID := ubig.NewI(int64(mathrand.Uint32()))
+	upkeep := keeper.UpkeepRegistration{
+		UpkeepID:   upkeepID,
+		ExecuteGas: uint32(150_000),
+		Registry:   registry,
+		RegistryID: registry.ID,
+		CheckData:  common.Hex2Bytes("ABC123"),
+	}
+	positioningConstant, err := keeper.CalcPositioningConstant(upkeepID, registry.ContractAddress)
+	require.NoError(t, err)
+	upkeep.PositioningConstant = positioningConstant
+	err = korm.UpsertUpkeep(ctx, &upkeep)
+	require.NoError(t, err)
+	return upkeep
+}
+
+func MustInsertPipelineRun(t *testing.T, db *sqlx.DB) (run pipeline.Run) {
+	require.NoError(t, db.Get(&run, `INSERT INTO pipeline_runs (state,pipeline_spec_id,pruning_key,created_at) VALUES ($1, 0, 0, NOW()) RETURNING *`, pipeline.RunStatusRunning))
+	return run
+}
+
+func MustInsertPipelineRunWithStatus(t *testing.T, db *sqlx.DB, pipelineSpecID int32, status pipeline.RunStatus, jobID int32) (run pipeline.Run) {
+	var finishedAt *time.Time
+	var outputs jsonserializable.JSONSerializable
+	var allErrors pipeline.RunErrors
+	var fatalErrors pipeline.RunErrors
+	now := time.Now()
+	switch status {
+	case pipeline.RunStatusCompleted:
+		finishedAt = &now
+		outputs = jsonserializable.JSONSerializable{
+			Val:   "foo",
+			Valid: true,
+		}
+	case pipeline.RunStatusErrored:
+		finishedAt = &now
+		allErrors = []null.String{null.StringFrom("oh no!")}
+		fatalErrors = []null.String{null.StringFrom("oh no!")}
+	case pipeline.RunStatusRunning, pipeline.RunStatusSuspended:
+		// leave empty
+	default:
+		t.Fatalf("unknown status: %s", status)
+	}
+	require.NoError(t, db.Get(&run, `INSERT INTO pipeline_runs (state,pipeline_spec_id,pruning_key,finished_at,outputs,all_errors,fatal_errors,created_at) VALUES ($1, $2, $3, $4, $5, $6, $7, NOW()) RETURNING *`, status, pipelineSpecID, jobID, finishedAt, outputs, allErrors, fatalErrors))
+	return run
+}
+
+func MustInsertPipelineSpec(t *testing.T, db *sqlx.DB) (spec pipeline.Spec) {
+	err := db.Get(&spec, `INSERT INTO pipeline_specs (dot_dag_source,created_at) VALUES ('',NOW()) RETURNING *`)
+	require.NoError(t, err)
+	return
+}
+
+func MustInsertUnfinishedPipelineTaskRun(t *testing.T, db *sqlx.DB, pipelineRunID int64) (tr pipeline.TaskRun) {
+	/* #nosec G404 */
+	require.NoError(t, db.Get(&tr, `INSERT INTO pipeline_task_runs (dot_id, pipeline_run_id, id, type, created_at) VALUES ($1,$2,$3, '', NOW()) RETURNING *`, strconv.Itoa(mathrand.Int()), pipelineRunID, uuid.New()))
+	return tr
+}
+
+func RandomLog(t *testing.T) types.Log {
+	t.Helper()
+
+	topics := make([]common.Hash, 4)
+	for i := range topics {
+		topics[i] = evmutils.NewHash()
+	}
+
+	return types.Log{
+		Address:     testutils.NewAddress(),
+		BlockHash:   evmutils.NewHash(),
+		BlockNumber: uint64(mathrand.Intn(9999999)),
+		Index:       uint(mathrand.Intn(9999999)),
+		Data:        MustRandomBytes(t, 512),
+		Topics:      []common.Hash{evmutils.NewHash(), evmutils.NewHash(), evmutils.NewHash(), evmutils.NewHash()},
+	}
+}
+
+func RawNewRoundLog(t *testing.T, contractAddr common.Address, blockHash common.Hash, blockNumber uint64, logIndex uint, removed bool) types.Log {
+	t.Helper()
+	topic := (flux_aggregator_wrapper.FluxAggregatorNewRound{}).Topic()
+	topics := []common.Hash{topic, evmutils.NewHash(), evmutils.NewHash()}
+	return RawNewRoundLogWithTopics(t, contractAddr, blockHash, blockNumber, logIndex, removed, topics)
+}
+
+func RawNewRoundLogWithTopics(t *testing.T, contractAddr common.Address, blockHash common.Hash, blockNumber uint64, logIndex uint, removed bool, topics []common.Hash) types.Log {
+	t.Helper()
+	return types.Log{
+		Address:     contractAddr,
+		BlockHash:   blockHash,
+		BlockNumber: blockNumber,
+		Index:       logIndex,
+		Topics:      topics,
+		Data:        []byte("aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"),
+		Removed:     removed,
+	}
+}
+
+func MustInsertExternalInitiator(t *testing.T, orm bridges.ORM) (ei bridges.ExternalInitiator) {
+	return MustInsertExternalInitiatorWithOpts(t, orm, ExternalInitiatorOpts{})
+}
+
+type ExternalInitiatorOpts struct {
+	NamePrefix     string
+	URL            *models.WebURL
+	OutgoingSecret string
+	OutgoingToken  string
+}
+
+func MustInsertExternalInitiatorWithOpts(t *testing.T, orm bridges.ORM, opts ExternalInitiatorOpts) (ei bridges.ExternalInitiator) {
+	ctx := testutils.Context(t)
+	var prefix string
+	if opts.NamePrefix != "" {
+		prefix = opts.NamePrefix
+	} else {
+		prefix = "ei"
+	}
+	ei.Name = fmt.Sprintf("%s-%s", prefix, uuid.New())
+	ei.URL = opts.URL
+	ei.OutgoingSecret = opts.OutgoingSecret
+	ei.OutgoingToken = opts.OutgoingToken
+	token := auth.NewToken()
+	ei.AccessKey = token.AccessKey
+	ei.Salt = utils.NewSecret(utils.DefaultSecretSize)
+	hashedSecret, err := auth.HashedSecret(token, ei.Salt)
+	require.NoError(t, err)
+	ei.HashedSecret = hashedSecret
+	err = orm.CreateExternalInitiator(ctx, &ei)
+	require.NoError(t, err)
+	return ei
 }
