@@ -1,19 +1,51 @@
-// Code generated by mockery v1.0.0. DO NOT EDIT.
+// Code generated by mockery v2.42.2. DO NOT EDIT.
 
 package mocks
 
 import (
 	big "math/big"
 
-	models "github.com/smartcontractkit/chainlink/core/store/models"
+	audit "github.com/smartcontractkit/chainlink/v2/core/logger/audit"
+
+	bridges "github.com/smartcontractkit/chainlink/v2/core/bridges"
+
+	chainlink "github.com/smartcontractkit/chainlink/v2/core/services/chainlink"
+
+	context "context"
+
+	feeds "github.com/smartcontractkit/chainlink/v2/core/services/feeds"
+
+	job "github.com/smartcontractkit/chainlink/v2/core/services/job"
+
+	jsonserializable "github.com/smartcontractkit/chainlink-common/pkg/utils/jsonserializable"
+
+	keystore "github.com/smartcontractkit/chainlink/v2/core/services/keystore"
+
+	logger "github.com/smartcontractkit/chainlink/v2/core/logger"
+
+	logpoller "github.com/smartcontractkit/chainlink/v2/core/chains/evm/logpoller"
 
 	mock "github.com/stretchr/testify/mock"
 
-	packr "github.com/gobuffalo/packr"
+	pipeline "github.com/smartcontractkit/chainlink/v2/core/services/pipeline"
 
-	store "github.com/smartcontractkit/chainlink/core/store"
+	plugins "github.com/smartcontractkit/chainlink/v2/plugins"
 
-	synchronization "github.com/smartcontractkit/chainlink/core/services/synchronization"
+	services "github.com/smartcontractkit/chainlink/v2/core/services"
+
+	sessions "github.com/smartcontractkit/chainlink/v2/core/sessions"
+
+	sqlutil "github.com/smartcontractkit/chainlink-common/pkg/sqlutil"
+
+	txmgr "github.com/smartcontractkit/chainlink/v2/core/chains/evm/txmgr"
+
+	types "github.com/smartcontractkit/chainlink/v2/core/chains/evm/types"
+
+	uuid "github.com/google/uuid"
+
+	webhook "github.com/smartcontractkit/chainlink/v2/core/services/webhook"
+
+	zapcore "go.uber.org/zap/zapcore"
 )
 
 // Application is an autogenerated mock type for the Application type
@@ -21,13 +53,17 @@ type Application struct {
 	mock.Mock
 }
 
-// AddJob provides a mock function with given fields: job
-func (_m *Application) AddJob(job models.JobSpec) error {
-	ret := _m.Called(job)
+// AddJobV2 provides a mock function with given fields: ctx, _a1
+func (_m *Application) AddJobV2(ctx context.Context, _a1 *job.Job) error {
+	ret := _m.Called(ctx, _a1)
+
+	if len(ret) == 0 {
+		panic("no return value specified for AddJobV2")
+	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(models.JobSpec) error); ok {
-		r0 = rf(job)
+	if rf, ok := ret.Get(0).(func(context.Context, *job.Job) error); ok {
+		r0 = rf(ctx, _a1)
 	} else {
 		r0 = ret.Error(0)
 	}
@@ -35,50 +71,145 @@ func (_m *Application) AddJob(job models.JobSpec) error {
 	return r0
 }
 
-// AddServiceAgreement provides a mock function with given fields: _a0
-func (_m *Application) AddServiceAgreement(_a0 *models.ServiceAgreement) error {
-	ret := _m.Called(_a0)
+// AuthenticationProvider provides a mock function with given fields:
+func (_m *Application) AuthenticationProvider() sessions.AuthenticationProvider {
+	ret := _m.Called()
 
-	var r0 error
-	if rf, ok := ret.Get(0).(func(*models.ServiceAgreement) error); ok {
-		r0 = rf(_a0)
-	} else {
-		r0 = ret.Error(0)
+	if len(ret) == 0 {
+		panic("no return value specified for AuthenticationProvider")
 	}
 
-	return r0
-}
-
-// ArchiveJob provides a mock function with given fields: _a0
-func (_m *Application) ArchiveJob(_a0 *models.ID) error {
-	ret := _m.Called(_a0)
-
-	var r0 error
-	if rf, ok := ret.Get(0).(func(*models.ID) error); ok {
-		r0 = rf(_a0)
-	} else {
-		r0 = ret.Error(0)
-	}
-
-	return r0
-}
-
-// Cancel provides a mock function with given fields: runID
-func (_m *Application) Cancel(runID *models.ID) (*models.JobRun, error) {
-	ret := _m.Called(runID)
-
-	var r0 *models.JobRun
-	if rf, ok := ret.Get(0).(func(*models.ID) *models.JobRun); ok {
-		r0 = rf(runID)
+	var r0 sessions.AuthenticationProvider
+	if rf, ok := ret.Get(0).(func() sessions.AuthenticationProvider); ok {
+		r0 = rf()
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).(*models.JobRun)
+			r0 = ret.Get(0).(sessions.AuthenticationProvider)
 		}
 	}
 
+	return r0
+}
+
+// BasicAdminUsersORM provides a mock function with given fields:
+func (_m *Application) BasicAdminUsersORM() sessions.BasicAdminUsersORM {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for BasicAdminUsersORM")
+	}
+
+	var r0 sessions.BasicAdminUsersORM
+	if rf, ok := ret.Get(0).(func() sessions.BasicAdminUsersORM); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(sessions.BasicAdminUsersORM)
+		}
+	}
+
+	return r0
+}
+
+// BridgeORM provides a mock function with given fields:
+func (_m *Application) BridgeORM() bridges.ORM {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for BridgeORM")
+	}
+
+	var r0 bridges.ORM
+	if rf, ok := ret.Get(0).(func() bridges.ORM); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(bridges.ORM)
+		}
+	}
+
+	return r0
+}
+
+// DeleteJob provides a mock function with given fields: ctx, jobID
+func (_m *Application) DeleteJob(ctx context.Context, jobID int32) error {
+	ret := _m.Called(ctx, jobID)
+
+	if len(ret) == 0 {
+		panic("no return value specified for DeleteJob")
+	}
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(context.Context, int32) error); ok {
+		r0 = rf(ctx, jobID)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
+// DeleteLogPollerDataAfter provides a mock function with given fields: ctx, chainID, start
+func (_m *Application) DeleteLogPollerDataAfter(ctx context.Context, chainID *big.Int, start int64) error {
+	ret := _m.Called(ctx, chainID, start)
+
+	if len(ret) == 0 {
+		panic("no return value specified for DeleteLogPollerDataAfter")
+	}
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(context.Context, *big.Int, int64) error); ok {
+		r0 = rf(ctx, chainID, start)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
+// EVMORM provides a mock function with given fields:
+func (_m *Application) EVMORM() types.Configs {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for EVMORM")
+	}
+
+	var r0 types.Configs
+	if rf, ok := ret.Get(0).(func() types.Configs); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(types.Configs)
+		}
+	}
+
+	return r0
+}
+
+// FindLCA provides a mock function with given fields: ctx, chainID
+func (_m *Application) FindLCA(ctx context.Context, chainID *big.Int) (*logpoller.LogPollerBlock, error) {
+	ret := _m.Called(ctx, chainID)
+
+	if len(ret) == 0 {
+		panic("no return value specified for FindLCA")
+	}
+
+	var r0 *logpoller.LogPollerBlock
 	var r1 error
-	if rf, ok := ret.Get(1).(func(*models.ID) error); ok {
-		r1 = rf(runID)
+	if rf, ok := ret.Get(0).(func(context.Context, *big.Int) (*logpoller.LogPollerBlock, error)); ok {
+		return rf(ctx, chainID)
+	}
+	if rf, ok := ret.Get(0).(func(context.Context, *big.Int) *logpoller.LogPollerBlock); ok {
+		r0 = rf(ctx, chainID)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*logpoller.LogPollerBlock)
+		}
+	}
+
+	if rf, ok := ret.Get(1).(func(context.Context, *big.Int) error); ok {
+		r1 = rf(ctx, chainID)
 	} else {
 		r1 = ret.Error(1)
 	}
@@ -86,22 +217,381 @@ func (_m *Application) Cancel(runID *models.ID) (*models.JobRun, error) {
 	return r0, r1
 }
 
-// Create provides a mock function with given fields: jobSpecID, initiator, creationHeight, runRequest
-func (_m *Application) Create(jobSpecID *models.ID, initiator *models.Initiator, creationHeight *big.Int, runRequest *models.RunRequest) (*models.JobRun, error) {
-	ret := _m.Called(jobSpecID, initiator, creationHeight, runRequest)
+// GetAuditLogger provides a mock function with given fields:
+func (_m *Application) GetAuditLogger() audit.AuditLogger {
+	ret := _m.Called()
 
-	var r0 *models.JobRun
-	if rf, ok := ret.Get(0).(func(*models.ID, *models.Initiator, *big.Int, *models.RunRequest) *models.JobRun); ok {
-		r0 = rf(jobSpecID, initiator, creationHeight, runRequest)
+	if len(ret) == 0 {
+		panic("no return value specified for GetAuditLogger")
+	}
+
+	var r0 audit.AuditLogger
+	if rf, ok := ret.Get(0).(func() audit.AuditLogger); ok {
+		r0 = rf()
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).(*models.JobRun)
+			r0 = ret.Get(0).(audit.AuditLogger)
 		}
 	}
 
+	return r0
+}
+
+// GetConfig provides a mock function with given fields:
+func (_m *Application) GetConfig() chainlink.GeneralConfig {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetConfig")
+	}
+
+	var r0 chainlink.GeneralConfig
+	if rf, ok := ret.Get(0).(func() chainlink.GeneralConfig); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(chainlink.GeneralConfig)
+		}
+	}
+
+	return r0
+}
+
+// GetDB provides a mock function with given fields:
+func (_m *Application) GetDB() sqlutil.DataSource {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetDB")
+	}
+
+	var r0 sqlutil.DataSource
+	if rf, ok := ret.Get(0).(func() sqlutil.DataSource); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(sqlutil.DataSource)
+		}
+	}
+
+	return r0
+}
+
+// GetExternalInitiatorManager provides a mock function with given fields:
+func (_m *Application) GetExternalInitiatorManager() webhook.ExternalInitiatorManager {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetExternalInitiatorManager")
+	}
+
+	var r0 webhook.ExternalInitiatorManager
+	if rf, ok := ret.Get(0).(func() webhook.ExternalInitiatorManager); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(webhook.ExternalInitiatorManager)
+		}
+	}
+
+	return r0
+}
+
+// GetFeedsService provides a mock function with given fields:
+func (_m *Application) GetFeedsService() feeds.Service {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetFeedsService")
+	}
+
+	var r0 feeds.Service
+	if rf, ok := ret.Get(0).(func() feeds.Service); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(feeds.Service)
+		}
+	}
+
+	return r0
+}
+
+// GetHealthChecker provides a mock function with given fields:
+func (_m *Application) GetHealthChecker() services.Checker {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetHealthChecker")
+	}
+
+	var r0 services.Checker
+	if rf, ok := ret.Get(0).(func() services.Checker); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(services.Checker)
+		}
+	}
+
+	return r0
+}
+
+// GetKeyStore provides a mock function with given fields:
+func (_m *Application) GetKeyStore() keystore.Master {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetKeyStore")
+	}
+
+	var r0 keystore.Master
+	if rf, ok := ret.Get(0).(func() keystore.Master); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(keystore.Master)
+		}
+	}
+
+	return r0
+}
+
+// GetLogger provides a mock function with given fields:
+func (_m *Application) GetLogger() logger.SugaredLogger {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetLogger")
+	}
+
+	var r0 logger.SugaredLogger
+	if rf, ok := ret.Get(0).(func() logger.SugaredLogger); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(logger.SugaredLogger)
+		}
+	}
+
+	return r0
+}
+
+// GetLoopRegistrarConfig provides a mock function with given fields:
+func (_m *Application) GetLoopRegistrarConfig() plugins.RegistrarConfig {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetLoopRegistrarConfig")
+	}
+
+	var r0 plugins.RegistrarConfig
+	if rf, ok := ret.Get(0).(func() plugins.RegistrarConfig); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(plugins.RegistrarConfig)
+		}
+	}
+
+	return r0
+}
+
+// GetLoopRegistry provides a mock function with given fields:
+func (_m *Application) GetLoopRegistry() *plugins.LoopRegistry {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetLoopRegistry")
+	}
+
+	var r0 *plugins.LoopRegistry
+	if rf, ok := ret.Get(0).(func() *plugins.LoopRegistry); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*plugins.LoopRegistry)
+		}
+	}
+
+	return r0
+}
+
+// GetRelayers provides a mock function with given fields:
+func (_m *Application) GetRelayers() chainlink.RelayerChainInteroperators {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetRelayers")
+	}
+
+	var r0 chainlink.RelayerChainInteroperators
+	if rf, ok := ret.Get(0).(func() chainlink.RelayerChainInteroperators); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(chainlink.RelayerChainInteroperators)
+		}
+	}
+
+	return r0
+}
+
+// GetWebAuthnConfiguration provides a mock function with given fields:
+func (_m *Application) GetWebAuthnConfiguration() sessions.WebAuthnConfiguration {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetWebAuthnConfiguration")
+	}
+
+	var r0 sessions.WebAuthnConfiguration
+	if rf, ok := ret.Get(0).(func() sessions.WebAuthnConfiguration); ok {
+		r0 = rf()
+	} else {
+		r0 = ret.Get(0).(sessions.WebAuthnConfiguration)
+	}
+
+	return r0
+}
+
+// ID provides a mock function with given fields:
+func (_m *Application) ID() uuid.UUID {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for ID")
+	}
+
+	var r0 uuid.UUID
+	if rf, ok := ret.Get(0).(func() uuid.UUID); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(uuid.UUID)
+		}
+	}
+
+	return r0
+}
+
+// JobORM provides a mock function with given fields:
+func (_m *Application) JobORM() job.ORM {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for JobORM")
+	}
+
+	var r0 job.ORM
+	if rf, ok := ret.Get(0).(func() job.ORM); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(job.ORM)
+		}
+	}
+
+	return r0
+}
+
+// JobSpawner provides a mock function with given fields:
+func (_m *Application) JobSpawner() job.Spawner {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for JobSpawner")
+	}
+
+	var r0 job.Spawner
+	if rf, ok := ret.Get(0).(func() job.Spawner); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(job.Spawner)
+		}
+	}
+
+	return r0
+}
+
+// PipelineORM provides a mock function with given fields:
+func (_m *Application) PipelineORM() pipeline.ORM {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for PipelineORM")
+	}
+
+	var r0 pipeline.ORM
+	if rf, ok := ret.Get(0).(func() pipeline.ORM); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(pipeline.ORM)
+		}
+	}
+
+	return r0
+}
+
+// ReplayFromBlock provides a mock function with given fields: chainID, number, forceBroadcast
+func (_m *Application) ReplayFromBlock(chainID *big.Int, number uint64, forceBroadcast bool) error {
+	ret := _m.Called(chainID, number, forceBroadcast)
+
+	if len(ret) == 0 {
+		panic("no return value specified for ReplayFromBlock")
+	}
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(*big.Int, uint64, bool) error); ok {
+		r0 = rf(chainID, number, forceBroadcast)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
+// ResumeJobV2 provides a mock function with given fields: ctx, taskID, result
+func (_m *Application) ResumeJobV2(ctx context.Context, taskID uuid.UUID, result pipeline.Result) error {
+	ret := _m.Called(ctx, taskID, result)
+
+	if len(ret) == 0 {
+		panic("no return value specified for ResumeJobV2")
+	}
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(context.Context, uuid.UUID, pipeline.Result) error); ok {
+		r0 = rf(ctx, taskID, result)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
+// RunJobV2 provides a mock function with given fields: ctx, jobID, meta
+func (_m *Application) RunJobV2(ctx context.Context, jobID int32, meta map[string]interface{}) (int64, error) {
+	ret := _m.Called(ctx, jobID, meta)
+
+	if len(ret) == 0 {
+		panic("no return value specified for RunJobV2")
+	}
+
+	var r0 int64
 	var r1 error
-	if rf, ok := ret.Get(1).(func(*models.ID, *models.Initiator, *big.Int, *models.RunRequest) error); ok {
-		r1 = rf(jobSpecID, initiator, creationHeight, runRequest)
+	if rf, ok := ret.Get(0).(func(context.Context, int32, map[string]interface{}) (int64, error)); ok {
+		return rf(ctx, jobID, meta)
+	}
+	if rf, ok := ret.Get(0).(func(context.Context, int32, map[string]interface{}) int64); ok {
+		r0 = rf(ctx, jobID, meta)
+	} else {
+		r0 = ret.Get(0).(int64)
+	}
+
+	if rf, ok := ret.Get(1).(func(context.Context, int32, map[string]interface{}) error); ok {
+		r1 = rf(ctx, jobID, meta)
 	} else {
 		r1 = ret.Error(1)
 	}
@@ -109,22 +599,27 @@ func (_m *Application) Create(jobSpecID *models.ID, initiator *models.Initiator,
 	return r0, r1
 }
 
-// CreateErrored provides a mock function with given fields: jobSpecID, initiator, err
-func (_m *Application) CreateErrored(jobSpecID *models.ID, initiator models.Initiator, err error) (*models.JobRun, error) {
-	ret := _m.Called(jobSpecID, initiator, err)
+// RunWebhookJobV2 provides a mock function with given fields: ctx, jobUUID, requestBody, meta
+func (_m *Application) RunWebhookJobV2(ctx context.Context, jobUUID uuid.UUID, requestBody string, meta jsonserializable.JSONSerializable) (int64, error) {
+	ret := _m.Called(ctx, jobUUID, requestBody, meta)
 
-	var r0 *models.JobRun
-	if rf, ok := ret.Get(0).(func(*models.ID, models.Initiator, error) *models.JobRun); ok {
-		r0 = rf(jobSpecID, initiator, err)
-	} else {
-		if ret.Get(0) != nil {
-			r0 = ret.Get(0).(*models.JobRun)
-		}
+	if len(ret) == 0 {
+		panic("no return value specified for RunWebhookJobV2")
 	}
 
+	var r0 int64
 	var r1 error
-	if rf, ok := ret.Get(1).(func(*models.ID, models.Initiator, error) error); ok {
-		r1 = rf(jobSpecID, initiator, err)
+	if rf, ok := ret.Get(0).(func(context.Context, uuid.UUID, string, jsonserializable.JSONSerializable) (int64, error)); ok {
+		return rf(ctx, jobUUID, requestBody, meta)
+	}
+	if rf, ok := ret.Get(0).(func(context.Context, uuid.UUID, string, jsonserializable.JSONSerializable) int64); ok {
+		r0 = rf(ctx, jobUUID, requestBody, meta)
+	} else {
+		r0 = ret.Get(0).(int64)
+	}
+
+	if rf, ok := ret.Get(1).(func(context.Context, uuid.UUID, string, jsonserializable.JSONSerializable) error); ok {
+		r1 = rf(ctx, jobUUID, requestBody, meta)
 	} else {
 		r1 = ret.Error(1)
 	}
@@ -132,59 +627,37 @@ func (_m *Application) CreateErrored(jobSpecID *models.ID, initiator models.Init
 	return r0, r1
 }
 
-// GetStatsPusher provides a mock function with given fields:
-func (_m *Application) GetStatsPusher() synchronization.StatsPusher {
+// SecretGenerator provides a mock function with given fields:
+func (_m *Application) SecretGenerator() chainlink.SecretGenerator {
 	ret := _m.Called()
 
-	var r0 synchronization.StatsPusher
-	if rf, ok := ret.Get(0).(func() synchronization.StatsPusher); ok {
+	if len(ret) == 0 {
+		panic("no return value specified for SecretGenerator")
+	}
+
+	var r0 chainlink.SecretGenerator
+	if rf, ok := ret.Get(0).(func() chainlink.SecretGenerator); ok {
 		r0 = rf()
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).(synchronization.StatsPusher)
+			r0 = ret.Get(0).(chainlink.SecretGenerator)
 		}
 	}
 
 	return r0
 }
 
-// GetStore provides a mock function with given fields:
-func (_m *Application) GetStore() *store.Store {
-	ret := _m.Called()
+// SetLogLevel provides a mock function with given fields: lvl
+func (_m *Application) SetLogLevel(lvl zapcore.Level) error {
+	ret := _m.Called(lvl)
 
-	var r0 *store.Store
-	if rf, ok := ret.Get(0).(func() *store.Store); ok {
-		r0 = rf()
-	} else {
-		if ret.Get(0) != nil {
-			r0 = ret.Get(0).(*store.Store)
-		}
+	if len(ret) == 0 {
+		panic("no return value specified for SetLogLevel")
 	}
-
-	return r0
-}
-
-// NewBox provides a mock function with given fields:
-func (_m *Application) NewBox() packr.Box {
-	ret := _m.Called()
-
-	var r0 packr.Box
-	if rf, ok := ret.Get(0).(func() packr.Box); ok {
-		r0 = rf()
-	} else {
-		r0 = ret.Get(0).(packr.Box)
-	}
-
-	return r0
-}
-
-// ResumeAllPendingNextBlock provides a mock function with given fields: currentBlockHeight
-func (_m *Application) ResumeAllPendingNextBlock(currentBlockHeight *big.Int) error {
-	ret := _m.Called(currentBlockHeight)
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(*big.Int) error); ok {
-		r0 = rf(currentBlockHeight)
+	if rf, ok := ret.Get(0).(func(zapcore.Level) error); ok {
+		r0 = rf(lvl)
 	} else {
 		r0 = ret.Error(0)
 	}
@@ -192,55 +665,17 @@ func (_m *Application) ResumeAllPendingNextBlock(currentBlockHeight *big.Int) er
 	return r0
 }
 
-// ResumeAllPendingConnection provides a mock function with given fields:
-func (_m *Application) ResumeAllPendingConnection() error {
-	ret := _m.Called()
+// Start provides a mock function with given fields: ctx
+func (_m *Application) Start(ctx context.Context) error {
+	ret := _m.Called(ctx)
 
-	var r0 error
-	if rf, ok := ret.Get(0).(func() error); ok {
-		r0 = rf()
-	} else {
-		r0 = ret.Error(0)
+	if len(ret) == 0 {
+		panic("no return value specified for Start")
 	}
 
-	return r0
-}
-
-// ResumeAllInProgress provides a mock function with given fields:
-func (_m *Application) ResumeAllInProgress() error {
-	ret := _m.Called()
-
 	var r0 error
-	if rf, ok := ret.Get(0).(func() error); ok {
-		r0 = rf()
-	} else {
-		r0 = ret.Error(0)
-	}
-
-	return r0
-}
-
-// ResumePendingBridge provides a mock function with given fields: runID, input
-func (_m *Application) ResumePendingBridge(runID *models.ID, input models.BridgeRunResult) error {
-	ret := _m.Called(runID, input)
-
-	var r0 error
-	if rf, ok := ret.Get(0).(func(*models.ID, models.BridgeRunResult) error); ok {
-		r0 = rf(runID, input)
-	} else {
-		r0 = ret.Error(0)
-	}
-
-	return r0
-}
-
-// Start provides a mock function with given fields:
-func (_m *Application) Start() error {
-	ret := _m.Called()
-
-	var r0 error
-	if rf, ok := ret.Get(0).(func() error); ok {
-		r0 = rf()
+	if rf, ok := ret.Get(0).(func(context.Context) error); ok {
+		r0 = rf(ctx)
 	} else {
 		r0 = ret.Error(0)
 	}
@@ -252,6 +687,10 @@ func (_m *Application) Start() error {
 func (_m *Application) Stop() error {
 	ret := _m.Called()
 
+	if len(ret) == 0 {
+		panic("no return value specified for Stop")
+	}
+
 	var r0 error
 	if rf, ok := ret.Get(0).(func() error); ok {
 		r0 = rf()
@@ -262,7 +701,41 @@ func (_m *Application) Stop() error {
 	return r0
 }
 
+// TxmStorageService provides a mock function with given fields:
+func (_m *Application) TxmStorageService() txmgr.EvmTxStore {
+	ret := _m.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for TxmStorageService")
+	}
+
+	var r0 txmgr.EvmTxStore
+	if rf, ok := ret.Get(0).(func() txmgr.EvmTxStore); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(txmgr.EvmTxStore)
+		}
+	}
+
+	return r0
+}
+
 // WakeSessionReaper provides a mock function with given fields:
 func (_m *Application) WakeSessionReaper() {
 	_m.Called()
+}
+
+// NewApplication creates a new instance of Application. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewApplication(t interface {
+	mock.TestingT
+	Cleanup(func())
+}) *Application {
+	mock := &Application{}
+	mock.Mock.Test(t)
+
+	t.Cleanup(func() { mock.AssertExpectations(t) })
+
+	return mock
 }
