@@ -1,3 +1,4 @@
+//go:build !windows
 // +build !windows
 
 package logger
@@ -9,7 +10,7 @@ func registerOSSinks() error {
 }
 
 // logFileURI returns the full path to the file the
-// ProductionLogger logs to, and uses zap's built in default file sink.
-func logFileURI(configRootDir string) string {
-	return filepath.ToSlash(filepath.Join(configRootDir, "log.jsonl"))
+// NewLogger logs to, and uses zap's built in default file sink.
+func (c Config) logFileURI() string {
+	return filepath.ToSlash(c.LogsFile())
 }
