@@ -1,3 +1,4 @@
+//go:build windows
 // +build windows
 
 package logger
@@ -14,8 +15,8 @@ import (
 // directory, with a custom scheme winfile:/// specifically tailored for
 // Windows to get around their handling of the file:// schema in uber.org/zap.
 // https://github.com/uber-go/zap/issues/621
-func logFileURI(configRootDir string) string {
-	return "winfile:///" + filepath.ToSlash(filepath.Join(configRootDir, "log.jsonl"))
+func (c Config) logFileURI() string {
+	return "winfile:///" + filepath.ToSlash(c.LogsFile())
 }
 
 func registerOSSinks() error {
