@@ -12,7 +12,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/smartcontractkit/chainlink/core/services/signatures/cryptotest"
+	"github.com/smartcontractkit/chainlink/v2/core/services/signatures/cryptotest"
 )
 
 var numPointSamples = 10
@@ -35,7 +35,7 @@ func TestPoint_CloneAndEqual(t *testing.T) {
 		assert.Equal(t, f, g, "clones should be equal")
 		g.Add(g, f)
 		assert.Equal(t, h, f,
-			"modifying a clone shouldn't change originial")
+			"modifying a clone shouldn't change original")
 	}
 }
 
