@@ -1,33 +1,9 @@
 package models
 
 import (
-	"fmt"
+	"errors"
 	"strings"
 )
-
-// DatabaseAccessError is an error that occurs during database access.
-type DatabaseAccessError struct {
-	msg string
-}
-
-func (e *DatabaseAccessError) Error() string { return e.msg }
-
-// NewDatabaseAccessError returns a database access error.
-func NewDatabaseAccessError(msg string) error {
-	return &DatabaseAccessError{msg}
-}
-
-// ValidationError is an error that occurs during validation.
-type ValidationError struct {
-	msg string
-}
-
-func (e *ValidationError) Error() string { return e.msg }
-
-// NewValidationError returns a validation error.
-func NewValidationError(msg string, values ...interface{}) error {
-	return &ValidationError{msg: fmt.Sprintf(msg, values...)}
-}
 
 // JSONAPIErrors holds errors conforming to the JSONAPI spec.
 type JSONAPIErrors struct {
@@ -74,12 +50,12 @@ func (jae *JSONAPIErrors) Add(detail string) {
 // Merge combines the arrays of the passed error if it is of type JSONAPIErrors,
 // otherwise simply adds a single error with the error string as detail.
 func (jae *JSONAPIErrors) Merge(e error) {
-	switch typed := e.(type) {
-	case *JSONAPIErrors:
-		jae.Errors = append(jae.Errors, typed.Errors...)
-	default:
-		jae.Add(e.Error())
+	var jsonErr *JSONAPIErrors
+	if errors.As(e, &jsonErr) {
+		jae.Errors = append(jae.Errors, jsonErr.Errors...)
+		return
 	}
+	jae.Add(e.Error())
 }
 
 // CoerceEmptyToNil will return nil if JSONAPIErrors has no errors.
