@@ -8,8 +8,6 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/smartcontractkit/chainlink/core/logger"
-
 	"golang.org/x/text/unicode/norm"
 )
 
@@ -20,7 +18,8 @@ import (
 // final representation will be consistent with the string produced by this
 // function due to differences in JSON implementations and information loss.
 // e.g:
-// 	JSON does not have a requirement to respect object key ordering.
+//
+//	JSON does not have a requirement to respect object key ordering.
 func NormalizedJSON(val []byte) (string, error) {
 	// Unmarshal into a generic interface{}
 	var data interface{}
@@ -34,7 +33,6 @@ func NormalizedJSON(val []byte) (string, error) {
 
 	// Wrap the buffer in a normalization writer
 	wc := norm.NFC.Writer(writer)
-	defer logger.ErrorIfCalling(wc.Close)
 
 	// Now marshal the generic interface
 	if err = marshal(wc, data); err != nil {
