@@ -1,99 +1,54 @@
-// Package utils is used for the common functions for dealing with
-// conversion to and from hex, bytes, and strings, formatting time.
+// Package utils is used for common functions and tools used across the codebase.
 package utils
 
 import (
+	"bytes"
+	"context"
+	"crypto/ed25519"
 	"crypto/rand"
 	"encoding/base64"
 	"encoding/hex"
 	"encoding/json"
+	"errors"
 	"fmt"
-	"math/big"
+	"math"
+	mrand "math/rand"
 	"sort"
-	"strconv"
 	"strings"
 	"sync"
+	"sync/atomic"
 	"time"
 
-	"github.com/smartcontractkit/chainlink/core/logger"
-
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/rlp"
+	"github.com/google/uuid"
 	"github.com/jpillora/backoff"
-	"github.com/pkg/errors"
-	uuid "github.com/satori/go.uuid"
-	"github.com/shopspring/decimal"
-	"github.com/tevino/abool"
+	pkgerrors "github.com/pkg/errors"
+	"github.com/robfig/cron/v3"
 	"golang.org/x/crypto/bcrypt"
 	"golang.org/x/crypto/sha3"
-	null "gopkg.in/guregu/null.v3"
+
+	ragep2ptypes "github.com/smartcontractkit/libocr/ragep2p/types"
+
+	"github.com/smartcontractkit/chainlink-common/pkg/services"
 )
 
-const (
-	// DefaultSecretSize is the entroy in bytes to generate a base64 string of 64 characters.
-	DefaultSecretSize = 48
-	// EVMWordByteLen the length of an EVM Word Byte
-	EVMWordByteLen = 32
-	// EVMWordHexLen the length of an EVM Word Hex
-	EVMWordHexLen = EVMWordByteLen * 2
-)
+// DefaultSecretSize is the entropy in bytes to generate a base64 string of 64 characters.
+const DefaultSecretSize = 48
 
-// ZeroAddress is an address of all zeroes, otherwise in Ethereum as
-// 0x0000000000000000000000000000000000000000
-var ZeroAddress = common.Address{}
-
-// EmptyHash is a hash of all zeroes, otherwise in Ethereum as
-// 0x0000000000000000000000000000000000000000000000000000000000000000
-var EmptyHash = common.Hash{}
-
-// WithoutZeroAddresses returns a list of addresses excluding the zero address.
-func WithoutZeroAddresses(addresses []common.Address) []common.Address {
-	var withoutZeros []common.Address
-	for _, address := range addresses {
-		if address != ZeroAddress {
-			withoutZeros = append(withoutZeros, address)
-		}
+func MustNewPeerID() string {
+	pubKey, _, err := ed25519.GenerateKey(rand.Reader)
+	if err != nil {
+		panic(err)
 	}
-	return withoutZeros
-}
-
-// HexToUint64 converts a given hex string to 64-bit unsigned integer.
-func HexToUint64(hex string) (uint64, error) {
-	return strconv.ParseUint(RemoveHexPrefix(hex), 16, 64)
-}
-
-// Uint64ToHex converts the given uint64 value to a hex-value string.
-func Uint64ToHex(i uint64) string {
-	return fmt.Sprintf("0x%x", i)
-}
-
-var maxUint256 = common.HexToHash("0x" + strings.Repeat("f", 64)).Big()
-
-// Uint256ToBytes(x) is x represented as the bytes of a uint256
-func Uint256ToBytes(x *big.Int) (uint256 []byte, err error) {
-	if x.Cmp(maxUint256) > 0 {
-		return nil, fmt.Errorf("too large to convert to uint256")
+	peerID, err := ragep2ptypes.PeerIDFromPublicKey(pubKey)
+	if err != nil {
+		panic(err)
 	}
-	uint256 = common.LeftPadBytes(x.Bytes(), EVMWordByteLen)
-	if x.Cmp(big.NewInt(0).SetBytes(uint256)) != 0 {
-		panic("failed to round-trip uint256 back to source big.Int")
-	}
-	return uint256, err
+	return peerID.String()
 }
 
 // ISO8601UTC formats given time to ISO8601.
 func ISO8601UTC(t time.Time) string {
 	return t.UTC().Format(time.RFC3339)
-}
-
-// NullISO8601UTC returns formatted time if valid, empty string otherwise.
-func NullISO8601UTC(t null.Time) string {
-	if t.Valid {
-		return ISO8601UTC(t.Time)
-	}
-	return ""
 }
 
 // DurationFromNow returns the amount of time since the Time
@@ -110,7 +65,7 @@ func FormatJSON(v interface{}) ([]byte, error) {
 // NewBytes32ID returns a randomly generated UUID that conforms to
 // Ethereum bytes32.
 func NewBytes32ID() string {
-	return strings.Replace(uuid.NewV4().String(), "-", "", -1)
+	return strings.ReplaceAll(uuid.New().String(), "-", "")
 }
 
 // NewSecret returns a new securely random sequence of n bytes of entropy.  The
@@ -121,43 +76,9 @@ func NewSecret(n int) string {
 	b := make([]byte, n)
 	_, err := rand.Read(b)
 	if err != nil {
-		panic(errors.Wrap(err, "generating secret failed"))
+		panic(pkgerrors.Wrap(err, "generating secret failed"))
 	}
 	return base64.StdEncoding.EncodeToString(b)
-}
-
-// RemoveHexPrefix removes the prefix (0x) of a given hex string.
-func RemoveHexPrefix(str string) string {
-	if HasHexPrefix(str) {
-		return str[2:]
-	}
-	return str
-}
-
-// HasHexPrefix returns true if the string starts with 0x.
-func HasHexPrefix(str string) bool {
-	return len(str) >= 2 && str[0] == '0' && (str[1] == 'x' || str[1] == 'X')
-}
-
-// DecodeEthereumTx takes an RLP hex encoded Ethereum transaction and
-// returns a Transaction struct with all the fields accessible.
-func DecodeEthereumTx(hex string) (types.Transaction, error) {
-	var tx types.Transaction
-	b, err := hexutil.Decode(hex)
-	if err != nil {
-		return tx, err
-	}
-	return tx, rlp.DecodeBytes(b, &tx)
-}
-
-// IsEmptyAddress checks that the address is empty, synonymous with the zero
-// account/address. No logs can come from this address, as there is no contract
-// present there.
-//
-// See https://stackoverflow.com/questions/48219716/what-is-address0-in-solidity
-// for the more info on the zero address.
-func IsEmptyAddress(addr common.Address) bool {
-	return addr == ZeroAddress
 }
 
 // StringToHex converts a standard string to a hex encoded string.
@@ -165,7 +86,7 @@ func StringToHex(in string) string {
 	return AddHexPrefix(hex.EncodeToString([]byte(in)))
 }
 
-// AddHexPrefix adds the previx (0x) to a given hex string.
+// AddHexPrefix adds the prefix (0x) to a given hex string.
 func AddHexPrefix(str string) string {
 	if len(str) < 2 || len(str) > 1 && strings.ToLower(str[0:2]) != "0x" {
 		str = "0x" + str
@@ -173,179 +94,30 @@ func AddHexPrefix(str string) string {
 	return str
 }
 
-// Sleeper interface is used for tasks that need to be done on some
-// interval, excluding Cron, like reconnecting.
-type Sleeper interface {
-	Reset()
-	Sleep()
-	After() time.Duration
-	Duration() time.Duration
-}
-
-// BackoffSleeper is a sleeper that backs off on subsequent attempts.
-type BackoffSleeper struct {
-	backoff.Backoff
-	beenRun *abool.AtomicBool
-}
-
-// NewBackoffSleeper returns a BackoffSleeper that is configured to
-// sleep for 0 seconds initially, then backs off from 1 second minimum
-// to 10 seconds maximum.
-func NewBackoffSleeper() *BackoffSleeper {
-	return &BackoffSleeper{
-		Backoff: backoff.Backoff{
-			Min: 1 * time.Second,
-			Max: 10 * time.Second,
-		},
-		beenRun: abool.New(),
-	}
-}
-
-// Sleep waits for the given duration, incrementing the back off.
-func (bs *BackoffSleeper) Sleep() {
-	if bs.beenRun.SetToIf(false, true) {
-		return
-	}
-	time.Sleep(bs.Backoff.Duration())
-}
-
-// After returns the duration for the next stop, and increments the backoff.
-func (bs *BackoffSleeper) After() time.Duration {
-	if bs.beenRun.SetToIf(false, true) {
-		return 0
-	}
-	return bs.Backoff.Duration()
-}
-
-// Duration returns the current duration value.
-func (bs *BackoffSleeper) Duration() time.Duration {
-	if !bs.beenRun.IsSet() {
-		return 0
-	}
-	return bs.ForAttempt(bs.Attempt())
-}
-
-// Reset resets the backoff intervals.
-func (bs *BackoffSleeper) Reset() {
-	bs.beenRun.UnSet()
-	bs.Backoff.Reset()
-}
-
-func RetryWithBackoff(chCancel <-chan struct{}, errPrefix string, fn func() error) (aborted bool) {
-	sleeper := NewBackoffSleeper()
-	sleeper.Reset()
-	for {
-		err := fn()
-		if err == nil {
+// IsEmpty returns true if bytes contains only zero values, or has len 0.
+func IsEmpty(bytes []byte) bool {
+	for _, b := range bytes {
+		if b != 0 {
 			return false
 		}
-
-		logger.Errorf("%v: %v", errPrefix, err)
-		select {
-		case <-chCancel:
-			return true
-		case <-time.After(sleeper.After()):
-			continue
-		}
 	}
+	return true
 }
 
-// MinBigs finds the minimum value of a list of big.Ints.
-func MinBigs(first *big.Int, bigs ...*big.Int) *big.Int {
-	min := first
-	for _, n := range bigs {
-		if min.Cmp(n) > 0 {
-			min = n
-		}
-	}
-	return min
+// UnmarshalToMap takes an input json string and returns a map[string]interface i.e. a raw object
+func UnmarshalToMap(input string) (map[string]interface{}, error) {
+	var output map[string]interface{}
+	err := json.Unmarshal([]byte(input), &output)
+	return output, err
 }
 
-// MaxBigs finds the maximum value of a list of big.Ints.
-func MaxBigs(first *big.Int, bigs ...*big.Int) *big.Int {
-	max := first
-	for _, n := range bigs {
-		if max.Cmp(n) < 0 {
-			max = n
-		}
+// MustUnmarshalToMap performs UnmarshalToMap, panics upon failure
+func MustUnmarshalToMap(input string) map[string]interface{} {
+	output, err := UnmarshalToMap(input)
+	if err != nil {
+		panic(err)
 	}
-	return max
-}
-
-// MaxUint32 finds the maximum value of a list of uint32s.
-func MaxUint32(first uint32, uints ...uint32) uint32 {
-	max := first
-	for _, n := range uints {
-		if n > max {
-			max = n
-		}
-	}
-	return max
-}
-
-// MaxInt finds the maximum value of a list of ints.
-func MaxInt(first int, ints ...int) int {
-	max := first
-	for _, n := range ints {
-		if n > max {
-			max = n
-		}
-	}
-	return max
-}
-
-// MinUint finds the minimum value of a list of uints.
-func MinUint(first uint, vals ...uint) uint {
-	min := first
-	for _, n := range vals {
-		if n < min {
-			min = n
-		}
-	}
-	return min
-}
-
-// CoerceInterfaceMapToStringMap converts map[interface{}]interface{} (interface maps) to
-// map[string]interface{} (string maps) and []interface{} with interface maps to string maps.
-// Relevant when serializing between CBOR and JSON.
-func CoerceInterfaceMapToStringMap(in interface{}) (interface{}, error) {
-	switch typed := in.(type) {
-	case map[string]interface{}:
-		for k, v := range typed {
-			coerced, err := CoerceInterfaceMapToStringMap(v)
-			if err != nil {
-				return nil, err
-			}
-			typed[k] = coerced
-		}
-		return typed, nil
-	case map[interface{}]interface{}:
-		m := map[string]interface{}{}
-		for k, v := range typed {
-			coercedKey, ok := k.(string)
-			if !ok {
-				return nil, fmt.Errorf("unable to coerce key %T %v to a string", k, k)
-			}
-			coerced, err := CoerceInterfaceMapToStringMap(v)
-			if err != nil {
-				return nil, err
-			}
-			m[coercedKey] = coerced
-		}
-		return m, nil
-	case []interface{}:
-		r := make([]interface{}, len(typed))
-		for i, v := range typed {
-			coerced, err := CoerceInterfaceMapToStringMap(v)
-			if err != nil {
-				return nil, err
-			}
-			r[i] = coerced
-		}
-		return r, nil
-	default:
-		return in, nil
-	}
+	return output
 }
 
 // HashPassword wraps around bcrypt.GenerateFromPassword for a friendlier API.
@@ -360,243 +132,119 @@ func CheckPasswordHash(password, hash string) bool {
 	return err == nil
 }
 
-// Keccak256 is a simplified interface for the legacy SHA3 implementation that
-// Ethereum uses.
-func Keccak256(in []byte) ([]byte, error) {
-	hash := sha3.NewLegacyKeccak256()
-	_, err := hash.Write(in)
-	return hash.Sum(nil), err
-}
-
 // Sha256 returns a hexadecimal encoded string of a hashed input
 func Sha256(in string) (string, error) {
 	hasher := sha3.New256()
 	_, err := hasher.Write([]byte(in))
 	if err != nil {
-		return "", errors.Wrap(err, "sha256 write error")
+		return "", pkgerrors.Wrap(err, "sha256 write error")
 	}
 	return hex.EncodeToString(hasher.Sum(nil)), nil
 }
 
-// StripBearer removes the 'Bearer: ' prefix from the HTTP Authorization header.
-func StripBearer(authorizationStr string) string {
-	return strings.TrimPrefix(strings.TrimSpace(authorizationStr), "Bearer ")
+// WithCloseChan wraps a context so that it is canceled if the passed in channel is closed.
+// Deprecated: Call [services.StopChan.Ctx] directly
+func WithCloseChan(parentCtx context.Context, chStop chan struct{}) (context.Context, context.CancelFunc) {
+	return services.StopChan(chStop).Ctx(parentCtx)
 }
 
-// IsQuoted checks if the first and last characters are either " or '.
-func IsQuoted(input []byte) bool {
-	return len(input) >= 2 &&
-		((input[0] == '"' && input[len(input)-1] == '"') ||
-			(input[0] == '\'' && input[len(input)-1] == '\''))
+// ContextFromChan creates a context that finishes when the provided channel receives or is closed.
+// Deprecated: Call [services.StopChan.NewCtx] directly.
+func ContextFromChan(chStop chan struct{}) (context.Context, context.CancelFunc) {
+	return services.StopChan(chStop).NewCtx()
 }
 
-// RemoveQuotes removes the first and last character if they are both either
-// " or ', otherwise it is a noop.
-func RemoveQuotes(input []byte) []byte {
-	if IsQuoted(input) {
-		return input[1 : len(input)-1]
-	}
-	return input
+// ContextFromChanWithTimeout creates a context with a timeout that finishes when the provided channel receives or is closed.
+// Deprecated: Call [services.StopChan.CtxCancel] directly
+func ContextFromChanWithTimeout(chStop chan struct{}, timeout time.Duration) (context.Context, context.CancelFunc) {
+	return services.StopChan(chStop).CtxCancel(context.WithTimeout(context.Background(), timeout))
 }
 
-// EIP55CapitalizedAddress returns true iff possibleAddressString has the correct
-// capitalization for an Ethereum address, per EIP 55
-func EIP55CapitalizedAddress(possibleAddressString string) bool {
-	if !HasHexPrefix(possibleAddressString) {
-		possibleAddressString = "0x" + possibleAddressString
-	}
-	EIP55Capitalized := common.HexToAddress(possibleAddressString).Hex()
-	return possibleAddressString == EIP55Capitalized
+// Deprecated: use services.StopChan
+type StopChan = services.StopChan
+
+// Deprecated: use services.StopRChan
+type StopRChan = services.StopRChan
+
+// BoundedQueue is a FIFO queue that discards older items when it reaches its capacity.
+type BoundedQueue[T any] struct {
+	capacity int
+	items    []T
+	mu       sync.RWMutex
 }
 
-// ParseEthereumAddress returns addressString as a go-ethereum Address, or an
-// error if it's invalid, e.g. if EIP 55 capitalization check fails
-func ParseEthereumAddress(addressString string) (common.Address, error) {
-	if !common.IsHexAddress(addressString) {
-		return common.Address{}, fmt.Errorf(
-			"not a valid Ethereum address: %s", addressString)
-	}
-	address := common.HexToAddress(addressString)
-	if !EIP55CapitalizedAddress(addressString) {
-		return common.Address{}, fmt.Errorf(
-			"%s treated as Ethereum address, but it has an invalid capitalization! "+
-				"The correctly-capitalized address would be %s, but "+
-				"check carefully before copying and pasting! ",
-			addressString, address.Hex())
-	}
-	return address, nil
+// NewBoundedQueue creates a new BoundedQueue instance
+func NewBoundedQueue[T any](capacity int) *BoundedQueue[T] {
+	var bq BoundedQueue[T]
+	bq.capacity = capacity
+	return &bq
 }
 
-// MustHash returns the keccak256 hash, or panics on failure.
-func MustHash(in string) common.Hash {
-	out, err := Keccak256([]byte(in))
-	if err != nil {
-		panic(err)
-	}
-	return common.BytesToHash(out)
-}
-
-// LogListeningAddress returns the LogListeningAddress
-func LogListeningAddress(address common.Address) string {
-	if address == ZeroAddress {
-		return "[all]"
-	}
-	return address.String()
-}
-
-// JustError takes a tuple and returns the last entry, the error.
-func JustError(_ interface{}, err error) error {
-	return err
-}
-
-var zero = big.NewInt(0)
-
-// CheckUint256(n) returns an error if n is out of bounds for a uint256
-func CheckUint256(n *big.Int) error {
-	if n.Cmp(zero) < 0 || n.Cmp(maxUint256) >= 0 {
-		return fmt.Errorf("number out of range for uint256")
-	}
-	return nil
-}
-
-// HexToUint256 returns the uint256 represented by s, or an error if it doesn't
-// represent one.
-func HexToUint256(s string) (*big.Int, error) {
-	rawNum, err := hexutil.Decode(s)
-	if err != nil {
-		return nil, errors.Wrapf(err, "while parsing %s as hex: ", s)
-	}
-	rv := big.NewInt(0).SetBytes(rawNum) // can't be negative number
-	if err := CheckUint256(rv); err != nil {
-		return nil, err
-	}
-	return rv, nil
-}
-
-// Uint256ToHex returns the hex representation of n, or error if out of bounds
-func Uint256ToHex(n *big.Int) (string, error) {
-	if err := CheckUint256(n); err != nil {
-		return "", err
-	}
-	return common.BigToHash(n).Hex(), nil
-}
-
-func DecimalFromBigInt(i *big.Int, precision int32) decimal.Decimal {
-	return decimal.NewFromBigInt(i, -precision)
-}
-
-func WaitGroupChan(wg *sync.WaitGroup) <-chan struct{} {
-	chAwait := make(chan struct{})
-	go func() {
-		defer close(chAwait)
-		wg.Wait()
-	}()
-	return chAwait
-}
-
-type DependentAwaiter interface {
-	AwaitDependents() <-chan struct{}
-	AddDependents(n int)
-	DependentReady()
-}
-
-type dependentAwaiter struct {
-	wg *sync.WaitGroup
-	ch <-chan struct{}
-}
-
-func NewDependentAwaiter() DependentAwaiter {
-	return &dependentAwaiter{
-		wg: &sync.WaitGroup{},
-	}
-}
-
-func (da *dependentAwaiter) AwaitDependents() <-chan struct{} {
-	if da.ch == nil {
-		da.ch = WaitGroupChan(da.wg)
-	}
-	return da.ch
-}
-
-func (da *dependentAwaiter) AddDependents(n int) {
-	da.wg.Add(n)
-}
-
-func (da *dependentAwaiter) DependentReady() {
-	da.wg.Done()
-}
-
-// FIFO queue that discards older items when it reaches its capacity.
-type BoundedQueue struct {
-	capacity uint
-	items    []interface{}
-	mu       *sync.RWMutex
-}
-
-func NewBoundedQueue(capacity uint) *BoundedQueue {
-	return &BoundedQueue{
-		capacity: capacity,
-		mu:       &sync.RWMutex{},
-	}
-}
-
-func (q *BoundedQueue) Add(x interface{}) {
+// Add appends items to a BoundedQueue
+func (q *BoundedQueue[T]) Add(x T) {
 	q.mu.Lock()
 	defer q.mu.Unlock()
 	q.items = append(q.items, x)
-	if uint(len(q.items)) > q.capacity {
-		excess := uint(len(q.items)) - q.capacity
+	if len(q.items) > q.capacity {
+		excess := len(q.items) - q.capacity
 		q.items = q.items[excess:]
 	}
 }
 
-func (q *BoundedQueue) Take() interface{} {
+// Take pulls the first item from the array and removes it
+func (q *BoundedQueue[T]) Take() (t T) {
 	q.mu.Lock()
 	defer q.mu.Unlock()
 	if len(q.items) == 0 {
-		return nil
+		return
 	}
-	x := q.items[0]
+	t = q.items[0]
 	q.items = q.items[1:]
-	return x
+	return
 }
 
-func (q *BoundedQueue) Empty() bool {
+// Empty check is a BoundedQueue is empty
+func (q *BoundedQueue[T]) Empty() bool {
 	q.mu.RLock()
 	defer q.mu.RUnlock()
 	return len(q.items) == 0
 }
 
-func (q *BoundedQueue) Full() bool {
+// Full checks if a BoundedQueue is over capacity.
+func (q *BoundedQueue[T]) Full() bool {
 	q.mu.RLock()
 	defer q.mu.RUnlock()
-	return uint(len(q.items)) >= q.capacity
+	return len(q.items) >= q.capacity
 }
 
-type BoundedPriorityQueue struct {
-	queues     map[uint]*BoundedQueue
+// BoundedPriorityQueue stores a series of BoundedQueues
+// with associated priorities and capacities
+type BoundedPriorityQueue[T any] struct {
+	queues     map[uint]*BoundedQueue[T]
 	priorities []uint
-	capacities map[uint]uint
-	mu         *sync.RWMutex
+	capacities map[uint]int
+	mu         sync.RWMutex
 }
 
-func NewBoundedPriorityQueue(capacities map[uint]uint) *BoundedPriorityQueue {
-	queues := make(map[uint]*BoundedQueue)
+// NewBoundedPriorityQueue creates a new BoundedPriorityQueue
+func NewBoundedPriorityQueue[T any](capacities map[uint]int) *BoundedPriorityQueue[T] {
+	queues := make(map[uint]*BoundedQueue[T])
 	var priorities []uint
 	for priority, capacity := range capacities {
 		priorities = append(priorities, priority)
-		queues[priority] = NewBoundedQueue(capacity)
+		queues[priority] = NewBoundedQueue[T](capacity)
 	}
 	sort.Slice(priorities, func(i, j int) bool { return priorities[i] < priorities[j] })
-	return &BoundedPriorityQueue{
+	bpq := BoundedPriorityQueue[T]{
 		queues:     queues,
 		priorities: priorities,
 		capacities: capacities,
-		mu:         &sync.RWMutex{},
 	}
+	return &bpq
 }
 
-func (q *BoundedPriorityQueue) Add(priority uint, x interface{}) {
+// Add pushes an item into a subque within a BoundedPriorityQueue
+func (q *BoundedPriorityQueue[T]) Add(priority uint, x T) {
 	q.mu.Lock()
 	defer q.mu.Unlock()
 
@@ -608,7 +256,8 @@ func (q *BoundedPriorityQueue) Add(priority uint, x interface{}) {
 	subqueue.Add(x)
 }
 
-func (q *BoundedPriorityQueue) Take() interface{} {
+// Take takes from the BoundedPriorityQueue's subque
+func (q *BoundedPriorityQueue[T]) Take() (t T) {
 	q.mu.Lock()
 	defer q.mu.Unlock()
 
@@ -619,10 +268,12 @@ func (q *BoundedPriorityQueue) Take() interface{} {
 		}
 		return queue.Take()
 	}
-	return nil
+	return
 }
 
-func (q *BoundedPriorityQueue) Empty() bool {
+// Empty checks the BoundedPriorityQueue
+// if all subqueues are empty
+func (q *BoundedPriorityQueue[T]) Empty() bool {
 	q.mu.RLock()
 	defer q.mu.RUnlock()
 
@@ -635,16 +286,311 @@ func (q *BoundedPriorityQueue) Empty() bool {
 	return true
 }
 
-// WrapIfError decorates an error with the given message.  It is intended to
-// be used with `defer` statements, like so:
-//
-// func SomeFunction() (err error) {
-//     defer WrapIfError(&err, "error in SomeFunction:")
-//
-//     ...
-// }
-func WrapIfError(err *error, msg string) {
-	if *err != nil {
-		*err = errors.Wrap(*err, msg)
+// TickerBase is an interface for pausable tickers.
+type TickerBase interface {
+	Resume()
+	Pause()
+	Destroy()
+	Ticks() <-chan time.Time
+}
+
+// PausableTicker stores a ticker with a duration
+type PausableTicker struct {
+	ticker   *time.Ticker
+	duration time.Duration
+	mu       *sync.RWMutex
+}
+
+// NewPausableTicker creates a new PausableTicker
+func NewPausableTicker(duration time.Duration) PausableTicker {
+	return PausableTicker{
+		duration: duration,
+		mu:       &sync.RWMutex{},
 	}
+}
+
+// Ticks retrieves the ticks from a PausableTicker
+func (t *PausableTicker) Ticks() <-chan time.Time {
+	t.mu.RLock()
+	defer t.mu.RUnlock()
+	if t.ticker == nil {
+		return nil
+	}
+	return t.ticker.C
+}
+
+// Pause pauses a PausableTicker
+func (t *PausableTicker) Pause() {
+	t.mu.Lock()
+	defer t.mu.Unlock()
+	if t.ticker != nil {
+		t.ticker.Stop()
+		t.ticker = nil
+	}
+}
+
+// Resume resumes a Ticker
+// using a PausibleTicker's duration
+func (t *PausableTicker) Resume() {
+	t.mu.Lock()
+	defer t.mu.Unlock()
+	if t.ticker == nil {
+		t.ticker = time.NewTicker(t.duration)
+	}
+}
+
+// Destroy pauses the PausibleTicker
+func (t *PausableTicker) Destroy() {
+	t.Pause()
+}
+
+// CronTicker is like a time.Ticker but for a cron schedule.
+type CronTicker struct {
+	*cron.Cron
+	ch      chan time.Time
+	beenRun atomic.Bool
+}
+
+// NewCronTicker returns a new CrontTicker for the given schedule.
+func NewCronTicker(schedule string) (CronTicker, error) {
+	cron := cron.New(cron.WithSeconds())
+	ch := make(chan time.Time, 1)
+	_, err := cron.AddFunc(schedule, func() {
+		select {
+		case ch <- time.Now():
+		default:
+		}
+	})
+	if err != nil {
+		return CronTicker{}, err
+	}
+	return CronTicker{Cron: cron, ch: ch}, nil
+}
+
+// Start - returns true if the CronTicker was actually started, false otherwise
+func (t *CronTicker) Start() bool {
+	if t.Cron != nil {
+		if t.beenRun.CompareAndSwap(false, true) {
+			t.Cron.Start()
+			return true
+		}
+	}
+	return false
+}
+
+// Stop - returns true if the CronTicker was actually stopped, false otherwise
+func (t *CronTicker) Stop() bool {
+	if t.Cron != nil {
+		if t.beenRun.CompareAndSwap(true, false) {
+			t.Cron.Stop()
+			return true
+		}
+	}
+	return false
+}
+
+// Ticks returns the underlying chanel.
+func (t *CronTicker) Ticks() <-chan time.Time {
+	return t.ch
+}
+
+// ValidateCronSchedule returns an error if the given schedule is invalid.
+func ValidateCronSchedule(schedule string) error {
+	if !(strings.HasPrefix(schedule, "CRON_TZ=") || strings.HasPrefix(schedule, "@every ")) {
+		return errors.New("cron schedule must specify a time zone using CRON_TZ, e.g. 'CRON_TZ=UTC 5 * * * *', or use the @every syntax, e.g. '@every 1h30m'")
+	}
+	parser := cron.NewParser(cron.SecondOptional | cron.Minute | cron.Hour | cron.Dom | cron.Month | cron.Dow | cron.Descriptor)
+	_, err := parser.Parse(schedule)
+	return pkgerrors.Wrapf(err, "invalid cron schedule '%v'", schedule)
+}
+
+// ResettableTimer stores a timer
+type ResettableTimer struct {
+	timer *time.Timer
+	mu    *sync.RWMutex
+}
+
+// NewResettableTimer creates a new ResettableTimer
+func NewResettableTimer() ResettableTimer {
+	return ResettableTimer{
+		mu: &sync.RWMutex{},
+	}
+}
+
+// Ticks retrieves the ticks from a ResettableTimer
+func (t *ResettableTimer) Ticks() <-chan time.Time {
+	t.mu.RLock()
+	defer t.mu.RUnlock()
+	if t.timer == nil {
+		return nil
+	}
+	return t.timer.C
+}
+
+// Stop stops a ResettableTimer
+func (t *ResettableTimer) Stop() {
+	t.mu.Lock()
+	defer t.mu.Unlock()
+	if t.timer != nil {
+		t.timer.Stop()
+		t.timer = nil
+	}
+}
+
+// Reset stops a ResettableTimer
+// and resets it with a new duration
+func (t *ResettableTimer) Reset(duration time.Duration) {
+	t.mu.Lock()
+	defer t.mu.Unlock()
+	if t.timer != nil {
+		t.timer.Stop()
+	}
+	t.timer = time.NewTimer(duration)
+}
+
+var (
+	ErrAlreadyStopped      = errors.New("already stopped")
+	ErrCannotStopUnstarted = errors.New("cannot stop unstarted service")
+)
+
+// StartStopOnce contains a StartStopOnceState integer
+// Deprecated: use services.StateMachine
+type StartStopOnce = services.StateMachine
+
+// WithJitter adds +/- 10% to a duration
+func WithJitter(d time.Duration) time.Duration {
+	// #nosec
+	if d == 0 {
+		return 0
+	}
+	// ensure non-zero arg to Intn to avoid panic
+	max := math.Max(float64(d.Abs())/5.0, 1.)
+	// #nosec - non critical randomness
+	jitter := mrand.Intn(int(max))
+	jitter = jitter - (jitter / 2)
+	return time.Duration(int(d) + jitter)
+}
+
+// NewRedialBackoff is a standard backoff to use for redialling or reconnecting to
+// unreachable network endpoints
+func NewRedialBackoff() backoff.Backoff {
+	return backoff.Backoff{
+		Min:    1 * time.Second,
+		Max:    15 * time.Second,
+		Jitter: true,
+	}
+}
+
+// KeyedMutex allows to lock based on particular values
+type KeyedMutex struct {
+	mutexes sync.Map
+}
+
+// LockInt64 locks the value for read/write
+func (m *KeyedMutex) LockInt64(key int64) func() {
+	value, _ := m.mutexes.LoadOrStore(key, new(sync.Mutex))
+	mtx := value.(*sync.Mutex)
+	mtx.Lock()
+
+	return mtx.Unlock
+}
+
+// BoxOutput formats its arguments as fmt.Printf, and encloses them in a box of
+// arrows pointing at their content, in order to better highlight it. See
+// ExampleBoxOutput
+func BoxOutput(errorMsgTemplate string, errorMsgValues ...interface{}) string {
+	errorMsgTemplate = fmt.Sprintf(errorMsgTemplate, errorMsgValues...)
+	lines := strings.Split(errorMsgTemplate, "\n")
+	maxlen := 0
+	for _, line := range lines {
+		if len(line) > maxlen {
+			maxlen = len(line)
+		}
+	}
+	internalLength := maxlen + 4
+	output := "↘" + strings.Repeat("↓", internalLength) + "↙\n" // top line
+	output += "→  " + strings.Repeat(" ", maxlen) + "  ←\n"
+	readme := strings.Repeat("README ", maxlen/7)
+	output += "→  " + readme + strings.Repeat(" ", maxlen-len(readme)) + "  ←\n"
+	output += "→  " + strings.Repeat(" ", maxlen) + "  ←\n"
+	for _, line := range lines {
+		output += "→  " + line + strings.Repeat(" ", maxlen-len(line)) + "  ←\n"
+	}
+	output += "→  " + strings.Repeat(" ", maxlen) + "  ←\n"
+	output += "→  " + readme + strings.Repeat(" ", maxlen-len(readme)) + "  ←\n"
+	output += "→  " + strings.Repeat(" ", maxlen) + "  ←\n"
+	return "\n" + output + "↗" + strings.Repeat("↑", internalLength) + "↖" + // bottom line
+		"\n\n"
+}
+
+// ConcatBytes appends a bunch of byte arrays into a single byte array
+func ConcatBytes(bufs ...[]byte) []byte {
+	return bytes.Join(bufs, []byte{})
+}
+
+func LeftPadBitString(input string, length int) string {
+	if len(input) >= length {
+		return input
+	}
+	return strings.Repeat("0", length-len(input)) + input
+}
+
+// ErrorBuffer uses joinedErrors interface to join multiple errors into a single error.
+// This is useful to track the most recent N errors in a service and flush them as a single error.
+type ErrorBuffer struct {
+	// buffer is a slice of errors
+	buffer []error
+
+	// cap is the maximum number of errors that the buffer can hold.
+	// Exceeding the cap results in discarding the oldest error
+	cap int
+
+	mu sync.RWMutex
+}
+
+func (eb *ErrorBuffer) Flush() (err error) {
+	eb.mu.RLock()
+	defer eb.mu.RUnlock()
+	err = errors.Join(eb.buffer...)
+	eb.buffer = nil
+	return
+}
+
+func (eb *ErrorBuffer) Append(incoming error) {
+	eb.mu.Lock()
+	defer eb.mu.Unlock()
+
+	if len(eb.buffer) == eb.cap && eb.cap != 0 {
+		eb.buffer = append(eb.buffer[1:], incoming)
+		return
+	}
+	eb.buffer = append(eb.buffer, incoming)
+}
+
+func (eb *ErrorBuffer) SetCap(cap int) {
+	eb.mu.Lock()
+	defer eb.mu.Unlock()
+	if len(eb.buffer) > cap {
+		eb.buffer = eb.buffer[len(eb.buffer)-cap:]
+	}
+	eb.cap = cap
+}
+
+// UnwrapError returns a list of underlying errors if passed error implements joinedError or return the err in a single-element list otherwise.
+//
+//nolint:errorlint // error type checks will fail on wrapped errors. Disabled since we are not doing checks on error types.
+func UnwrapError(err error) []error {
+	joined, ok := err.(interface{ Unwrap() []error })
+	if !ok {
+		return []error{err}
+	}
+	return joined.Unwrap()
+}
+
+// DeleteUnstable destructively removes slice element at index i
+// It does no bounds checking and may re-order the slice
+func DeleteUnstable[T any](s []T, i int) []T {
+	s[i] = s[len(s)-1]
+	s = s[:len(s)-1]
+	return s
 }
