@@ -1,66 +1,54 @@
 package utils_test
 
 import (
+	"context"
+	"encoding/hex"
 	"fmt"
-	"math/big"
-	"reflect"
-	"strings"
-	"sync"
+	"sync/atomic"
 	"testing"
 	"time"
 
-	"github.com/smartcontractkit/chainlink/core/internal/cltest"
-	"github.com/smartcontractkit/chainlink/core/utils"
+	"github.com/pkg/errors"
 
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
+	"github.com/smartcontractkit/chainlink/v2/core/internal/cltest"
+	"github.com/smartcontractkit/chainlink/v2/core/internal/testutils"
+	"github.com/smartcontractkit/chainlink/v2/core/utils"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go.uber.org/multierr"
 )
 
 func TestUtils_NewBytes32ID(t *testing.T) {
 	t.Parallel()
+
 	id := utils.NewBytes32ID()
 	assert.NotContains(t, id, "-")
 }
 
 func TestUtils_NewSecret(t *testing.T) {
 	t.Parallel()
+
 	tests := []struct {
 		numOfBytes int
 		wantStrLen int
 	}{
 		{12, 16}, {24, 32}, {48, 64}, {96, 128},
 	}
-	for _, tt := range tests {
-		test := tt
-		secret := utils.NewSecret(test.numOfBytes)
-		assert.Equal(t, test.wantStrLen, len(secret))
-	}
-}
+	for _, test := range tests {
+		test := test
 
-func TestUtils_IsEmptyAddress(t *testing.T) {
-	tests := []struct {
-		name string
-		addr common.Address
-		want bool
-	}{
-		{"zero address", common.Address{}, true},
-		{"non-zero address", cltest.NewAddress(), false},
-	}
-
-	for _, tt := range tests {
-		test := tt
-		t.Run(test.name, func(t *testing.T) {
+		t.Run(fmt.Sprintf("%d_%d", test.numOfBytes, test.wantStrLen), func(t *testing.T) {
 			t.Parallel()
-			actual := utils.IsEmptyAddress(test.addr)
-			assert.Equal(t, test.want, actual)
+
+			secret := utils.NewSecret(test.numOfBytes)
+			assert.Equal(t, test.wantStrLen, len(secret))
 		})
 	}
 }
 
 func TestUtils_StringToHex(t *testing.T) {
+	t.Parallel()
+
 	tests := []struct {
 		utf8 string
 		hex  string
@@ -70,311 +58,29 @@ func TestUtils_StringToHex(t *testing.T) {
 		{"", "0x"},
 	}
 
-	for _, tt := range tests {
-		test := tt
+	for _, test := range tests {
+		test := test
+
 		t.Run(test.utf8, func(t *testing.T) {
 			t.Parallel()
+
 			assert.Equal(t, test.hex, utils.StringToHex(test.utf8))
 		})
 	}
 }
 
-func TestUtils_BackoffSleeper(t *testing.T) {
-	bs := utils.NewBackoffSleeper()
-	assert.Equal(t, time.Duration(0), bs.Duration(), "should initially return immediately")
-	bs.Sleep()
-
-	d := 1 * time.Nanosecond
-	bs.Min = d
-	bs.Factor = 2
-	assert.Equal(t, d, bs.Duration())
-	bs.Sleep()
-
-	d2 := 2 * time.Nanosecond
-	assert.Equal(t, d2, bs.Duration())
-
-	bs.Reset()
-	assert.Equal(t, time.Duration(0), bs.Duration(), "should initially return immediately")
-}
-
 func TestUtils_DurationFromNow(t *testing.T) {
 	t.Parallel()
+
 	future := time.Now().Add(time.Second)
 	duration := utils.DurationFromNow(future)
 	assert.True(t, 0 < duration)
 }
 
-func TestCoerceInterfaceMapToStringMap(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name      string
-		input     interface{}
-		want      interface{}
-		wantError bool
-	}{
-		{"empty map", map[interface{}]interface{}{}, map[string]interface{}{}, false},
-		{"simple map", map[interface{}]interface{}{"key": "value"}, map[string]interface{}{"key": "value"}, false},
-		{"int map", map[int]interface{}{1: "value"}, map[int]interface{}{1: "value"}, false},
-		{"error map", map[interface{}]interface{}{1: "value"}, map[int]interface{}{}, true},
-		{
-			"nested string map map",
-			map[string]interface{}{"key": map[interface{}]interface{}{"nk": "nv"}},
-			map[string]interface{}{"key": map[string]interface{}{"nk": "nv"}},
-			false,
-		},
-		{
-			"nested map map",
-			map[interface{}]interface{}{"key": map[interface{}]interface{}{"nk": "nv"}},
-			map[string]interface{}{"key": map[string]interface{}{"nk": "nv"}},
-			false,
-		},
-		{
-			"nested map array",
-			map[interface{}]interface{}{"key": []interface{}{1, "value"}},
-			map[string]interface{}{"key": []interface{}{1, "value"}},
-			false,
-		},
-		{"empty array", []interface{}{}, []interface{}{}, false},
-		{"simple array", []interface{}{1, "value"}, []interface{}{1, "value"}, false},
-		{
-			"error array",
-			[]interface{}{map[interface{}]interface{}{1: "value"}},
-			[]interface{}{},
-			true,
-		},
-		{
-			"nested array map",
-			[]interface{}{map[interface{}]interface{}{"key": map[interface{}]interface{}{"nk": "nv"}}},
-			[]interface{}{map[string]interface{}{"key": map[string]interface{}{"nk": "nv"}}},
-			false,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			decoded, err := utils.CoerceInterfaceMapToStringMap(test.input)
-			if test.wantError {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-				assert.True(t, reflect.DeepEqual(test.want, decoded))
-			}
-		})
-	}
-}
-
-func TestKeccak256(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name  string
-		input string
-		want  string
-	}{
-		{"basic", "0xf00b", "0x2433bb36d5f9b14e4fea87c2d32d79abfe34e56808b891e471f4400fca2a336c"},
-		{"long input", "0xf00b2433bb36d5f9b14e4fea87c2d32d79abfe34e56808b891e471f4400fca2a336c", "0x6b917c56ad7bea7d09132b9e1e29bb5d9aa7d32d067c638dfa886bbbf6874cdf"},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			input, err := hexutil.Decode(test.input)
-			assert.NoError(t, err)
-			result, err := utils.Keccak256(input)
-			assert.NoError(t, err)
-
-			assert.Equal(t, test.want, hexutil.Encode(result))
-		})
-	}
-}
-
-// From https://github.com/ethereum/EIPs/blob/master/EIPS/eip-55.md#test-cases
-var testAddresses = []string{
-	"0x52908400098527886E0F7030069857D2E4169EE7",
-	"0x8617E340B3D01FA5F11F306F4090FD50E238070D",
-	"0xde709f2102306220921060314715629080e2fb77",
-	"0x27b1fdb04752bbc536007a920d24acb045561c26",
-	"0x5aAeb6053F3E94C9b9A09f33669435E7Ef1BeAed",
-	"0xfB6916095ca1df60bB79Ce92cE3Ea74c37c5d359",
-	"0xdbF03B407c01E7cD3CBea99509d93f8DDDC8C6FB",
-	"0xD1220A0cf47c7B9Be7A2E6BA89F429762e7b9aDb",
-}
-
-func TestClient_EIP55CapitalizedAddress(t *testing.T) {
-	valid := utils.EIP55CapitalizedAddress
-	for _, address := range testAddresses {
-		assert.True(t, valid(address))
-		assert.False(t, valid(strings.ToLower(address)) &&
-			valid(strings.ToUpper(address)))
-	}
-}
-
-func TestClient_ParseEthereumAddress(t *testing.T) {
-	parse := utils.ParseEthereumAddress
-	for _, address := range testAddresses {
-		a1, err := parse(address)
-		assert.NoError(t, err)
-		no0xPrefix := address[2:]
-		a2, err := parse(no0xPrefix)
-		assert.NoError(t, err)
-		assert.True(t, a1 == a2)
-		_, lowerErr := parse(strings.ToLower(address))
-		_, upperErr := parse(strings.ToUpper(address))
-		shouldBeError := multierr.Combine(lowerErr, upperErr)
-		assert.Error(t, shouldBeError)
-		assert.True(t, strings.Contains(shouldBeError.Error(), no0xPrefix))
-	}
-	_, notHexErr := parse("0xCeci n'est pas une chaîne hexadécimale")
-	assert.Error(t, notHexErr)
-	_, tooLongErr := parse("0x0123456789abcdef0123456789abcdef0123456789abcdef")
-	assert.Error(t, tooLongErr)
-}
-
-func TestMinBigs(t *testing.T) {
-	tests := []struct {
-		min, max string
-	}{
-		{"0", "0"},
-		{"-1", "0"},
-		{"99", "100"},
-		{"0", "1"},
-		{"4294967295", "4294967296"},
-	}
-
-	for _, test := range tests {
-		t.Run(fmt.Sprintf("%s < %s", test.min, test.max), func(t *testing.T) {
-			left, ok := big.NewInt(0).SetString(test.min, 10)
-			require.True(t, ok)
-			right, ok := big.NewInt(0).SetString(test.max, 10)
-			require.True(t, ok)
-
-			min := utils.MinBigs(left, right)
-			assert.Equal(t, left, min)
-		})
-	}
-}
-
-func TestMaxUint32(t *testing.T) {
-	tests := []struct {
-		name        string
-		expectation uint32
-		vals        []uint32
-	}{
-		{"single", 9, []uint32{9}},
-		{"positives", 5, []uint32{3, 4, 5}},
-		{"equal", 3, []uint32{3, 3, 3}},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			actual := utils.MaxUint32(test.vals[0], test.vals[1:len(test.vals)]...)
-			assert.Equal(t, test.expectation, actual)
-		})
-	}
-}
-
-func TestMaxInt(t *testing.T) {
-	tests := []struct {
-		name        string
-		expectation int
-		vals        []int
-	}{
-		{"negatives", -1, []int{-1, -2, -9}},
-		{"positives", 5, []int{3, 4, 5}},
-		{"both", 5, []int{-1, -2, -9, 3, 4, 5}},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			actual := utils.MaxInt(test.vals[0], test.vals[1:len(test.vals)]...)
-			assert.Equal(t, test.expectation, actual)
-		})
-	}
-}
-
-func TestMinUint(t *testing.T) {
-	tests := []struct {
-		name        string
-		expectation uint
-		vals        []uint
-	}{
-		{"single", 9, []uint{9}},
-		{"positives", 3, []uint{3, 4, 5}},
-		{"equal", 3, []uint{3, 3, 3}},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			actual := utils.MinUint(test.vals[0], test.vals[1:len(test.vals)]...)
-			assert.Equal(t, test.expectation, actual)
-		})
-	}
-}
-
-func TestWaitGroupChan(t *testing.T) {
-	t.Parallel()
-
-	wg := &sync.WaitGroup{}
-	wg.Add(2)
-
-	ch := utils.WaitGroupChan(wg)
-
-	select {
-	case <-ch:
-		t.Fatal("should not fire immediately")
-	default:
-	}
-
-	wg.Done()
-
-	select {
-	case <-ch:
-		t.Fatal("should not fire until finished")
-	default:
-	}
-
-	go func() {
-		time.Sleep(2 * time.Second)
-		wg.Done()
-	}()
-
-	cltest.CallbackOrTimeout(t, "WaitGroupChan fires", func() {
-		<-ch
-	}, 5*time.Second)
-}
-
-func TestDependentAwaiter(t *testing.T) {
-	da := utils.NewDependentAwaiter()
-	da.AddDependents(2)
-
-	select {
-	case <-da.AwaitDependents():
-		t.Fatal("should not fire immediately")
-	default:
-	}
-
-	da.DependentReady()
-
-	select {
-	case <-da.AwaitDependents():
-		t.Fatal("should not fire until finished")
-	default:
-	}
-
-	go func() {
-		time.Sleep(2 * time.Second)
-		da.DependentReady()
-	}()
-
-	cltest.CallbackOrTimeout(t, "dependents are now ready", func() {
-		<-da.AwaitDependents()
-	}, 5*time.Second)
-}
-
 func TestBoundedQueue(t *testing.T) {
 	t.Parallel()
 
-	q := utils.NewBoundedQueue(3)
+	q := utils.NewBoundedQueue[int](3)
 	require.True(t, q.Empty())
 	require.False(t, q.Full())
 
@@ -382,11 +88,10 @@ func TestBoundedQueue(t *testing.T) {
 	require.False(t, q.Empty())
 	require.False(t, q.Full())
 
-	x := q.Take().(int)
+	x := q.Take()
 	require.Equal(t, 1, x)
 
-	iface := q.Take()
-	require.Nil(t, iface)
+	require.Zero(t, q.Take())
 	require.True(t, q.Empty())
 	require.False(t, q.Full())
 
@@ -396,17 +101,17 @@ func TestBoundedQueue(t *testing.T) {
 	q.Add(4)
 	require.True(t, q.Full())
 
-	x = q.Take().(int)
+	x = q.Take()
 	require.Equal(t, 2, x)
 	require.False(t, q.Empty())
 	require.False(t, q.Full())
 
-	x = q.Take().(int)
+	x = q.Take()
 	require.Equal(t, 3, x)
 	require.False(t, q.Empty())
 	require.False(t, q.Full())
 
-	x = q.Take().(int)
+	x = q.Take()
 	require.Equal(t, 4, x)
 	require.True(t, q.Empty())
 	require.False(t, q.Full())
@@ -415,7 +120,7 @@ func TestBoundedQueue(t *testing.T) {
 func TestBoundedPriorityQueue(t *testing.T) {
 	t.Parallel()
 
-	q := utils.NewBoundedPriorityQueue(map[uint]uint{
+	q := utils.NewBoundedPriorityQueue[int](map[uint]int{
 		1: 3,
 		2: 1,
 	})
@@ -424,12 +129,11 @@ func TestBoundedPriorityQueue(t *testing.T) {
 	q.Add(1, 1)
 	require.False(t, q.Empty())
 
-	x := q.Take().(int)
+	x := q.Take()
 	require.Equal(t, 1, x)
 	require.True(t, q.Empty())
 
-	iface := q.Take()
-	require.Nil(t, iface)
+	require.Zero(t, q.Take())
 	require.True(t, q.Empty())
 
 	q.Add(2, 1)
@@ -437,32 +141,412 @@ func TestBoundedPriorityQueue(t *testing.T) {
 	q.Add(1, 3)
 	q.Add(1, 4)
 
-	x = q.Take().(int)
+	x = q.Take()
 	require.Equal(t, 2, x)
 	require.False(t, q.Empty())
 
-	x = q.Take().(int)
+	x = q.Take()
 	require.Equal(t, 3, x)
 	require.False(t, q.Empty())
 
-	x = q.Take().(int)
+	x = q.Take()
 	require.Equal(t, 4, x)
 	require.False(t, q.Empty())
 
-	x = q.Take().(int)
+	x = q.Take()
 	require.Equal(t, 1, x)
 	require.True(t, q.Empty())
 
-	iface = q.Take()
-	require.Nil(t, iface)
+	require.Zero(t, q.Take())
 
 	q.Add(2, 1)
 	q.Add(2, 2)
 
-	x = q.Take().(int)
+	x = q.Take()
 	require.Equal(t, 2, x)
 	require.True(t, q.Empty())
 
-	iface = q.Take()
-	require.Nil(t, iface)
+	require.Zero(t, q.Take())
+}
+
+func Test_WithJitter(t *testing.T) {
+	t.Parallel()
+
+	d := 10 * time.Second
+
+	for i := 0; i < 32; i++ {
+		r := utils.WithJitter(d)
+		require.GreaterOrEqual(t, int(r), int(9*time.Second))
+		require.LessOrEqual(t, int(r), int(11*time.Second))
+	}
+}
+
+func TestIsEmpty(t *testing.T) {
+	t.Parallel()
+
+	b := make([]byte, 32)
+	assert.True(t, utils.IsEmpty(b))
+
+	b[10] = 1
+	assert.False(t, utils.IsEmpty(b))
+}
+
+func TestHashPassword(t *testing.T) {
+	t.Parallel()
+
+	h, err := utils.HashPassword("Qwerty123!")
+	assert.NoError(t, err)
+	assert.NotEmpty(t, h)
+
+	ok := utils.CheckPasswordHash("Qwerty123!", h)
+	assert.True(t, ok)
+
+	ok = utils.CheckPasswordHash("God", h)
+	assert.False(t, ok)
+}
+
+func TestBoxOutput(t *testing.T) {
+	t.Parallel()
+
+	output := utils.BoxOutput("some error %d %s", 123, "foo")
+	const expected = "\n" +
+		"↘↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↙\n" +
+		"→                      ←\n" +
+		"→  README README       ←\n" +
+		"→                      ←\n" +
+		"→  some error 123 foo  ←\n" +
+		"→                      ←\n" +
+		"→  README README       ←\n" +
+		"→                      ←\n" +
+		"↗↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↖\n" +
+		"\n"
+	assert.Equal(t, expected, output)
+}
+
+func TestISO8601UTC(t *testing.T) {
+	t.Parallel()
+
+	ts := time.Unix(1651818206, 0)
+	str := utils.ISO8601UTC(ts)
+	assert.Equal(t, "2022-05-06T06:23:26Z", str)
+}
+
+func TestFormatJSON(t *testing.T) {
+	t.Parallel()
+
+	json := `{"foo":123}`
+	formatted, err := utils.FormatJSON(json)
+	assert.NoError(t, err)
+	assert.Equal(t, "\"{\\\"foo\\\":123}\"", string(formatted))
+}
+
+func TestMustUnmarshalToMap(t *testing.T) {
+	t.Parallel()
+
+	json := `{"foo":123.45}`
+	expected := make(map[string]interface{})
+	expected["foo"] = 123.45
+	m := utils.MustUnmarshalToMap(json)
+	assert.Equal(t, expected, m)
+
+	assert.Panics(t, func() {
+		utils.MustUnmarshalToMap("")
+	})
+
+	assert.Panics(t, func() {
+		utils.MustUnmarshalToMap("123")
+	})
+}
+
+func TestSha256(t *testing.T) {
+	t.Parallel()
+
+	hexHash, err := utils.Sha256("test")
+	assert.NoError(t, err)
+
+	hash, err := hex.DecodeString(hexHash)
+	assert.NoError(t, err)
+	assert.Len(t, hash, 32)
+}
+
+func TestWithCloseChan(t *testing.T) {
+	t.Parallel()
+
+	assertCtxCancelled := func(ctx context.Context, t *testing.T) {
+		select {
+		case <-ctx.Done():
+		case <-time.After(testutils.WaitTimeout(t)):
+			assert.FailNow(t, "context was not cancelled")
+		}
+	}
+
+	t.Run("closing channel", func(t *testing.T) {
+		t.Parallel()
+
+		ch := make(chan struct{})
+		ctx, cancel := utils.WithCloseChan(testutils.Context(t), ch)
+		defer cancel()
+
+		close(ch)
+
+		assertCtxCancelled(ctx, t)
+	})
+
+	t.Run("cancelling ctx", func(t *testing.T) {
+		t.Parallel()
+
+		ch := make(chan struct{})
+		defer close(ch)
+		ctx, cancel := utils.WithCloseChan(testutils.Context(t), ch)
+		cancel()
+
+		assertCtxCancelled(ctx, t)
+	})
+
+	t.Run("cancelling parent ctx", func(t *testing.T) {
+		t.Parallel()
+
+		ch := make(chan struct{})
+		defer close(ch)
+		pctx, pcancel := context.WithCancel(testutils.Context(t))
+		ctx, cancel := utils.WithCloseChan(pctx, ch)
+		defer cancel()
+
+		pcancel()
+
+		assertCtxCancelled(ctx, t)
+	})
+}
+
+func TestContextFromChan(t *testing.T) {
+	t.Parallel()
+
+	ch := make(chan struct{})
+	ctx, cancel := utils.ContextFromChan(ch)
+	defer cancel()
+
+	close(ch)
+
+	select {
+	case <-ctx.Done():
+	case <-time.After(testutils.WaitTimeout(t)):
+		assert.FailNow(t, "context was not cancelled")
+	}
+}
+
+func TestContextFromChanWithTimeout(t *testing.T) {
+	t.Parallel()
+
+	assertCtxCancelled := func(ctx context.Context, t *testing.T) {
+		select {
+		case <-ctx.Done():
+		case <-time.After(testutils.WaitTimeout(t)):
+			assert.FailNow(t, "context was not cancelled")
+		}
+	}
+
+	t.Run("small deadline", func(t *testing.T) {
+		t.Parallel()
+
+		ch := make(chan struct{})
+		ctx, cancel := utils.ContextFromChanWithTimeout(ch, testutils.TestInterval)
+		defer cancel()
+
+		assertCtxCancelled(ctx, t)
+	})
+
+	t.Run("stopped", func(t *testing.T) {
+		t.Parallel()
+
+		ch := make(chan struct{})
+		ctx, cancel := utils.ContextFromChanWithTimeout(ch, testutils.WaitTimeout(t))
+		defer cancel()
+
+		ch <- struct{}{}
+
+		assertCtxCancelled(ctx, t)
+	})
+}
+
+func TestLeftPadBitString(t *testing.T) {
+	t.Parallel()
+
+	for _, test := range []struct {
+		str      string
+		len      int
+		expected string
+	}{
+		{"abc", 10, "0000000abc"},
+		{"abc", 0, "abc"},
+		{"abc", 2, "abc"},
+		{"abc", 3, "abc"},
+		{"abc", -10, "abc"},
+	} {
+		s := utils.LeftPadBitString(test.str, test.len)
+		assert.Equal(t, test.expected, s)
+	}
+}
+
+func TestKeyedMutex(t *testing.T) {
+	t.Parallel()
+
+	var km utils.KeyedMutex
+	unlock1 := km.LockInt64(1)
+	unlock2 := km.LockInt64(2)
+
+	awaiter := cltest.NewAwaiter()
+	go func() {
+		km.LockInt64(1)()
+		km.LockInt64(2)()
+		awaiter.ItHappened()
+	}()
+
+	unlock2()
+	unlock1()
+	awaiter.AwaitOrFail(t)
+}
+
+func TestValidateCronSchedule(t *testing.T) {
+	t.Parallel()
+
+	err := utils.ValidateCronSchedule("")
+	assert.Error(t, err)
+
+	err = utils.ValidateCronSchedule("CRON_TZ=UTC 5 * * * *")
+	assert.NoError(t, err)
+
+	err = utils.ValidateCronSchedule("@every 1h30m")
+	assert.NoError(t, err)
+
+	err = utils.ValidateCronSchedule("@every xyz")
+	assert.Error(t, err)
+}
+
+func TestPausableTicker(t *testing.T) {
+	t.Parallel()
+
+	var counter atomic.Int32
+
+	pt := utils.NewPausableTicker(testutils.TestInterval)
+	assert.Nil(t, pt.Ticks())
+	defer pt.Destroy()
+
+	followNTicks := func(n int32, awaiter cltest.Awaiter) {
+		for range pt.Ticks() {
+			if counter.Add(1) == n {
+				awaiter.ItHappened()
+			}
+		}
+	}
+
+	pt.Resume()
+
+	wait10 := cltest.NewAwaiter()
+	go followNTicks(10, wait10)
+
+	wait10.AwaitOrFail(t)
+
+	pt.Pause()
+	time.Sleep(10 * testutils.TestInterval)
+	assert.Less(t, counter.Load(), int32(20))
+	pt.Resume()
+
+	wait20 := cltest.NewAwaiter()
+	go followNTicks(20, wait20)
+
+	wait20.AwaitOrFail(t)
+}
+
+func TestCronTicker(t *testing.T) {
+	t.Parallel()
+
+	var counter atomic.Int32
+
+	ct, err := utils.NewCronTicker("@every 100ms")
+	assert.NoError(t, err)
+
+	awaiter := cltest.NewAwaiter()
+
+	go func() {
+		for range ct.Ticks() {
+			if counter.Add(1) == 2 {
+				awaiter.ItHappened()
+			}
+		}
+	}()
+
+	assert.True(t, ct.Start())
+	assert.True(t, ct.Stop())
+	assert.Zero(t, counter.Load())
+
+	assert.True(t, ct.Start())
+
+	awaiter.AwaitOrFail(t)
+
+	assert.True(t, ct.Stop())
+	c := counter.Load()
+	time.Sleep(1 * time.Second)
+	assert.Equal(t, c, counter.Load())
+}
+
+func TestErrorBuffer(t *testing.T) {
+	t.Parallel()
+
+	err1 := errors.New("err1")
+	err2 := errors.New("err2")
+	err3 := errors.New("err3")
+
+	t.Run("happy path", func(t *testing.T) {
+		t.Parallel()
+		buff := utils.ErrorBuffer{}
+		buff.Append(err1)
+		buff.Append(err2)
+		combined := buff.Flush()
+		errs := utils.UnwrapError(combined)
+		assert.Equal(t, 2, len(errs))
+		assert.Equal(t, err1.Error(), errs[0].Error())
+		assert.Equal(t, err2.Error(), errs[1].Error())
+	})
+
+	t.Run("ovewrite oldest error when cap exceeded", func(t *testing.T) {
+		t.Parallel()
+		buff := utils.ErrorBuffer{}
+		buff.SetCap(2)
+		buff.Append(err1)
+		buff.Append(err2)
+		buff.Append(err3)
+		combined := buff.Flush()
+		errs := utils.UnwrapError(combined)
+		assert.Equal(t, 2, len(errs))
+		assert.Equal(t, err2.Error(), errs[0].Error())
+		assert.Equal(t, err3.Error(), errs[1].Error())
+	})
+
+	t.Run("does not overwrite the buffer if cap == 0", func(t *testing.T) {
+		t.Parallel()
+		buff := utils.ErrorBuffer{}
+		for i := 1; i <= 20; i++ {
+			buff.Append(errors.Errorf("err#%d", i))
+		}
+
+		combined := buff.Flush()
+		errs := utils.UnwrapError(combined)
+		assert.Equal(t, 20, len(errs))
+		assert.Equal(t, "err#20", errs[19].Error())
+	})
+
+	t.Run("UnwrapError returns the a single element err array if passed err is not a joinedError", func(t *testing.T) {
+		t.Parallel()
+		errs := utils.UnwrapError(err1)
+		assert.Equal(t, 1, len(errs))
+		assert.Equal(t, err1.Error(), errs[0].Error())
+	})
+
+	t.Run("flushing an empty err buffer is a nil error", func(t *testing.T) {
+		t.Parallel()
+		buff := utils.ErrorBuffer{}
+
+		combined := buff.Flush()
+		require.Nil(t, combined)
+	})
 }
