@@ -6,73 +6,79 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/smartcontractkit/chainlink/core/auth"
-	"github.com/smartcontractkit/chainlink/core/internal/cltest"
-	"github.com/smartcontractkit/chainlink/core/store/models"
-	"github.com/smartcontractkit/chainlink/core/web"
+	"github.com/google/uuid"
+
+	"github.com/smartcontractkit/chainlink/v2/core/auth"
+	"github.com/smartcontractkit/chainlink/v2/core/bridges"
+	"github.com/smartcontractkit/chainlink/v2/core/internal/cltest"
+	"github.com/smartcontractkit/chainlink/v2/core/internal/testutils"
+	clhttptest "github.com/smartcontractkit/chainlink/v2/core/internal/testutils/httptest"
+	"github.com/smartcontractkit/chainlink/v2/core/web"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestTokenAuthRequired_NoCredentials(t *testing.T) {
-	app, cleanup := cltest.NewApplicationWithKey(t, cltest.LenientEthMock)
-	defer cleanup()
-	require.NoError(t, app.Start())
+	ctx := testutils.Context(t)
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(ctx))
 
-	router := web.Router(app)
+	router := web.Router(t, app, nil)
 	ts := httptest.NewServer(router)
 	defer ts.Close()
 
-	resp, err := http.Post(ts.URL+"/v2/specs/", web.MediaType, bytes.NewBufferString("{}"))
+	req, err := http.NewRequestWithContext(ctx, "POST", ts.URL+"/v2/jobs/", bytes.NewBufferString("{}"))
+	require.NoError(t, err)
+	req.Header.Set("Content-Type", web.MediaType)
+	resp, err := http.DefaultClient.Do(req)
 	require.NoError(t, err)
 
 	assert.Equal(t, http.StatusUnauthorized, resp.StatusCode)
 }
 
 func TestTokenAuthRequired_SessionCredentials(t *testing.T) {
-	app, cleanup := cltest.NewApplicationWithKey(t, cltest.LenientEthMock)
-	defer cleanup()
-	require.NoError(t, app.Start())
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(testutils.Context(t)))
 
-	router := web.Router(app)
+	router := web.Router(t, app, nil)
 	ts := httptest.NewServer(router)
 	defer ts.Close()
 
-	client := app.NewHTTPClient()
-	resp, cleanup := client.Post("/v2/specs/", nil)
+	client := app.NewHTTPClient(nil)
+	resp, cleanup := client.Post("/v2/bridge_types/", nil)
 	defer cleanup()
 
-	assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
+	assert.Equal(t, http.StatusUnprocessableEntity, resp.StatusCode)
 }
 
 func TestTokenAuthRequired_TokenCredentials(t *testing.T) {
-	app, cleanup := cltest.NewApplicationWithKey(t, cltest.LenientEthMock)
-	defer cleanup()
-	require.NoError(t, app.Start())
+	ctx := testutils.Context(t)
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(ctx))
 
-	router := web.Router(app)
+	router := web.Router(t, app, nil)
 	ts := httptest.NewServer(router)
 	defer ts.Close()
 
 	eia := auth.NewToken()
 	url := cltest.WebURL(t, "http://localhost:8888")
-	eir := &models.ExternalInitiatorRequest{
-		Name: "bitcoin",
+	eir := &bridges.ExternalInitiatorRequest{
+		Name: uuid.New().String(),
 		URL:  &url,
 	}
-	ea, err := models.NewExternalInitiator(eia, eir)
+	ea, err := bridges.NewExternalInitiator(eia, eir)
 	require.NoError(t, err)
-	err = app.GetStore().CreateExternalInitiator(ea)
+	err = app.BridgeORM().CreateExternalInitiator(ctx, ea)
 	require.NoError(t, err)
 
-	request, err := http.NewRequest("GET", ts.URL+"/v2/ping/", bytes.NewBufferString("{}"))
+	request, err := http.NewRequestWithContext(ctx, "GET", ts.URL+"/v2/ping/", bytes.NewBufferString("{}"))
 	require.NoError(t, err)
 	request.Header.Set("Content-Type", web.MediaType)
 	request.Header.Set("X-Chainlink-EA-AccessKey", eia.AccessKey)
 	request.Header.Set("X-Chainlink-EA-Secret", eia.Secret)
 
-	client := http.Client{}
+	client := clhttptest.NewTestLocalOnlyHTTPClient()
 	resp, err := client.Do(request)
 	require.NoError(t, err)
 
@@ -80,32 +86,32 @@ func TestTokenAuthRequired_TokenCredentials(t *testing.T) {
 }
 
 func TestTokenAuthRequired_BadTokenCredentials(t *testing.T) {
-	app, cleanup := cltest.NewApplicationWithKey(t, cltest.LenientEthMock)
-	defer cleanup()
-	require.NoError(t, app.Start())
+	ctx := testutils.Context(t)
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(ctx))
 
-	router := web.Router(app)
+	router := web.Router(t, app, nil)
 	ts := httptest.NewServer(router)
 	defer ts.Close()
 
 	eia := auth.NewToken()
 	url := cltest.WebURL(t, "http://localhost:8888")
-	eir := &models.ExternalInitiatorRequest{
-		Name: "bitcoin",
+	eir := &bridges.ExternalInitiatorRequest{
+		Name: uuid.New().String(),
 		URL:  &url,
 	}
-	ea, err := models.NewExternalInitiator(eia, eir)
+	ea, err := bridges.NewExternalInitiator(eia, eir)
 	require.NoError(t, err)
-	err = app.GetStore().CreateExternalInitiator(ea)
+	err = app.BridgeORM().CreateExternalInitiator(ctx, ea)
 	require.NoError(t, err)
 
-	request, err := http.NewRequest("GET", ts.URL+"/v2/ping/", bytes.NewBufferString("{}"))
+	request, err := http.NewRequestWithContext(ctx, "GET", ts.URL+"/v2/ping/", bytes.NewBufferString("{}"))
 	require.NoError(t, err)
 	request.Header.Set("Content-Type", web.MediaType)
 	request.Header.Set("X-Chainlink-EA-AccessKey", eia.AccessKey)
 	request.Header.Set("X-Chainlink-EA-Secret", "every unpleasant commercial color from aquamarine to beige")
 
-	client := http.Client{}
+	client := clhttptest.NewTestLocalOnlyHTTPClient()
 	resp, err := client.Do(request)
 	require.NoError(t, err)
 
@@ -113,19 +119,19 @@ func TestTokenAuthRequired_BadTokenCredentials(t *testing.T) {
 }
 
 func TestSessions_RateLimited(t *testing.T) {
-	app, cleanup := cltest.NewApplicationWithKey(t, cltest.LenientEthMock)
-	defer cleanup()
-	require.NoError(t, app.Start())
+	ctx := testutils.Context(t)
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(ctx))
 
-	router := web.Router(app)
+	router := web.Router(t, app, nil)
 	ts := httptest.NewServer(router)
 	defer ts.Close()
 
-	client := http.Client{}
+	client := clhttptest.NewTestLocalOnlyHTTPClient()
 	input := `{"email":"brute@force.com", "password": "wrongpassword"}`
 
 	for i := 0; i < 5; i++ {
-		request, err := http.NewRequest("POST", ts.URL+"/sessions", bytes.NewBufferString(input))
+		request, err := http.NewRequestWithContext(ctx, "POST", ts.URL+"/sessions", bytes.NewBufferString(input))
 		require.NoError(t, err)
 
 		resp, err := client.Do(request)
@@ -133,7 +139,7 @@ func TestSessions_RateLimited(t *testing.T) {
 		assert.Equal(t, http.StatusUnauthorized, resp.StatusCode)
 	}
 
-	request, err := http.NewRequest("POST", ts.URL+"/sessions", bytes.NewBufferString(input))
+	request, err := http.NewRequestWithContext(ctx, "POST", ts.URL+"/sessions", bytes.NewBufferString(input))
 	require.NoError(t, err)
 
 	resp, err := client.Do(request)
@@ -142,18 +148,18 @@ func TestSessions_RateLimited(t *testing.T) {
 }
 
 func TestRouter_LargePOSTBody(t *testing.T) {
-	app, cleanup := cltest.NewApplicationWithKey(t, cltest.LenientEthMock)
-	defer cleanup()
-	require.NoError(t, app.Start())
+	ctx := testutils.Context(t)
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(ctx))
 
-	router := web.Router(app)
+	router := web.Router(t, app, nil)
 	ts := httptest.NewServer(router)
 	defer ts.Close()
 
-	client := http.Client{}
+	client := clhttptest.NewTestLocalOnlyHTTPClient()
 
 	body := string(make([]byte, 70000))
-	request, err := http.NewRequest("POST", ts.URL+"/sessions", bytes.NewBufferString(body))
+	request, err := http.NewRequestWithContext(ctx, "POST", ts.URL+"/sessions", bytes.NewBufferString(body))
 	require.NoError(t, err)
 
 	resp, err := client.Do(request)
@@ -162,14 +168,16 @@ func TestRouter_LargePOSTBody(t *testing.T) {
 }
 
 func TestRouter_GinHelmetHeaders(t *testing.T) {
-	app, cleanup := cltest.NewApplicationWithKey(t, cltest.LenientEthMock)
-	defer cleanup()
-	require.NoError(t, app.Start())
+	ctx := testutils.Context(t)
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(ctx))
 
-	router := web.Router(app)
+	router := web.Router(t, app, nil)
 	ts := httptest.NewServer(router)
 	defer ts.Close()
-	res, err := http.Get(ts.URL)
+	req, err := http.NewRequestWithContext(ctx, "GET", ts.URL, nil)
+	require.NoError(t, err)
+	res, err := http.DefaultClient.Do(req)
 	require.NoError(t, err)
 	for _, tt := range []struct {
 		HelmetName  string
