@@ -3,132 +3,266 @@ package web_test
 import (
 	"bytes"
 	"encoding/json"
+	"fmt"
 	"net/http"
+	"net/url"
+	"strings"
 	"testing"
-
-	"github.com/ethereum/go-ethereum/accounts"
-	"github.com/smartcontractkit/chainlink/core/auth"
-	"github.com/smartcontractkit/chainlink/core/internal/cltest"
-	"github.com/smartcontractkit/chainlink/core/internal/mocks"
-	"github.com/smartcontractkit/chainlink/core/store/models"
-	"github.com/smartcontractkit/chainlink/core/store/presenters"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"github.com/smartcontractkit/chainlink/v2/core/auth"
+	"github.com/smartcontractkit/chainlink/v2/core/internal/cltest"
+	"github.com/smartcontractkit/chainlink/v2/core/internal/testutils"
+	"github.com/smartcontractkit/chainlink/v2/core/sessions"
+	"github.com/smartcontractkit/chainlink/v2/core/utils"
 )
 
 func TestUserController_UpdatePassword(t *testing.T) {
 	t.Parallel()
 
-	app, cleanup := cltest.NewApplicationWithKey(t, cltest.LenientEthMock)
-	defer cleanup()
-	require.NoError(t, app.Start())
-	client := app.NewHTTPClient()
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(testutils.Context(t)))
 
-	// Invalid request
-	resp, cleanup := client.Patch("/v2/user/password", bytes.NewBufferString(""))
-	defer cleanup()
+	u := cltest.User{}
+	client := app.NewHTTPClient(&u)
+
+	testCases := []struct {
+		name           string
+		reqBody        string
+		wantStatusCode int
+		wantErrCount   int
+		wantErrMessage string
+	}{
+		{
+			name:           "Invalid request",
+			reqBody:        "",
+			wantStatusCode: http.StatusUnprocessableEntity,
+			wantErrCount:   1,
+		},
+		{
+			name:           "Incorrect old password",
+			reqBody:        `{"oldPassword": "wrong password"}`,
+			wantStatusCode: http.StatusConflict,
+			wantErrCount:   1,
+			wantErrMessage: "old password does not match",
+		},
+		{
+			name:           "Insufficient length of new password",
+			reqBody:        fmt.Sprintf(`{"newPassword": "%v", "oldPassword": "%v"}`, "foo", cltest.Password),
+			wantStatusCode: http.StatusUnprocessableEntity,
+			wantErrCount:   1,
+			wantErrMessage: fmt.Sprintf("%s	%s\n", utils.ErrMsgHeader, "password is less than 16 characters long"),
+		},
+		{
+			name:           "New password includes api email",
+			reqBody:        fmt.Sprintf(`{"newPassword": "%slonglonglonglong", "oldPassword": "%s"}`, u.Email, cltest.Password),
+			wantStatusCode: http.StatusUnprocessableEntity,
+			wantErrCount:   1,
+			wantErrMessage: fmt.Sprintf("%s	%s%s\n", utils.ErrMsgHeader, "password may not contain: ", fmt.Sprintf(`"%s"`, u.Email)),
+		},
+		{
+			name:           "Success",
+			reqBody:        fmt.Sprintf(`{"newPassword": "%v", "oldPassword": "%v"}`, cltest.Password, cltest.Password),
+			wantStatusCode: http.StatusOK,
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			resp, cleanup := client.Patch("/v2/user/password", bytes.NewBufferString(tc.reqBody))
+			t.Cleanup(cleanup)
+			errors := cltest.ParseJSONAPIErrors(t, resp.Body)
+
+			require.Equal(t, tc.wantStatusCode, resp.StatusCode)
+			assert.Len(t, errors.Errors, tc.wantErrCount)
+			if tc.wantErrMessage != "" {
+				assert.Equal(t, tc.wantErrMessage, errors.Errors[0].Detail)
+			}
+		})
+	}
+}
+
+func TestUserController_CreateUser(t *testing.T) {
+	t.Parallel()
+
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(testutils.Context(t)))
+
+	client := app.NewHTTPClient(nil)
+
+	longPassword := strings.Repeat("x", sessions.MaxBcryptPasswordLength+1)
+
+	testCases := []struct {
+		name           string
+		reqBody        string
+		wantStatusCode int
+		wantErrCount   int
+		wantErrMessage string
+	}{
+		{
+			name:           "Invalid request",
+			reqBody:        "",
+			wantStatusCode: http.StatusUnprocessableEntity,
+			wantErrCount:   1,
+		},
+		{
+			name:           "Wrong email format",
+			reqBody:        fmt.Sprintf(`{"email": "12345678", "role": "view", "password": "%v"}`, cltest.Password),
+			wantStatusCode: http.StatusBadRequest,
+			wantErrCount:   1,
+			wantErrMessage: "mail: missing '@' or angle-addr",
+		},
+		{
+			name:           "Empty email format",
+			reqBody:        fmt.Sprintf(`{"email": "", "role": "view", "password": "%v"}`, cltest.Password),
+			wantStatusCode: http.StatusBadRequest,
+			wantErrCount:   1,
+			wantErrMessage: "Must enter an email",
+		},
+		{
+			name:           "Empty role",
+			reqBody:        fmt.Sprintf(`{"email": "abc@email.com", "role": "", "password": "%v"}`, cltest.Password),
+			wantStatusCode: http.StatusBadRequest,
+			wantErrCount:   1,
+			wantErrMessage: "Invalid role",
+		},
+		{
+			name:           "Too long password",
+			reqBody:        fmt.Sprintf(`{"email": "abc@email.com", "role": "view", "password": "%v"}`, longPassword),
+			wantStatusCode: http.StatusBadRequest,
+			wantErrCount:   1,
+			wantErrMessage: "must enter a password less than 50 characters",
+		},
+		{
+			name:           "Too short password",
+			reqBody:        `{"email": "abc@email.com", "role": "view", "password": "short"}`,
+			wantStatusCode: http.StatusBadRequest,
+			wantErrCount:   1,
+			wantErrMessage: "Must be at least 16 characters long",
+		},
+		{
+			name:           "Empty password",
+			reqBody:        `{"email": "abc@email.com", "role": "view", "password": ""}`,
+			wantStatusCode: http.StatusBadRequest,
+			wantErrCount:   1,
+			wantErrMessage: "Must be at least 16 characters long",
+		},
+		{
+			name:           "Password contains email",
+			reqBody:        `{"email": "asd@email.com", "role": "view", "password": "asd@email.comasd@email.comasd@email.com"}`,
+			wantStatusCode: http.StatusBadRequest,
+			wantErrCount:   1,
+			wantErrMessage: `password may not contain: "asd@email.com"`,
+		},
+		{
+			name:           "Success",
+			reqBody:        fmt.Sprintf(`{"email": "%s", "role": "edit", "password": "%v"}`, cltest.MustRandomUser(t).Email, cltest.Password),
+			wantStatusCode: http.StatusOK,
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			resp, cleanup := client.Post("/v2/users", bytes.NewBufferString(tc.reqBody))
+			t.Cleanup(cleanup)
+			errors := cltest.ParseJSONAPIErrors(t, resp.Body)
+
+			require.Equal(t, tc.wantStatusCode, resp.StatusCode)
+			assert.Len(t, errors.Errors, tc.wantErrCount)
+			if tc.wantErrMessage != "" {
+				assert.Contains(t, errors.Errors[0].Detail, tc.wantErrMessage)
+			}
+		})
+	}
+}
+
+func TestUserController_UpdateRole(t *testing.T) {
+	t.Parallel()
+	ctx := testutils.Context(t)
+
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(testutils.Context(t)))
+
+	client := app.NewHTTPClient(nil)
+	user := cltest.MustRandomUser(t)
+	err := app.AuthenticationProvider().CreateUser(ctx, &user)
+	require.NoError(t, err)
+
+	testCases := []struct {
+		name           string
+		reqBody        string
+		wantStatusCode int
+		wantErrCount   int
+		wantErrMessage string
+	}{
+		{
+			name:           "Invalid request",
+			reqBody:        "",
+			wantStatusCode: http.StatusUnprocessableEntity,
+			wantErrCount:   1,
+		},
+		{
+			name:           "Success",
+			reqBody:        fmt.Sprintf(`{"email": "%s", "newRole": "edit"}`, user.Email),
+			wantStatusCode: http.StatusOK,
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			resp, cleanup := client.Patch("/v2/users", bytes.NewBufferString(tc.reqBody))
+			t.Cleanup(cleanup)
+			errors := cltest.ParseJSONAPIErrors(t, resp.Body)
+
+			require.Equal(t, tc.wantStatusCode, resp.StatusCode)
+			assert.Len(t, errors.Errors, tc.wantErrCount)
+			if tc.wantErrMessage != "" {
+				assert.Contains(t, errors.Errors[0].Detail, tc.wantErrMessage)
+			}
+		})
+	}
+}
+
+func TestUserController_DeleteUser(t *testing.T) {
+	t.Parallel()
+	ctx := testutils.Context(t)
+
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(testutils.Context(t)))
+
+	client := app.NewHTTPClient(nil)
+	user := cltest.MustRandomUser(t)
+	err := app.AuthenticationProvider().CreateUser(ctx, &user)
+	require.NoError(t, err)
+
+	resp, cleanup := client.Delete(fmt.Sprintf("/v2/users/%s", url.QueryEscape(user.Email)))
+	t.Cleanup(cleanup)
 	errors := cltest.ParseJSONAPIErrors(t, resp.Body)
-	require.Equal(t, http.StatusUnprocessableEntity, resp.StatusCode)
-	assert.Len(t, errors.Errors, 1)
-
-	// Old password is wrong
-	resp, cleanup = client.Patch(
-		"/v2/user/password",
-		bytes.NewBufferString(`{"oldPassword": "wrong password"}`))
-	defer cleanup()
-	errors = cltest.ParseJSONAPIErrors(t, resp.Body)
-	require.Equal(t, http.StatusConflict, resp.StatusCode)
-	assert.Len(t, errors.Errors, 1)
-	assert.Equal(t, "old password does not match", errors.Errors[0].Detail)
-
-	// Success
-	resp, cleanup = client.Patch(
-		"/v2/user/password",
-		bytes.NewBufferString(`{"newPassword": "password", "oldPassword": "password"}`))
-	defer cleanup()
-	errors = cltest.ParseJSONAPIErrors(t, resp.Body)
-	assert.Len(t, errors.Errors, 0)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
-}
+	assert.Empty(t, errors.Errors)
 
-func TestUserController_AccountBalances_NoAccounts(t *testing.T) {
-	t.Parallel()
-
-	app, cleanup := cltest.NewApplication(t, cltest.LenientEthMock)
-	kst := new(mocks.KeyStoreInterface)
-	kst.On("Accounts").Return([]accounts.Account{})
-	app.Store.KeyStore = kst
-	defer cleanup()
-	require.NoError(t, app.Start())
-
-	client := app.NewHTTPClient()
-
-	resp, cleanup := client.Get("/v2/user/balances")
-	defer cleanup()
-
-	balances := []presenters.AccountBalance{}
-	err := cltest.ParseJSONAPIResponse(t, resp, &balances)
-	assert.NoError(t, err)
-
-	assert.Equal(t, http.StatusOK, resp.StatusCode)
-	assert.Len(t, balances, 0)
-	kst.AssertExpectations(t)
-}
-
-func TestUserController_AccountBalances_Success(t *testing.T) {
-	t.Parallel()
-
-	app, cleanup := cltest.NewApplicationWithKey(t,
-		cltest.LenientEthMock,
-		cltest.EthMockRegisterChainID,
-		cltest.EthMockRegisterGetBalance,
-	)
-	defer cleanup()
-	require.NoError(t, app.Start())
-
-	app.AddUnlockedKey()
-	client := app.NewHTTPClient()
-
-	ethMock := app.EthMock
-	ethMock.Context("first wallet", func(ethMock *cltest.EthMock) {
-		ethMock.Register("eth_getBalance", "0x100")
-		ethMock.Register("eth_call", "0x100")
-	})
-	ethMock.Context("second wallet", func(ethMock *cltest.EthMock) {
-		ethMock.Register("eth_getBalance", "0x1")
-		ethMock.Register("eth_call", "0x1")
-	})
-
-	resp, cleanup := client.Get("/v2/user/balances")
-	defer cleanup()
-	require.Equal(t, http.StatusOK, resp.StatusCode)
-
-	expectedAccounts := app.Store.KeyStore.Accounts()
-	actualBalances := []presenters.AccountBalance{}
-	err := cltest.ParseJSONAPIResponse(t, resp, &actualBalances)
-	assert.NoError(t, err)
-
-	first := actualBalances[0]
-	assert.Equal(t, expectedAccounts[0].Address.Hex(), first.Address)
-	assert.Equal(t, "0.000000000000000256", first.EthBalance.String())
-	assert.Equal(t, "0.000000000000000256", first.LinkBalance.String())
-
-	second := actualBalances[1]
-	assert.Equal(t, expectedAccounts[1].Address.Hex(), second.Address)
-	assert.Equal(t, "0.000000000000000001", second.EthBalance.String())
-	assert.Equal(t, "0.000000000000000001", second.LinkBalance.String())
+	// second attempt would fail
+	resp, cleanup = client.Delete(fmt.Sprintf("/v2/users/%s", url.QueryEscape(user.Email)))
+	t.Cleanup(cleanup)
+	errors = cltest.ParseJSONAPIErrors(t, resp.Body)
+	assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
+	assert.Len(t, errors.Errors, 1)
+	assert.Contains(t, errors.Errors[0].Detail, "specified user not found")
 }
 
 func TestUserController_NewAPIToken(t *testing.T) {
 	t.Parallel()
 
-	app, cleanup := cltest.NewApplicationWithKey(t, cltest.LenientEthMock)
-	defer cleanup()
-	require.NoError(t, app.Start())
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(testutils.Context(t)))
 
-	client := app.NewHTTPClient()
-	req, err := json.Marshal(models.ChangeAuthTokenRequest{
+	client := app.NewHTTPClient(nil)
+	req, err := json.Marshal(sessions.ChangeAuthTokenRequest{
 		Password: cltest.Password,
 	})
 	require.NoError(t, err)
@@ -146,12 +280,11 @@ func TestUserController_NewAPIToken(t *testing.T) {
 func TestUserController_NewAPIToken_unauthorized(t *testing.T) {
 	t.Parallel()
 
-	app, cleanup := cltest.NewApplicationWithKey(t, cltest.LenientEthMock)
-	defer cleanup()
-	require.NoError(t, app.Start())
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(testutils.Context(t)))
 
-	client := app.NewHTTPClient()
-	req, err := json.Marshal(models.ChangeAuthTokenRequest{
+	client := app.NewHTTPClient(nil)
+	req, err := json.Marshal(sessions.ChangeAuthTokenRequest{
 		Password: "wrong-password",
 	})
 	require.NoError(t, err)
@@ -163,12 +296,11 @@ func TestUserController_NewAPIToken_unauthorized(t *testing.T) {
 func TestUserController_DeleteAPIKey(t *testing.T) {
 	t.Parallel()
 
-	app, cleanup := cltest.NewApplicationWithKey(t, cltest.LenientEthMock)
-	defer cleanup()
-	require.NoError(t, app.Start())
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(testutils.Context(t)))
 
-	client := app.NewHTTPClient()
-	req, err := json.Marshal(models.ChangeAuthTokenRequest{
+	client := app.NewHTTPClient(nil)
+	req, err := json.Marshal(sessions.ChangeAuthTokenRequest{
 		Password: cltest.Password,
 	})
 	require.NoError(t, err)
@@ -181,12 +313,11 @@ func TestUserController_DeleteAPIKey(t *testing.T) {
 func TestUserController_DeleteAPIKey_unauthorized(t *testing.T) {
 	t.Parallel()
 
-	app, cleanup := cltest.NewApplicationWithKey(t, cltest.LenientEthMock)
-	defer cleanup()
-	require.NoError(t, app.Start())
+	app := cltest.NewApplicationEVMDisabled(t)
+	require.NoError(t, app.Start(testutils.Context(t)))
 
-	client := app.NewHTTPClient()
-	req, err := json.Marshal(models.ChangeAuthTokenRequest{
+	client := app.NewHTTPClient(nil)
+	req, err := json.Marshal(sessions.ChangeAuthTokenRequest{
 		Password: "wrong-password",
 	})
 	require.NoError(t, err)
